### 2020-10-01 #2710 @whedon
[PRE REVIEW]: ParaMonte: A high-performance serial/parallel Monte Carlo simulation library for C, C++, Fortran
**Submitting author:** @shahmoradi (<a href="http://orcid.org/0000-0002-9720-8937">Amir Shahmoradi</a>)
**Repository:** <a href="https://github.com/cdslaborg/paramonte" target ="_blank">https://github.com/cdslaborg/paramonte</a>
**Version:** v1.2.0
