### 2020-10-01 #2710 @Kevin-Mattheus-Moerman
@shahmoradi I think our system cannot see your paper (which appears to be at `https://github.com/cdslaborg/paramonte/tree/master/pubs/ParaMonte/kernel/JOSS2020`), can you move the files to your main folder or put them in a `paper` folder?

@openjournals/dev can you confirm this is the issue here? Thanks
