### 2020-10-01 #2710 @xuanxu
> I think our system cannot see your paper (which appears to be at `https://github.com/cdslaborg/paramonte/tree/master/pubs/ParaMonte/kernel/JOSS2020`), can you move the files to your main folder or put them in a `paper` folder?

That should not be a problem, whedon can find the paper in any subfolder
