+++ ACKNOWLEDGMENT.md
@@ -10,22 +10,42 @@ library's main publications as listed here:
   
 ```text  
 
-    @article{2020arXiv200809589S,
-               author = {Shahmoradi, Amir and Bagheri, Fatemeh},
-                title = "{ParaDRAM: A Cross-Language Toolbox for Parallel High-Performance Delayed-Rejection Adaptive Metropolis Markov Chain Monte Carlo Simulations}",
-              journal = {arXiv e-prints},
-             keywords = {Computer Science - Computational Engineering, Finance, and Science, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Statistics - Computation, Statistics - Machine Learning},
-                 year = 2020,
-                month = aug,
-                  eid = {arXiv:2008.09589},
-                pages = {arXiv:2008.09589},
-        archivePrefix = {arXiv},
-               eprint = {2008.09589},
-         primaryClass = {cs.CE},
-               adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200809589S},
-              adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-    }
+@article{2020arXiv200809589S,
+           author = { {Shahmoradi}, Amir and {Bagheri}, Fatemeh},
+            title = "{ParaDRAM: A Cross-Language Toolbox for Parallel High-Performance Delayed-Rejection Adaptive Metropolis Markov Chain Monte Carlo Simulations}",
+          journal = {arXiv e-prints},
+         keywords = {Computer Science - Computational Engineering, Finance, and Science, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Statistics - Computation, Statistics - Machine Learning},
+             year = 2020,
+            month = aug,
+              eid = {arXiv:2008.09589},
+            pages = {arXiv:2008.09589},
+    archivePrefix = {arXiv},
+           eprint = {2008.09589},
+     primaryClass = {cs.CE},
+           adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200809589S},
+          adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
-```    
+```  
+  
+```text  
+
+@article{2020arXiv200914229S,
+       author = { {Shahmoradi}, Amir and {Bagheri}, Fatemeh},
+        title = "{ParaMonte: A high-performance serial/parallel Monte Carlo simulation library for C, C++, Fortran}",
+      journal = {arXiv e-prints},
+     keywords = {Computer Science - Mathematical Software, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Quantitative Biology - Quantitative Methods, Statistics - Machine Learning},
+         year = 2020,
+        month = sep,
+          eid = {arXiv:2009.14229},
+        pages = {arXiv:2009.14229},
+archivePrefix = {arXiv},
+       eprint = {2009.14229},
+ primaryClass = {cs.MS},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200914229S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+```  
   
 Visit [the ParaMonte library homepage](https://www.cdslab.org/paramonte/notes/overview/preface/#how-to-acknowledge-the-use-of-the-paramonte-library-in-your-work) to access the PDF version of these files free of charge.  
