### 2020-10-03 b49a9d1
New ParaMonte sampler specification added.
A new simulation specification `overwriteRequested` has
been added to all ParaMonte samplers. If `True` and the
ParaMonte sampler detects an existing set of old simulation
output files in the output path of the current simulation with
the same names as the output file names of the current simulation,
then, the ParaMonte sampler will overwrite the existing simulation files.
