+++ src/ParaMonte/SpecBase_OverwriteRequested_mod.f90
@@ -40,26 +40,26 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-module SpecBase_InputFileHasPriority_mod
+module SpecBase_OverwriteRequested_mod
 
     implicit none
 
     ! namelist input
-    logical                         :: inputFileHasPriority
+    logical                         :: overwriteRequested
 
-    type                            :: InputFileHasPriority_type
+    type                            :: OverwriteRequested_type
         logical                     :: val
         logical                     :: def
         character(:), allocatable   :: desc
     contains
-        procedure, pass             :: set => setInputFileHasPriority, nullifyNameListVar
-    end type InputFileHasPriority_type
+        procedure, pass             :: set => setOverwriteRequested, nullifyNameListVar
+    end type OverwriteRequested_type
 
-    interface InputFileHasPriority_type
-        module procedure            :: constructInputFileHasPriority
-    end interface InputFileHasPriority_type
+    interface OverwriteRequested_type
+        module procedure            :: constructOverwriteRequested
+    end interface OverwriteRequested_type
 
-    private :: constructInputFileHasPriority, setInputFileHasPriority
+    private :: constructOverwriteRequested, setOverwriteRequested
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -67,51 +67,45 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-    function constructInputFileHasPriority(methodName) result(InputFileHasPriorityObj)
+    function constructOverwriteRequested(methodName) result(OverwriteRequestedObj)
 #if defined DLL_ENABLED && !defined CFI_ENABLED
-        !DEC$ ATTRIBUTES DLLEXPORT :: constructInputFileHasPriority
+        !DEC$ ATTRIBUTES DLLEXPORT :: constructOverwriteRequested
 #endif
         use String_mod, only: log2str
         implicit none
         character(*), intent(in)        :: methodName
-        type(InputFileHasPriority_type) :: InputFileHasPriorityObj
-        InputFileHasPriorityObj%def = .false.
-        InputFileHasPriorityObj%desc = &
-        "A logical (boolean) variable. If TRUE (or .true. or true or .t. from within an input file), then the &
-        &input specifications of the sampler will be read from the input file provided by the user, and the &
-        &simulation specification assignments from within the programming language environment (if any are made) &
-        &will be completely ignored. If inputFileHasPriority is FALSE, then all simulation specifications of the "//methodName//" &
[…]
