+++ src/ParaMonte/SpecBase_mod.f90
@@ -52,6 +52,7 @@ module SpecBase_mod
     use SpecBase_VariableNameList_mod               , only: VariableNameList_type
     use SpecBase_RestartFileFormat_mod              , only: RestartFileFormat_type
     use SpecBase_OutputColumnWidth_mod              , only: OutputColumnWidth_type
+    use SpecBase_OverwriteRequested_mod             , only: OverwriteRequested_type
     use SpecBase_OutputRealPrecision_mod            , only: OutputRealPrecision_type
     use SpecBase_SilentModeRequested_mod            , only: SilentModeRequested_type
     use SpecBase_DomainLowerLimitVec_mod            , only: DomainLowerLimitVec_type
@@ -77,6 +78,7 @@ module SpecBase_mod
     use SpecBase_VariableNameList_mod               , only: variableNameList
     use SpecBase_RestartFileFormat_mod              , only: restartFileFormat
     use SpecBase_OutputColumnWidth_mod              , only: outputColumnWidth
+    use SpecBase_OverwriteRequested_mod             , only: overwriteRequested
     use SpecBase_OutputRealPrecision_mod            , only: outputRealPrecision
     use SpecBase_SilentModeRequested_mod            , only: silentModeRequested
     use SpecBase_DomainLowerLimitVec_mod            , only: domainLowerLimitVec
@@ -84,7 +86,7 @@ module SpecBase_mod
     use SpecBase_ParallelizationModel_mod           , only: ParallelizationModel
     use SpecBase_InputFileHasPriority_mod           , only: inputFileHasPriority
     use SpecBase_ProgressReportPeriod_mod           , only: progressReportPeriod
-    use SpecBase_TargetAcceptanceRate_mod           , only: TargetAcceptanceRate
+    use SpecBase_TargetAcceptanceRate_mod           , only: targetAcceptanceRate
     use SpecBase_MpiFinalizeRequested_mod           , only: mpiFinalizeRequested
     use SpecBase_MaxNumDomainCheckToWarn_mod        , only: maxNumDomainCheckToWarn
     use SpecBase_MaxNumDomainCheckToStop_mod        , only: maxNumDomainCheckToStop
@@ -105,6 +107,7 @@ module SpecBase_mod
         type(VariableNameList_type)             :: VariableNameList
         type(RestartFileFormat_type)            :: RestartFileFormat
         type(OutputColumnWidth_type)            :: OutputColumnWidth
+        type(OverwriteRequested_type)           :: OverwriteRequested
         type(OutputRealPrecision_type)          :: OutputRealPrecision
         type(SilentModeRequested_type)          :: SilentModeRequested
         type(DomainLowerLimitVec_type)          :: domainLowerLimitVec
@@ -155,6 +158,7 @@ contains
         SpecBase%VariableNameList                   = VariableNameList_type(nd,methodName)
         SpecBase%RestartFileFormat                  = RestartFileFormat_type(methodName)
         SpecBase%OutputColumnWidth                  = OutputColumnWidth_type(methodName)
+        SpecBase%OverwriteRequested                 = OverwriteRequested_type(methodName)
         SpecBase%OutputRealPrecision                = OutputRealPrecision_type(methodName)
         SpecBase%SilentModeRequested                = SilentModeRequested_type(methodName)
         SpecBase%DomainLowerLimitVec                = DomainLowerLimitVec_type(methodName)
@@ -190,6 +194,7 @@ contains
         call SpecBase%VariableNameList              %nullifyNameListVar(nd)
         call SpecBase%RestartFileFormat             %nullifyNameListVar()
         call SpecBase%OutputColumnWidth             %nullifyNameListVar()
+        call SpecBase%OverwriteRequested            %nullifyNameListVar()
[…]
