### 2020-10-03 ce1833e
minor enhancements
