+++ src/interface/MATLAB/README.md
@@ -69,6 +69,20 @@ For a quick start with some MATLAB Live Script examples, visit [this ParaMonte d
 The corresponding example source files (the `*.mlx` files) can be downloaded from the [paramontex GitHub repository](https://github.com/cdslaborg/paramontex/tree/master/MATLAB/mlx), 
 a repository dedicated to the ParaMonte library examples.  
 
+The following example code samples a 4-dimensional MultiVariate Normal (MNV) distribution via the ParaDRAM sampler in serial mode,  
+
+```matlab  
+addpath(genpath("./"),"-begin") % change this path to the root directory of paramonte
+pm = paramonte();
+pmpd = pm.ParaDRAM();
+getLogFunc = @(x) -0.5 * sum( x.^2 );
+pmpd.runSampler ( 4 ... assume a 4-dimensional objective function
+                , getLogFunc ...           the objective function
+                );
+```  
+
+To learn about the post-processing and visualization tools of the `ParaMonte::Python` library, visit [this this documentation page](https://www.cdslab.org/paramonte/notes/examples/matlab/mlx/).  
+
   
 Installation  
 ============  
