### 2020-10-03 2a5fa56
minor enhancements
