### 2020-10-03 f5c00b7
minor enhancements
