### 2020-10-03 6d91a99
minor enhancements
