### 2020-10-03 14eb884
minor enhancements
