+++ src/interface/Python/paramonte/vis/LineScatterPlot.py
@@ -42,7 +42,6 @@
 
 import numpy as np
 import typing as tp
-import pandas as pd
 import weakref as wref
 import _paramonte as pm
 import _pmutils as pmutils
@@ -450,10 +449,10 @@ class LineScatterPlot(BasePlot):
     ################################################################################################################################
 
     def __init__( self
-                , plotType      : str
-                , dataFrame     : tp.Optional[ pd.DataFrame ] = None
-                , methodName    : tp.Optional[ str ] = "ParaMonte"
-                , reportEnabled : tp.Optional[ bool ] = True
+                , plotType      # : str
+                , dataFrame     = None # : tp.Optional[ pd.DataFrame ] = None
+                , methodName    = "ParaMonte" # : tp.Optional[ str ] = "ParaMonte"
+                , reportEnabled = True # : tp.Optional[ bool ] = True
                 , resetPlot     = None
                 ):
 
