+++ src/interface/Python/paramonte/vis/DensityPlot.py
@@ -477,10 +477,10 @@ class DensityPlot(BasePlot):
     ################################################################################################################################
 
     def __init__( self
-                , plotType      : str
-                , dataFrame     : tp.Optional[ pd.DataFrame ] = None
-                , methodName    : tp.Optional[ str ] = "ParaMonte"
-                , reportEnabled : tp.Optional[ bool ] = True
+                , plotType      # : str
+                , dataFrame     = None # : tp.Optional[ pd.DataFrame ] = None
+                , methodName    = "ParaMonte" # : tp.Optional[ str ] = "ParaMonte"
+                , reportEnabled = True # : tp.Optional[ bool ] = True
                 , resetPlot     = None
                 ):
 
