### 2020-10-03 2d54993
minor enhancements
