### 2020-10-03 bdb65d4
minor enhancements
