+++ src/ParaMonte/filelist.txt
@@ -2,101 +2,101 @@
 ! used for building ParaMonte library.
 ! any line preceded by the exclamation mark (!) is comment and will be ignored.
 !
-  !ifport.f90
-  Constants_mod.f90
-  Misc_mod.f90
-  !EconomicsToolbox_mod.f90
-  JaggedArray_mod.f90
-  String_mod.f90
-  Decoration_mod.f90
-  Decoration_mod@Routines_smod.f90
-  Err_mod.f90
-  Optimization_mod.f90
-  DateTime_mod.f90
-  System_mod.f90
-  Path_mod.f90
-  FileList_mod.f90
-  Timer_mod.f90
-  TimerCPU_mod.f90
-  RandomSeed_mod.f90
-  FileContents_mod.f90
-  File_mod.f90
-  Sort_mod.f90
-  Math_mod.f90
-  Matrix_mod.f90
-  Statistics_mod.f90
-  Parallelism_mod.f90
-  CorrCoef_mod.f90
-  TranGaus_mod.f90
-  CrossCorr_mod.f90
-  Integration_mod.f90
-  QuadPackSPR_mod.f90
-  QuadPackDPR_mod.f
-  Cosmology_mod.f90
-  StarFormation_mod.f90
-  BandSpectrum_mod.f90
-  Batse_mod.f90
-  !
-  SpecBase_SampleSize_mod.f90
-  SpecBase_RandomSeed_mod.f90
-  SpecBase_Description_mod.f90
-  SpecBase_InterfaceType_mod.f90
-  SpecBase_OutputFileName_mod.f90
-  SpecBase_OutputDelimiter_mod.f90
-  SpecBase_ChainFileFormat_mod.f90
-  SpecBase_VariableNameList_mod.f90
-  SpecBase_RestartFileFormat_mod.f90
-  SpecBase_OutputColumnWidth_mod.f90
-  SpecBase_OverwriteRequested_mod.f90
-  SpecBase_SystemInfoFilePath_mod.f90
-  SpecBase_OutputRealPrecision_mod.f90
-  SpecBase_SilentModeRequested_mod.f90
-  SpecBase_DomainLowerLimitVec_mod.f90
-  SpecBase_DomainUpperLimitVec_mod.f90
-  SpecBase_ParallelizationModel_mod.f90
-  SpecBase_InputFileHasPriority_mod.f90
-  SpecBase_ProgressReportPeriod_mod.f90
-  SpecBase_TargetAcceptanceRate_mod.f90
-  SpecBase_MpiFinalizeRequested_mod.f90
-  SpecBase_MaxNumDomainCheckToWarn_mod.f90
-  SpecBase_MaxNumDomainCheckToStop_mod.f90
-  SpecBase_mod.f90
-  !
-  SpecMCMC_ChainSize_mod.f90
-  SpecMCMC_ScaleFactor_mod.f90
-  SpecMCMC_StartPointVec_mod.f90
-  SpecMCMC_ProposalModel_mod.f90
-  SpecMCMC_ProposalStartStdVec_mod.f90
-  SpecMCMC_ProposalStartCorMat_mod.f90
-  SpecMCMC_ProposalStartCovMat_mod.f90
-  SpecMCMC_SampleRefinementCount_mod.f90
-  SpecMCMC_SampleRefinementMethod_mod.f90
-  SpecMCMC_RandomStartPointRequested_mod.f90
-  SpecMCMC_RandomStartPointDomainLowerLimitVec_mod.f90
[…]
