### 2020-10-03 13e1632
minor enhancements
