### 2020-10-03 855ecef
minor enhancements
