### 2020-10-04 51c8e82
minor enhancements
