+++ src/interface/MATLAB/paramonte/interface/@paramonte/paramonte.m
@@ -2220,7 +2220,7 @@ classdef paramonte %< dynamicprops
                 + "command and takes only a few seconds to install. " + newline ...
                 + "You can get the installation command from this page: " + newline ...
                 + newline ...
-                + "    " + pm.website.home.install.macos.prereqs.cmd.url + newline ...
+                + "    " + self.website.home.install.macos.prereqs.cmd.url + newline ...
                 + newline ...
                 + "Once you have manually installed the missing component, retry " + newline ...
                 + "the following commands on your MATLAB command line, " + newline ...
