### 2020-10-05 9bd63f5
minor enhancements
