### 2020-10-06 6ec247c
minor enhancements
