+++ ACKNOWLEDGMENT.md
@@ -4,69 +4,86 @@ As per the ParaMonte library license agreement terms, if you use any parts of
 this library for any purposes, kindly acknowledge the use of ParaMonte in your 
 work (education/research/industry/development/...) by citing the ParaMonte 
 library's main publications as listed here:  
-  
-  
-**BibTeX citation entries:**  
-  
-```text  
 
-@article{2020arXiv200809589S,
-           author = { {Shahmoradi}, Amir and {Bagheri}, Fatemeh},
-            title = "{ParaDRAM: A Cross-Language Toolbox for Parallel High-Performance Delayed-Rejection Adaptive Metropolis Markov Chain Monte Carlo Simulations}",
+### The ParaMonte Python library  
+
++   Amir Shahmoradi, Fatemeh Bagheri, Joshua Alexander Osborne (2020). 
+    **Fast fully-reproducible streamlined serial/parallel Monte Carlo/MCMC simulations and visualizations via `ParaMonte::Python` library.**. 
+    Journal of Open Source Software (JOSS), to be submitted, [**PDF link**](https://www.cdslab.org/pubs/2020_Shahmoradi_III.pdf).  
+    **BibTeX citation entries:**  
+    ```text  
+    
+    @article{2020arXiv201000724S,
+           author = { {Shahmoradi}, Amir and {Bagheri}, Fatemeh and {Osborne}, Joshua Alexand
+            er},
+            title = "{Fast fully-reproducible serial/parallel Monte Carlo and MCMC simulations and visualizations via ParaMonte::Python library}",
           journal = {arXiv e-prints},
-         keywords = {Computer Science - Computational Engineering, Finance, and Science, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Statistics - Computation, Statistics - Machine Learning},
+         keywords = {Computer Science - Mathematical Software, Astrophysics - Instrumentation and Methods for Astrophysics, Quantitative Biology - Quantitative Methods, Statistics - Machine Learning},
              year = 2020,
-            month = aug,
-              eid = {arXiv:2008.09589},
-            pages = {arXiv:2008.09589},
+            month = oct,
+              eid = {arXiv:2010.00724},
+            pages = {arXiv:2010.00724},
     archivePrefix = {arXiv},
-           eprint = {2008.09589},
-     primaryClass = {cs.CE},
-           adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200809589S},
+           eprint = {2010.00724},
+     primaryClass = {cs.MS},
+           adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv201000724S},
           adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
+    }
+    
+    ``` 
 
-```  
+### The ParaMonte C/C++/Fortran library  
+
++   Amir Shahmoradi, Fatemeh Bagheri (2020). 
+    **ParaMonte: A high-performance serial/parallel Monte Carlo simulation library for C, C++, Fortran**. 
+    Journal of Open Source Software (JOSS), submitted, [**PDF link**](https://www.cdslab.org/pubs/2020_Shahmoradi_II.pdf).  
+    **BibTeX citation entries:**  
+    ```text  
+    
+    @article{2020arXiv200914229S,
+           author = { {Shahmoradi}, Amir and {Bagheri}, Fatemeh},
+            title = "{ParaMonte: A high-performance serial/parallel Monte Carlo simulation library for C, C++, Fortran}",
+          journal = {arXiv e-prints},
[…]
