### 2020-10-08 342b3d7
enhancements to MatDRAM sampler
