+++ src/interface/MATLAB/paramonte/kernel/ParaDRAMProposalSymmetric_class.m
@@ -87,7 +87,7 @@ classdef ParaDRAMProposalSymmetric_class < handle
 %                    mc_DomainUpperLimitVec              ...
 %                    mc_MaxNumDomainCheckToWarnMsg       ...
 %                    mc_MaxNumDomainCheckToStopMsg       ...
-%                    mc_negativeHellingerDistSqMsg       ...
+%                    mc_negativeAdaptationMeasureMsg       ...
 %                    mc_restartFileFormat                ...
 %                    mc_methodBrand                      ...
 %                    mc_methodName                       ...
@@ -185,7 +185,7 @@ classdef ParaDRAMProposalSymmetric_class < handle
 
             % This will be used for Domain boundary checking during the simulation
 
-            self.Global.mc_negativeHellingerDistSqMsg   = self.SUB_CLASS_NAME + "@doAdaptation(): Non-positive adaptation measure detected, "           ...
+            self.Global.mc_negativeAdaptationMeasureMsg   = self.SUB_CLASS_NAME + "@doAdaptation(): Non-positive adaptation measure detected, "           ...
                                             + "possibly due to round-off error: ";
 
             self.Global.mc_MaxNumDomainCheckToWarnMsg   = self.SUB_CLASS_NAME + "@getNew(): " + num2str(self.Global.mc_MaxNumDomainCheckToWarn) + " proposals were "...
@@ -290,8 +290,8 @@ classdef ParaDRAMProposalSymmetric_class < handle
     %*******************************************************************************************************************************
     
         % ATTN: This routine may need further correction for the delayed rejection method
-        function hellingerDistSq =  doAutoTune  ( self              ...
-                                                , hellingerDistSq   ...
+        function adaptationMeasure =  doAutoTune( self              ...
+                                                , adaptationMeasure   ...
                                                 , AutoTuneScaleSq   ...
                                                 )
 %            global  comv_chol               ...
@@ -315,7 +315,7 @@ classdef ParaDRAMProposalSymmetric_class < handle
                 self.Global.comv_chol(1,1,1)    = sqrt(AutoTuneScaleSq(1));
             end
 
-            % compute the hellingerDistSq (adaptivity)
+            % compute the adaptationMeasure (adaptivity)
             logSqrtDetNew       = sum(log(diag(self.Global.comv_chol(:,:,1))));
             CovMatUpperCurrent  = 0.5 * (self.Global.comv_covMat(:,:,1) + CovMatUpperOld);
             [logSqrtDetSum, singularityOccurred] = getLogSqrtDetPosDefMat(1, CovMatUpperCurrent);
@@ -329,21 +329,21 @@ classdef ParaDRAMProposalSymmetric_class < handle
                 self.Global.mv_Err.abort();
                 return
             end
-            hellingerDistSq = 1 - exp(0.5 * (self.Global.mv_logSqrtDetOld_save + logSqrtDetNew) - logSqrtDetSum);
+            adaptationMeasure = 1 - exp(0.5 * (self.Global.mv_logSqrtDetOld_save + logSqrtDetNew) - logSqrtDetSum);
 
         end
 
     %*******************************************************************************************************************************
     %*******************************************************************************************************************************
 
-        function [samplerUpdateSucceeded, hellingerDistSq]  = doAdaptation  ( self                      ...
+        function [samplerUpdateSucceeded, adaptationMeasure]  = doAdaptation( self                      ...
                                                                             , nd                        ...
                                                                             , chainSize                 ...
                                                                             , Chain                     ...
                                                                             , ChainWeight               ...
                                                                             , samplerUpdateIsGreedy     ...
[…]
