+++ src/interface/MATLAB/paramonte/kernel/@ParaDRAM_class/runSampler.m
@@ -56,7 +56,7 @@ function runSampler ( self          ...
     self.setupParaMonte ( ndim                      ...
                         , Constants.PMSM.MatDRAM    ...
                         , "May 23 2018"             ...
-                        , "1.0.0"                   ...
+                        , self.version              ...
                         );
 
     %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -86,6 +86,7 @@ function runSampler ( self          ...
                                     , self.spec.variableNameList            ...
                                     , self.spec.restartFileFormat           ...
                                     , self.spec.outputColumnWidth           ...
+                                    , self.spec.overwriteRequested          ...
                                     , self.spec.outputRealPrecision         ...
                                     , self.spec.silentModeRequested         ...
                                     , self.spec.domainLowerLimitVec         ...
