+++ src/interface/MATLAB/paramonte/kernel/SpecBase_SilentModeRequested_class.m
@@ -60,7 +60,7 @@ classdef SpecBase_SilentModeRequested_class < handle
         function self = SpecBase_SilentModeRequested_class(methodName)
             self.def        = false;
             self.isFalse    = true;
-            self.desc       = "If silentModeRequested = true (or T, both case-insensitive), then the following contents will not be printed in the "...
+            self.desc       = "If silentModeRequested = true, then the following contents will not be printed in the "...
                             + "output report file of " + methodName + ":" + newline + newline...
                             + "    - " + methodName + " interface, compiler, and platform specifications." + newline...
                             + "    - " + methodName + " simulation specification-descriptions." + newline + newline...
