+++ src/interface/MATLAB/paramonte/kernel/@ParaDRAM_class/runKernel.m
@@ -388,7 +388,7 @@ if flag2, self.writeOutput2(); end
                                                                                     , self.Chain.Weight (numFunCallAcceptedLastAdaptation:self.Stats.NumFunCall.accepted)       ...
                                                                                     , samplerUpdateIsGreedy                                                                     ...
                                                                                     , meanAccRateSinceStart                                                                     ...
-                                                                                    , chainAdaptationMeasure                                                                    ...
+                                                                                    , chainAdaptationMeasure ...
                                                                                     ) ;
 
             self.Chain.Weight(self.Stats.NumFunCall.accepted) = dummy;  % needed for the restart mode, not needed in the fresh run
