+++ src/interface/MATLAB/paramonte/interface/@paramonte/paramonte.m
@@ -246,7 +246,7 @@ classdef paramonte %< dynamicprops
                 self.version.(versionType) = Version_class(self.path.auxil,versionType);
             end
 
-            self.prereqs = struct();
+            %self.prereqs = struct();
 
             %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
             %%%% set up website
@@ -632,7 +632,7 @@ classdef paramonte %< dynamicprops
 
                     % get the dependency list
 
-                    self.getPrereqs(false); % sets the self.prereqs component.
+                    self.prereqs = self.getPrereqs(false); % sets the self.prereqs component.
 
                     % search for the MPI library
 
@@ -689,8 +689,12 @@ classdef paramonte %< dynamicprops
                                                     + "    (only needed for parallel simulations) on your system now (y/n)? " ...
                                                     );
                             if isYes
-                                self.installMPI();
-                                self.writeVerificationStatusFile("Testing");
+                                succeeded = self.installMPI();
+                                if succeeded
+                                    self.writeVerificationStatusFile("Testing");
+                                else
+                                    self.writeVerificationStatusFile("False");
+                                end
                             else
                                 self.Err.msg    = "Skipping the MPI library installation... " + newline ...
                                                 + "It is now the user's responsibility to ensure an MPI runtime library" ...
@@ -712,7 +716,12 @@ classdef paramonte %< dynamicprops
                 else
 
                     self.warnForUnsupportedPlatform();
-                    self.build();
+                    succeeded = self.build();
+                    if succeeded
+                        self.writeVerificationStatusFile("Testing");
+                    else
+                        self.writeVerificationStatusFile("False");
+                    end
 
                 end
 
@@ -949,7 +958,9 @@ classdef paramonte %< dynamicprops
 
         %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-        function build(self,varargin)
+        function succeeded = build(self,varargin)
+
+            succeeded = false;
 
             if nargin==1
                 flags = "";
@@ -973,7 +984,8 @@ classdef paramonte %< dynamicprops
                                 + "please visit the following page for the installation instructions:" + newline + newline ...
                                 + "    " + href(self.website.home.url) + newline + newline ...
                                 + "Follow the website's instructions to build the ParaMonte library on your system.";
-                self.Err.abort();
+                self.Err.note();
+                return
 
             else
 
@@ -1040,7 +1052,7 @@ classdef paramonte %< dynamicprops
                     [errorOccurred1, ~] = system( ['find ', pmGitRootDir, ' -type f -iname \"*.sh\" -exec chmod +x {} \;'], "-echo" );
                     [errorOccurred2, ~] = system( pmGitInstallScriptPath + " --lang matlab --test_enabled true --exam_enabled false --yes-to-all " + flags, "-echo" );
                     if errorOccurred1 || errorOccurred2
-                        self.Err.msg    = "Local installation of ParaMonte failed." + newline ...
+                        self.Err.msg    = "The local installation of ParaMonte failed." + newline ...
                                         + "Please report this issue at " + newline + newline ...
                                         + "    " + href(self.website.github.issues.url);
                         self.Err.abort();
@@ -1115,10 +1127,13 @@ classdef paramonte %< dynamicprops
 
[…]
