### 2020-10-10 cc06051
minor enhancements
