### 2020-10-10 a63e1a9
minor enhancements
