### 2020-10-10 ae95968
minor enhancements
