### 2020-10-10 dab556a
minor enhancements
