+++ src/interface/MATLAB/paramonte/interface/@paramonte/paramonte.m
@@ -739,6 +739,10 @@ classdef paramonte %< dynamicprops
 
             end
 
+            if verificationEnabledString=="Testing"
+                self.writeVerificationStatusFile("False");
+            end
+
         end
 
         %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
