### 2020-10-10 6f3cade
minor enhancements
