### 2020-10-11 a861681
minor enhancements
