### 2020-10-11 88f2671
minor enhancements
