### 2020-10-11 64c123d
minor enhancements
