### 2020-10-11 401ab1a
minor enhancements
