+++ example/buildParaMonteExample.sh
@@ -59,7 +59,38 @@ echo >&2
 # setup examples' interface language
 #:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
-abort() {
+BUILD_NAME="buildParaMonteExample"
+verify() {
+    if [ $1 -eq 0 ]; then
+        echo >&2 "-- ${BUILD_NAME} - ParaMonte $2 appears to have succeeded."
+    else
+        echo >&2
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: ParaMonte $2 appears to have failed."
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: If the source of the error cannot be identified,"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: consider a fresh installation of ParaMonte's required compilers by calling"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
+        echo >&2 "    -- ${BUILD_NAME} - FATAL:     ./install --fresh"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: If the error happens during the installation of ParaMonte prerequisites"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: it is possible that the current existing GCC compiler collection installed"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: on your system cannot compile the downloaded version of GCC that is required"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: for ParaMonte build. In such case, make sure you have a GCC compiler collection"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: version 7.1 or newer installed on your system, with an updated PATH environmental"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: variable, then reinstall ParaMonte."
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: If all ParaMonte installation attempts fail, please report this issue at"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
+        echo >&2 "    -- ${BUILD_NAME} - FATAL:     https://github.com/shahmoradi/paramonte/issues"
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
+        echo >&2 "    -- ${BUILD_NAME} - FATAL: or by contacting the ParaMonte authors directly (e.g., shahmoradi@utexas.edu)."
+        echo >&2
+        echo >&2 "    -- ${BUILD_NAME} - gracefully exiting."
+        echo >&2
+        exit 1
+    fi
+}
+
+printCopyFailMsg() {
     echo >&2
     echo >&2 "-- ParaMonteExample${LANG_NAME} - Copy action failed. Please resolve the error. Gracefully exiting..."
     echo >&2
@@ -128,9 +159,6 @@ fi
 # build examples
 #::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
[…]
