### 2020-10-11 92fc5fa
minor enhancements
