### 2020-10-11 b73c7a2
minor enhancements
