### 2020-10-11 8278c58
MatDRAM unix build script.
