+++ buildMatDRAM.sh
@@ -111,7 +111,7 @@ fi
 LANG_FILE_EXT="m"
 LANG_NAME="MATLAB"
 
-if ! [ -z ${LANG_NAME+x} ]; then
+if [ -z ${LANG_NAME+x} ]; then
     echo >&2 
     echo >&2 "-- ${BUILD_NAME}Example - Fatal Error: unrecognized or no language specified. exiting..."
     echo >&2 
@@ -128,7 +128,7 @@ EXAM_LIST="mvn"
 
 # set and make example directories
 
-MatDRAM_BLD_DIR="${ParaMonte_ROOT_DIR}/build/MatDRAM"
+MatDRAM_BLD_DIR="${ParaMonte_ROOT_DIR}/build/libparamonte_MatDRAM"
 MatDRAM_EXP_DIR="${MatDRAM_BLD_DIR}/example"
 
 ParaMonteInterface_SRC_DIR_CURRENT="${ParaMonteInterface_SRC_DIR}/${LANG_NAME}"
@@ -149,8 +149,8 @@ do
         verify $? "deletion of the old files"
         echo >&2 "-- ${BUILD_NAME} - regenerating the MatDRAM library ${EXAM_NAME} example directory: ${MatDRAM_BLD_DIR_CURRENT}"
     fi
-    mkdir "${MatDRAM_BLD_DIR_CURRENT}/"
-    verify $? "directory creation"
+    mkdir -p "${MatDRAM_BLD_DIR_CURRENT}/"
+    verify $? "ParaMonte::MatDRAM example directory creation"
 
     # The MatDRAM library example required files
 
@@ -172,13 +172,15 @@ do
 
     # The MatDRAM library license file
 
+    echo >&2
     echo >&2 "-- ${BUILD_NAME} - copying the MatDRAM library license file..."
-    echo >&2 "-- ${BUILD_NAME} - from: ${ParaMonte_ROOT_DIR}LICENSE.md"
+    echo >&2 "-- ${BUILD_NAME} - from: ${ParaMonte_ROOT_DIR}/LICENSE.md"
     echo >&2 "-- ${BUILD_NAME} -   to: ${MatDRAM_BLD_DIR_CURRENT}/LICENSE.md"
-    cp "${ParaMonte_ROOT_DIR}LICENSE.md" "${MatDRAM_BLD_DIR_CURRENT}/LICENSE.md" || printCopyFailMsg
+    cp "${ParaMonte_ROOT_DIR}/LICENSE.md" "${MatDRAM_BLD_DIR_CURRENT}/LICENSE.md" || printCopyFailMsg
 
     # The MatDRAM library README file
 
+    echo >&2
     echo >&2 "-- ${BUILD_NAME} - copying the MatDRAM library license file..."
     echo >&2 "-- ${BUILD_NAME} - from: ${ParaMonteInterface_SRC_DIR_CURRENT}/README.MatDRAM.md"
     echo >&2 "-- ${BUILD_NAME} -   to: ${MatDRAM_BLD_DIR_CURRENT}/README.MatDRAM.md"
@@ -186,7 +188,8 @@ do
 
     # The MatDRAM library interface files
 
-    echo >&2 "-- ${BUILD_NAME} - from: ${ParaMonteInterface_SRC_DIR_CURRENT}/MatDRAM"
+    echo >&2
+    echo >&2 "-- ${BUILD_NAME} - from: ${ParaMonteInterface_SRC_DIR_CURRENT}/libparamonte_MatDRAM"
[…]
