### 2020-10-11 253c285
minor enhancements
