### 2020-10-12 #2741 @whedon
[REVIEW]: ParaMonte: A high-performance serial/parallel Monte Carlo simulation library for C, C++, Fortran
**Submitting author:** @shahmoradi (<a href="http://orcid.org/0000-0002-9720-8937">Amir Shahmoradi</a>)
**Repository:** <a href="https://github.com/cdslaborg/paramonte" target ="_blank">https://github.com/cdslaborg/paramonte</a>
**Version:** v1.5.1

- [X] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)  **I don't have enough domain knowledge to conduct a study on the topic**

- [X] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.   Reviewer: There are build instructions, but
- [X] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?   **There are manual steps, but not automatic CI**
- [X] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support    **Since it's hosted on GitHub it's following the typical workflow, issues, PR**
