### 2020-10-12 4a25ee7
minor enhancements
