+++ src/ParaMonte/ParaMCMCRefinedChain_mod.f90
@@ -69,6 +69,10 @@ module ParaMCMCRefinedChain_mod
     end type RefinedChain_type
 
    type Method_type
+       logical :: isMax = .false.
+       logical :: isMed = .false.
+       logical :: isMin = .false.
+       logical :: isAvg = .false.
        logical :: isBatchMeans = .false.
        logical :: isCutoffAutoCorr = .false.
        logical :: isViaCompactChain = .true.
@@ -97,7 +101,8 @@ contains
         use, intrinsic :: iso_fortran_env, only: output_unit
         use ParaMonteChainFileContents_mod, only: ChainFileContents_type, NUM_DEF_COL
         use CrossCorr_mod, only: getBatchMeansIAC, getCumSumIAC, getMaxCumSumIAC
-        use String_mod, only: getLowerCase
+        use String_mod, only: getLowerCase, replaceStr
+        use Sort_mod, only: median
 
         implicit none
 
@@ -108,12 +113,12 @@ contains
         type(Err_type)              , intent(out)               :: Err
         character(*)                , intent(in), optional      :: sampleRefinementMethod
         integer(IK)                 , intent(in), optional      :: burninLoc, refinedChainSize, sampleRefinementCount
-        integer(IK)                                             :: burninLocDefault, i, countCompact
+        integer(IK)                                             :: burninLocDefault, i, countCompact, ndimPlusOne
         integer(IK)                                             :: maxRefinementCurrent, maxRefinementCount
-        real(RK)                                                :: integratedAutoCorrTime
+        real(RK)                                                :: integratedAutoCorrTime, ndimPlusOneInverse
         logical                                                 :: refinedChainSizeIsPresent
         logical                                                 :: maxRefinementCountIsReached
-        character(:)    , allocatable                           :: sampleRefinementMethodLowerCase
+        character(:)    , allocatable                           :: sampleRefinementMethodDefault, sampleRefinementMethodDefaultLowerCase, srmethod
         type(Count_type), allocatable                           :: DumCount(:)
         real(RK)        , allocatable                           :: DumIAC(:,:), DumVec(:)
         integer(IK)     , allocatable                           :: SampleWeight(:)
@@ -162,29 +167,46 @@ contains
         if (CFC%Count%verbose == 0_IK) CFC%Count%verbose = sum(CFC%Weight(1:CFC%Count%compact))
         RefinedChain%Count(RefinedChain%numRefinement)%verbose = CFC%Count%verbose
 
-        if (present(sampleRefinementMethod)) then
-            sampleRefinementMethodLowerCase = getLowerCase(sampleRefinementMethod)
-            if (index(sampleRefinementMethodLowerCase,getLowerCase(BATCH_MEANS_METHOD_NAME))>0) then
-                Method%isBatchMeans = .true.
-            elseif (index(sampleRefinementMethodLowerCase,getLowerCase(CUTOFF_AUTOCORR_METHOD_NAME))>0 .or. index(sampleRefinementMethodLowerCase,"cutoff")>0) then
-                Method%isCutoffAutoCorr = .true.
-            elseif (index(sampleRefinementMethodLowerCase,getLowerCase(MAX_CUMSUM_AUTOCORR_METHOD_NAME))>0 .or. index(sampleRefinementMethodLowerCase,"cumsum")>0) then
-                Method%isMaxCumSumAutoCorr = .true.
-            else
-                Err%occurred = .true.
[…]
