### 2020-10-12 9ec82ac
minor enhancements
