### 2020-10-12 adba03c
minor enhancements
