### 2020-10-13 b21cb7a
minor enhancements
