### 2020-10-15 3b1bd21
minor enhancements
