### 2020-10-15 ab35625
minor enhancements
