### 2020-10-15 4ed4a16
minor enhancements
