+++ src/interface/MATLAB/CHANGES.md
@@ -5,6 +5,12 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 ## **Version 2.x.x**  
 
+### Version  2.1.2 -- October 15, 2020  
+
+**Minor enhancements**  
+
++   The `checkForUpdate()` method of `paramonte` class now functions as expected.  
+
 ### Version  2.1.1 -- October 9, 2020  
 
 **Minor enhancements**  
