### 2020-10-15 4eea350
minor enhancements
