### 2020-10-15 b969126
minor enhancements
