+++ src/interface/MATLAB/paramonte/interface/@paramonte/paramonte.m
@@ -774,36 +774,57 @@ classdef paramonte %< dynamicprops
             %end
 
             if getVecLen(latestVersionString)
+
                 currentVersionTriplet = getVersionTriplet(currentVersionString);
                 latestVersionTriplet = getVersionTriplet(latestVersionString);
-                updateAvailable =   (latestVersionTriplet(1)  > currentVersionTriplet(1)) ...
-                                ||  (latestVersionTriplet(1) == currentVersionTriplet(1) && latestVersionTriplet(2)  > currentVersionTriplet(2)) ...
-                                ||  (latestVersionTriplet(1) == currentVersionTriplet(1) && latestVersionTriplet(2) == currentVersionTriplet(2) && latestVersionTriplet(3) > currentVersionTriplet(3));
-            else
-                updateAvailable = false;
-            end
 
-            if updateAvailable
-                latestVersionFile = self.website.github.release.latest.url + "/libparamonte_MATLAB";
-                if self.platform.isWin32
-                    latestVersionFile = latestVersionFile + ".zip";
+                if all(currentVersionTriplet==latestVersionTriplet)
+
+                    self.Err.msg = "It appears that you have the latest version of the ParaMonte::MATLAB library. " + newline;
+
                 else
-                    latestVersionFile = latestVersionFile + ".tar.gz";
+
+                    updateAvailable =   (latestVersionTriplet(1)  > currentVersionTriplet(1)) ...
+                                    ||  (latestVersionTriplet(1) == currentVersionTriplet(1) && latestVersionTriplet(2)  > currentVersionTriplet(2)) ...
+                                    ||  (latestVersionTriplet(1) == currentVersionTriplet(1) && latestVersionTriplet(2) == currentVersionTriplet(2) && latestVersionTriplet(3) > currentVersionTriplet(3));
+
+                    if updateAvailable
+                        latestVersionFile = self.website.github.release.latest.url + "/libparamonte_MATLAB";
+                        if self.platform.isWin32
+                            latestVersionFile = latestVersionFile + ".zip";
+                        else
+                            latestVersionFile = latestVersionFile + ".tar.gz";
+                        end
+                        self.Err.msg    = "A newer version (" + latestVersionString + ") of the ParaMonte library appears " + newline ...
+                                        + "to be available on the GitHub repository or, is in preparation " + newline ...
+                                        + "for release. The currently-installed version on your system is: " + currentVersionString + newline ...
+                                        + "You can download the latest version of the ParaMonte MATLAB library from " + newline ...
+                                        + newline ...
+                                        + "    " + href(latestVersionFile) + newline ...
+                                        + newline ...
+                                        ;
+                    else
+                        self.Err.msg    = "Looks like you have a version of the ParaMonte::MATLAB library (" + currentVersionString + ") " + newline ...
+                                        + "that is newer than the most recent release version (" + latestVersionString + "). Good for you!" + newline;
+                    end
+
                 end
-                self.Err.msg    = "A newer version (" + latestVersionString + ") of the ParaMonte library appears " + newline ...
-                                + "to be available on the GitHub repository or, is in preparation " + newline ...
-                                + "for release. The currently-installed version on your system is: " + currentVersionString + newline ...
-                                + "You can download the latest version of the ParaMonte MATLAB library from " + newline ...
-                                + newline ...
-                                + "    " + href(latestVersionFile) + newline ...
-                                + newline ...
-                                ;
+
             else
-                self.Err.msg    = "It appears that you have the latest version of the ParaMonte library. " + newline;
+
+                self.Err.msg    = "Failed to capture the latest version of ParaMonte::MATLAB library from " + newline ...
+                                + "the project's repository. If you believe this behavior is unexpected, " + newline ...
[…]
