### 2020-10-15 d2bc1e9
minor enhancements
