### 2020-10-16 c4e89af
minor enhancements
