+++ auxil/btar.sh
@@ -45,6 +45,26 @@
 #
 #   ./auxil/btar.sh --dir ./bin/
 #
+
+####################################################################################################################################
+# auxil
+####################################################################################################################################
+
+reportBadValue()
+{
+    echo >&2 ""
+    echo >&2 "-- ParaMonte - FATAL: The requested input value $2 specified with "
+    echo >&2 "-- ParaMonte - FATAL: the input flag $1 is not supported."
+    if ! [ -z ${3+x} ]; then
+    echo >&2 "-- ParaMonte - FATAL: $3"
+    fi
+    echo >&2 ""
+    echo >&2 "-- ParaMonte - gracefully exiting."
+    echo >&2 ""
+    echo >&2 ""
+    exit 1
+}
+
 ####################################################################################################################################
 # parse arguments
 ####################################################################################################################################
@@ -52,10 +72,15 @@
 AUXIL_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 TARGET_DIR="${AUXIL_DIR}/../bin/"
 
+unset LANG_LIST # must be all lower case
+
 while [ "$1" != "" ]; do
     case $1 in
         -d | --dir )    shift
                         TARGET_DIR=$1
+                        ;;
+        -L | --lang )   shift
+                        LANG_LIST="$1"
                         ;;
         -h | --help )   #usage
                         echo >&2 "-- ParaMonte - Sadly, there is no help here, because you should not be here if you do not know why you are here."
@@ -68,6 +93,41 @@ while [ "$1" != "" ]; do
     shift
 done
 
+if ! ([ -z ${LANG_LIST+x} ] || [ "${LANG_LIST}" = "all" ]); then
+    ALL_ENABLED=false
+    for LANG in "${LANG_LIST}"; do
+        if  [[ $LANG == [cC] ]]; then C_ENABLED=true; else C_ENABLED=false; fi
+        if  [[ $LANG == "c++" || $LANG == "C++" ]]; then CPP_ENABLED=true; else CPP_ENABLED=false; fi
+        if  [[ $LANG == [fF][oO][rR][tT][rR][aA][nN] ]]; then Fortran_ENABLED=true; else Fortran_ENABLED=false; fi
+        if  [[ $LANG == [mM][aA][tT][lL][aA][bB] ]]; then MATLAB_ENABLED=true; else MATLAB_ENABLED=false; fi
+        if  [[ $LANG == [pP][yY][tT][hH][oO][nN] ]]; then Python_ENABLED=true; else Python_ENABLED=false; fi
+
+        #echo ${LANG}
+        #echo ${LANG_LIST}
+        #echo ${C_ENABLED}
+        #echo ${CPP_ENABLED}
+        #echo ${Fortran_ENABLED}
+        #echo ${MATLAB_ENABLED} 
+        #echo ${Python_ENABLED} 
+
+        if [ ${C_ENABLED} = "false" ] \
+        && [ ${CPP_ENABLED} = "false" ] \
+        && [ ${Fortran_ENABLED} = "false" ] \
+        && [ ${MATLAB_ENABLED} = "false" ] \
+        && [ ${Python_ENABLED} = "false" ]; then
+            reportBadValue "-L or --lang" $LANG
+        fi
+    done
+    if [ "${LANG_LIST}" = "" ]; then
+        unset LANG_LIST
+    #else
+    #    LANG_LIST="$(getLowerCase $LANG_LIST)"
+    fi
+else
+    ALL_ENABLED=true
+fi
+
+
 if [ -d "${TARGET_DIR}" ]; then
     echo >&2
[…]
