### 2020-10-18 3cddffe
minor enhancements
