### 2020-10-18 24b00be
minor enhancements
