### 2020-10-18 740d17c
minor enhancements
