### 2020-10-19 c4785eb
minor enhancements
