### 2020-10-19 603fd5e
minor enhancements
