### 2020-10-20 e4a3520
minor enhancements
