### 2020-10-20 15e90c5
minor enhancements
