### 2020-10-21 c2e0d36
minor enhancements
