### 2020-10-21 c9f4b4d
minor enhancements
