+++ src/ParaMonte/ParaDRAM_mod@Input_smod.f90
@@ -40,9 +40,9 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-submodule (ParaDRAM_mod) Setup_smod
+submodule (ParaDRAM_mod) Input_smod
 #define PARADRAM ParaDRAM
-#include "ParaDXXX_mod@Setup_smod.inc.f90"
+#include "ParaDXXX_mod@Input_smod.inc.f90"
 #undef PARADRAM
-end submodule Setup_smod
+end submodule Input_smod
 
