+++ src/ParaMonte/ParaDISE_mod@Input_smod.f90
@@ -40,9 +40,9 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-submodule (ParaDISE_mod) Kernel_smod
+submodule (ParaDISE_mod) Input_smod
 #define PARADISE ParaDISE
-#include "ParaDXXX_mod@Kernel_smod.inc.f90"
+#include "ParaDXXX_mod@Input_smod.inc.f90"
 #undef PARADISE
-end submodule Kernel_smod
+end submodule Input_smod
 
