+++ src/ParaMonte/ParaDISEProposalAbstract_mod.f90
@@ -42,7 +42,7 @@
 
 module ParaDISEProposalAbstract_mod
 #define PARADISE ParaDISE 
-#include "ParaDRAMProposalAbstract_mod.inc.f90"
+#include "ParaDXXXProposalAbstract_mod.inc.f90"
 #undef ParaDISE
 end module ParaDISEProposalAbstract_mod
 
