+++ src/ParaMonte/ParaDISE_mod@Setup_smod.f90
@@ -42,7 +42,7 @@
 
 submodule (ParaDISE_mod) Setup_smod
 #define PARADISE ParaDISE
-#include "ParaDRAM_mod@Setup_smod.inc.f90"
+#include "ParaDXXX_mod@Setup_smod.inc.f90"
 #undef PARADISE
 end submodule Setup_smod
 
