### 2020-10-29 1a1d769
minor enhancements
