### 2020-10-29 b884a98
minor enhancements
