### 2020-10-29 d6d65d3
minor enhancements
