### 2020-10-29 9d22e7a
minor enhancements
