### 2020-10-29 c94a90e
minor enhancements
