### 2020-10-29 f896d53
minor enhancements
