### 2020-10-29 f0630e8
minor enhancements
