### 2020-10-29 086e3a3
minor enhancements
