+++ src/interface/MATLAB/README.md
@@ -190,6 +190,7 @@ Example usage instructions
         ```bash  
         sudo matlab
         ```  
+        **The superuser (`sudo`) permission is only required for the installation of the ParaMonte library and not the MPI libraries**.  
         If `matlab` is not recognized on your command line as an application, seek help from 
         [this ParaMonte documentation page](https://www.cdslab.org/paramonte/notes/troubleshooting/bash-matlab-command-not-found/). 
         Once the MATLAB interactive environment opens, navigate to the root folder of the ParaMonte library (where the LICENSE file exists) 
@@ -200,7 +201,7 @@ Example usage instructions
         pm.verify(); % verify the integrity of the ParaMonte library on your system.
         ```  
         If needed, follow any extra instructions provided by the library on your MATLAB command prompt. 
-        If you do not intend to run simulations in parallel say NO (`n`) to any MPI library installations. 
+        **If you do not intend to run simulations in parallel say NO (`n`) to any MPI library installations**. 
         Before beginning to use the ParaMonte library, we strongly recommend that you close your current MATLAB 
         session and the Bash terminal from which you initiated the MATLAB session in **sudo** mode, **entirely**. 
         Then follow the instructions below on how to run your ParaMonte-enabled simulations.  
@@ -216,7 +217,6 @@ Example usage instructions
         pm.verify(); % verify the integrity of the ParaMonte library on your system.
         ```  
         Then, answer YES (`y`) to any permission requests by the ParaMonte library to install the MPI libraries on your system.  
-        **The superuser (`sudo`) permission is only required for the installation of the ParaMonte library and not the MPI libraries**.  
 
 +   **Running the ParaMonte simulations**  
     For complete up-to-date detailed instructions, visit: https://www.cdslab.org/paramonte/notes/run/matlab/
