### 2020-10-29 cf62bc4
minor enhancements
