### 2020-10-29 5411b50
minor enhancements
