### 2020-10-29 8f0c50d
minor enhancements
