+++ auxil/genReadme.py
@@ -692,7 +692,7 @@ readme["examples"]["matlab"] = readme["examples"]["title"] + """
         pm.verify(); % verify the integrity of the ParaMonte library on your system.
         ```  
         If needed, follow any extra instructions provided by the library on your MATLAB command prompt.  
-        **If you do not intend to run simulations in parallel say NO (`n`) to any MPI library installations**.  
+        **If you do not intend to run simulations in parallel say NO (`n`) to any MPI library installation requests via ParaMonte**.  
         Before beginning to use the ParaMonte library, we strongly recommend that you close your current MATLAB 
         session and the Bash terminal from which you initiated the MATLAB session in **sudo** mode, **entirely**. 
         Then follow the instructions below on how to run your ParaMonte-enabled simulations.  
