### 2020-10-29 48ffd64
minor enhancements
