### 2020-10-29 338acc6
minor enhancements
