### 2020-10-29 157a7c7
minor enhancements
