### 2020-10-30 95e28c3
minor enhancements
