+++ src/ParaMonte/ParaDXXX_mod@Setup_smod.inc.f90
@@ -1734,238 +1734,6 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-!
-!    module subroutine getSpecFromInputFile(self,nd)
-!#if defined DLL_ENABLED && !defined CFI_ENABLED
-!        !DEC$ ATTRIBUTES DLLEXPORT :: getSpecFromInputFile
-!#endif
-!        use Constants_mod, only: IK, RK
-!        use String_mod, only: num2str
-!
-!        ! ParaMonte namelist variables
-!        use SpecBase_SampleSize_mod                         , only: sampleSize
-!        use SpecBase_RandomSeed_mod                         , only: randomSeed
-!        use SpecBase_Description_mod                        , only: description
-!        use SpecBase_OutputFileName_mod                     , only: outputFileName
-!        use SpecBase_OutputDelimiter_mod                    , only: outputDelimiter
-!        use SpecBase_ChainFileFormat_mod                    , only: chainFileFormat
-!        use SpecBase_VariableNameList_mod                   , only: variableNameList
-!        use SpecBase_RestartFileFormat_mod                  , only: restartFileFormat
-!        use SpecBase_OutputColumnWidth_mod                  , only: outputColumnWidth
-!        use SpecBase_OverwriteRequested_mod                 , only: overwriteRequested
-!        use SpecBase_OutputRealPrecision_mod                , only: outputRealPrecision
-!        use SpecBase_SilentModeRequested_mod                , only: silentModeRequested
-!        use SpecBase_DomainLowerLimitVec_mod                , only: domainLowerLimitVec
-!        use SpecBase_DomainUpperLimitVec_mod                , only: domainUpperLimitVec
-!        use SpecBase_ParallelizationModel_mod               , only: parallelizationModel
-!        use SpecBase_InputFileHasPriority_mod               , only: inputFileHasPriority
-!        use SpecBase_ProgressReportPeriod_mod               , only: progressReportPeriod
-!        use SpecBase_targetAcceptanceRate_mod               , only: targetAcceptanceRate
-!        use SpecBase_MpiFinalizeRequested_mod               , only: mpiFinalizeRequested
-!        use SpecBase_MaxNumDomainCheckToWarn_mod            , only: maxNumDomainCheckToWarn
-!        use SpecBase_MaxNumDomainCheckToStop_mod            , only: maxNumDomainCheckToStop
-!        use SpecBase_SystemInfoFilePath_mod                 , only: systemInfoFilePath
-!        use SpecBase_InterfaceType_mod                      , only: interfaceType
-!
-!        ! ParaMCMC namelist variables
-!        use SpecMCMC_ChainSize_mod                          , only: chainSize
-!        use SpecMCMC_ScaleFactor_mod                        , only: scaleFactor
-!        use SpecMCMC_StartPointVec_mod                      , only: startPointVec
-!        use SpecMCMC_ProposalModel_mod                      , only: proposalModel
-!        use SpecMCMC_proposalStartCovMat_mod                , only: proposalStartCovMat
-!        use SpecMCMC_proposalStartCorMat_mod                , only: proposalStartCorMat
[…]
