+++ example/build.sh
@@ -145,11 +145,13 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"_fortran_".* ]]; then
     EXAMPLE_LANGUAGE=Fortran
     if [[ "$PMLIB_FULL_NAME" =~ .*"_intel_".* ]]; then
         PM_COMPILER_SUITE=intel
-        COMPILER_LIST=ifort
+        #COMPILER_LIST="ifort gfortran"
+        declare -a COMPILER_LIST=("ifort" "gfortran")
     fi
     if [[ "$PMLIB_FULL_NAME" =~ .*"_gnu_".* ]]; then
         PM_COMPILER_SUITE=gnu
-        COMPILER_LIST=gfortran
+        #COMPILER_LIST="gfortran ifort"
+        declare -a COMPILER_LIST=("gfortran" "ifort")
     fi
 fi
 
@@ -158,11 +160,13 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"_c_".* ]]; then
     EXAMPLE_LANGUAGE=C
     if [[ "$PMLIB_FULL_NAME" =~ .*"_intel_".* ]]; then
         PM_COMPILER_SUITE=intel
-        COMPILER_LIST=icc
+        #COMPILER_LIST="icc gcc"
+        declare -a COMPILER_LIST=("icc" "gcc")
     fi
     if [[ "$PMLIB_FULL_NAME" =~ .*"_gnu_".* ]]; then
         PM_COMPILER_SUITE=gnu
-        COMPILER_LIST=gcc
+        #COMPILER_LIST="gcc icc"
+        declare -a COMPILER_LIST=("gcc" "icc")
     fi
 fi
 
@@ -171,11 +175,13 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"_cpp_".* ]]; then
     EXAMPLE_LANGUAGE=C++
     if [[ "$PMLIB_FULL_NAME" =~ .*"_intel_".* ]]; then
         PM_COMPILER_SUITE=intel
-        COMPILER_LIST=icpc
+        #COMPILER_LIST="icpc g++"
+        declare -a COMPILER_LIST=("icpc" "g++")
     fi
     if [[ "$PMLIB_FULL_NAME" =~ .*"_gnu_".* ]]; then
         PM_COMPILER_SUITE=gnu
-        COMPILER_LIST=g++
+        #COMPILER_LIST="g++ icpc"
+        declare -a COMPILER_LIST=("g++" "icpc")
     fi
 fi
 
@@ -244,17 +250,17 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"mpi".* ]]; then
 fi
 echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - MPI_ENABLED: ${MPI_ENABLED}"
 
-if [ "${MPI_ENABLED}" = "true" ]; then
-    if [ "${EXAMPLE_LANGUAGE}" = "Fortran" ]; then
-        COMPILER_LIST="mpiifort mpifort"
-    fi
-    if [ "${EXAMPLE_LANGUAGE}" = "C" ]; then
-        COMPILER_LIST="mpiicc mpicc"
-    fi
-    if [ "${EXAMPLE_LANGUAGE}" = "C++" ]; then
-        COMPILER_LIST="mpiicpc mpicxx mpic++ mpicc"
-    fi
-fi
+#if [ "${MPI_ENABLED}" = "true" ]; then
+#    if [ "${EXAMPLE_LANGUAGE}" = "Fortran" ]; then
+#        COMPILER_LIST="mpiifort mpifort"
+#    fi
+#    if [ "${EXAMPLE_LANGUAGE}" = "C" ]; then
+#        COMPILER_LIST="mpiicc mpicc"
+#    fi
[…]
