+++ src/ParaMonte/ParaMonte@runParaDRAM.f90
@@ -46,7 +46,7 @@
 ! The procedural C interface to ParaDRAM
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-#if (defined MATLAB_ENABLED || defined PYTHON_ENABLED)
+#if (defined MATLAB_ENABLED || defined PYTHON_ENABLED || defined R_ENABLED)
 function &
 #else
 subroutine &
@@ -74,7 +74,7 @@ subroutine &
     character(:), allocatable                               :: inputFileStr
     type(ParaDRAM_type)                                     :: self
     integer                                                 :: i
-#if (defined MATLAB_ENABLED || defined PYTHON_ENABLED)
+#if (defined MATLAB_ENABLED || defined PYTHON_ENABLED || defined R_ENABLED)
     integer(IK)                                             :: runParaDRAM
     runParaDRAM = 0_IK
 #endif
@@ -110,7 +110,7 @@ subroutine &
     end if
     nullify(getLogFunc)
 
-#if (defined MATLAB_ENABLED || defined PYTHON_ENABLED)
+#if (defined MATLAB_ENABLED || defined PYTHON_ENABLED || defined R_ENABLED)
     if (self%Err%occurred) runParaDRAM = -1_IK
 end function runParaDRAM
 #else
