### 2020-11-03 #2741 @williamfgc
I have a few initial recommendations upon my review. I hope the authors find this helpful to transition ParaMonte from academic code to production quality software:

1. I couldn't find the use of automated test frameworks for unit and functional tests. I'd recommend using `ctest` as the project has CMake support and create automated tests. It seems there are a set of tests under [src/test](https://github.com/cdslaborg/paramonte/blob/master/src/test/CMakeLists.txt), but they are not executed. Also a [Fortran unit testing framework](http://fortranwiki.org/fortran/show/Unit+testing+frameworks) could be a good alternative.
2. CI: this project is missing continuous integration. I find it difficult to accept it without comprehensive testing when changes are made to the repo. CI builds trusts among users.
3. Contributing: it's not clear the contribution policy. This is tied to 3. and the lack of CI infrastructure (either GitHub Actions, CircleCI, Travis, etc.).
4. Installation: add support for CMake installation. While there is a large effort to support different platform via scripts, it would be beneficial to reuse CMake portable capabilities to build and install the library portably. This could be added to their CI.
5. Code coverage: once CI is installed please add code coverage (e.g. Coverity, Codecov)
6. Code documentation: it's hard to find API level documentation (e.g. doxygen) so any person new to the project can pick up the author's intention.
7. The paper focuses too much on previous work publications, rather than exposing the merits of the software infrastructure and (most importantly) impact. I'd focus it on the community impact: use cases, important integration with other projects, etc.
8.  Distribution and packaging (optional): it would be good to have the option for standard distribution packages (conda, spack). Install scripts have to be constantly updated and maintained. Without proper CI, bugs could pile up pretty quickly.

I'd be happy to expand on the above if required, but I believe they are necessary to meet minimal software quality standards.
