### 2020-11-03 29c881c
GNU Application builds on Windows added.
The ParaMonte library and the build scripts have been modified
to enable compilation and linking of GNU codebase with the ParaMonte
libraries prebuilt via the Intel Compilers on Windows.
