+++ example/build.sh
@@ -114,6 +114,39 @@ while [ "$1" != "" ]; do
 done
 
 ####################################################################################################################################
+# get the platform name. required to define __WIN64__ for binding GFortran applications to Intel-prebuilt ParaMonte libraries on Windows.
+####################################################################################################################################
+
+UNAME_PLATFORM="$(uname -s)"
+case "${UNAME_PLATFORM}" in
+    Linux*)     PLATFORM=linux;;
+    Darwin*)    PLATFORM=darwin;;
+    CYGWIN*)    PLATFORM=cygwin;;
+    MINGW*)     PLATFORM=mingw;;
+    *)          PLATFORM="unknown:${UNAME_PLATFORM}"
+esac
+if [[ "$PLATFORM" =~ .*"unknown".* ]]; then
+    echo >&2
+    echo >&2 "-- ${BUILD_NAME} - FATAL: build failed. unrecognized platform - ${PLATFORM}"
+    echo >&2 "-- ${BUILD_NAME} - supported platforms include: Linux, Darwin, CYGWIN, MINGW"
+    echo >&2 "-- ${BUILD_NAME} - ParaMonte build has been only tested on Linux and Darwin platforms."
+    echo >&2
+    echo >&2 "-- ${BUILD_NAME} - gracefully exiting."
+    echo >&2
+    exit 1
+else
+    export PLATFORM
+fi
+
+if [[ "${UNAME_PLATFORM}" =~ .*"Darwin".* ]]; then
+    isMacOS=true
+    OSNAME="macOS"
+else
+    isMacOS=false
+    OSNAME="Linux"
+fi
+
+####################################################################################################################################
 # build ParaMonteExample objects and executable
 ####################################################################################################################################
 
@@ -125,7 +158,9 @@ if [ -z ${FOR_COARRAY_NUM_IMAGES+x} ]; then
 fi
 export FOR_COARRAY_NUM_IMAGES
 
-PMLIB_FULL_PATH="$(ls -d ${FILE_DIR}/*libparamonte_* | sort -V | tail -n1)"
+PMLIB_EXTENSION=""
+if [ "$PLATFORM" = "mingw" ] || [ "$PLATFORM" = "cygwin" ]; then PMLIB_EXTENSION=".dll"; fi
+PMLIB_FULL_PATH="$(ls -d ${FILE_DIR}/*libparamonte_*${PMLIB_EXTENSION} | sort -V | tail -n1)"
 PMLIB_FULL_NAME=${PMLIB_FULL_PATH##*/}
 PMLIB_BASE_NAME=${PMLIB_FULL_NAME%.*}
 
@@ -153,6 +188,7 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"_fortran_".* ]]; then
         #COMPILER_LIST="gfortran ifort"
         declare -a COMPILER_LIST=("gfortran" "ifort")
     fi
+    if [ "$PLATFORM" = "mingw" ] || [ "$PLATFORM" = "cygwin" ]; then declare -a COMPILER_LIST=("gfortran"); fi
 fi
 
 if [[ "$PMLIB_FULL_NAME" =~ .*"_c_".* ]]; then
[…]
