### 2020-11-03 fd77703
minor enhancements
