### 2020-11-03 a7a827f
minor enhancements
