+++ example/build.sh
@@ -445,7 +445,7 @@ do
 
     echo >&2
     echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - linking ParaMonte example with ${LINKER}"
-    echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} paramonte.o logfunc.o main.o ${PMLIB_FULL_NAME} -o ${PM_EXAM_EXE_NAME}"
+    echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} ${SRC_FILES/$FILE_EXT/.o} ${PMLIB_FULL_NAME} -o ${PM_EXAM_EXE_NAME}"
 
     ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} paramonte.o logfunc.o main.o ${PMLIB_FULL_NAME} -o ${PM_EXAM_EXE_NAME}
 
