### 2020-11-03 de7add8
minor enhancements
