### 2020-11-03 bd2ca57
minor enhancements
