### 2020-11-03 b75d3a5
minor enhancements
