### 2020-11-03 95caf3c
minor enhancements
