### 2020-11-08 0fbbe12
kernel routines now fully documented.
