### 2020-11-08 a2e1d71
minor enhancements
