### 2020-11-08 ba52ac8
minor enhancements
