### 2020-11-08 512a8b9
minor enhancements
