### 2020-11-09 a211123
minor enhancements
