+++ auxil/genReadme.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 ####################################################################################################################################
 ####################################################################################################################################
 ####
@@ -84,6 +85,15 @@ banner = """
 <br><br>
 <a href="https://twitter.com/intent/tweet?text=ParaMonte%20-%20Plain%20Powerfull%20Parallel%20Monte%20Carlo%20Library:&url=https%3A%2F%2Fgithub.com%2Fcdslaborg%2Fparamonte" target="_blank"><img src="https://img.shields.io/twitter/url?style=social&url=https%3A%2F%2Fgithub.com%2Fcdslaborg%2Fparamonte" alt="Twitter" /></a> 
 </div>
+  
+  
+[Overview](#paramonte-plain-powerful-parallel-monte-carlo-library) | 
+[Installation](#installation) | 
+[Dependencies](#dependencies) | 
+[Parallelism](#parallelism) | 
+[License](#license) | 
+[Authors](#authors-and-contributors) | 
+[Examples](#example-usage-instructions)  
   
   
 ParaMonte: Plain Powerful Parallel Monte Carlo Library
