+++ src/interface/Python/README.md
@@ -24,9 +24,7 @@
 <a href="https://joss.theoj.org/papers/f964b6e22c71515c310fbe3843ad4513"><img src="https://joss.theoj.org/papers/f964b6e22c71515c310fbe3843ad4513/status.svg"></a>
 <br><br>
 <a href="https://twitter.com/intent/tweet?text=ParaMonte%20-%20Plain%20Powerfull%20Parallel%20Monte%20Carlo%20Library:&url=https%3A%2F%2Fgithub.com%2Fcdslaborg%2Fparamonte" target="_blank"><img src="https://img.shields.io/twitter/url?style=social&url=https%3A%2F%2Fgithub.com%2Fcdslaborg%2Fparamonte" alt="Twitter" /></a> 
-</div>
-  
-  
+<br><br>
 [Overview](#paramonte-plain-powerful-parallel-monte-carlo-library) | 
 [Installation](#installation) | 
 [Dependencies](#dependencies) | 
@@ -34,6 +32,7 @@
 [License](#license) | 
 [Authors](#authors-and-contributors) | 
 [Examples](#example-usage-instructions)  
+</div>
   
   
 ParaMonte: Plain Powerful Parallel Monte Carlo Library
