### 2020-11-09 6e4a60d
minor enhancements
