### 2020-11-09 c5af119
minor enhancements
