+++ src/interface/Fortran/README.md
@@ -29,8 +29,9 @@
 <a href="#installation">Installation</a> | 
 <a href="#dependencies">Dependencies</a> | 
 <a href="#parallelism">Parallelism</a> | 
-<a href="#authors-and-contributors">Authors</a> | 
+<a href="#citing-paramonte">Acknowledgments</a> | 
 <a href="#license">License</a> | 
+<a href="#authors-and-contributors">Authors</a> | 
 <a href="#example-usage-instructions">Examples</a> 
 </div>
   
