### 2020-11-09 e27396b
minor enhancements
