### 2020-11-09 eedaf09
minor enhancements
