### 2020-11-09 1e52d9b
minor enhancements
