### 2020-11-09 10b2645
minor enhancements
