### 2020-11-09 c0ca0db
minor enhancements
