### 2020-11-09 665903a
minor enhancements
