+++ src/interface/C/README.md
@@ -83,10 +83,10 @@ Alternatively, you can build the library from the source in the [GitHub reposito
 The ParaMonte library installation/build process is fully automated for all of the supported programming languages. 
 Currently, the following compiler suites are supported for builds from source:  
   
-| Compiler Suite | Linux | macOS | Windows (64bit) |  
-|---------------:|:-----:|:-----:|:---------------:|  
-| GNU > 7.1      |&check;|&check;| &cross;         |  
-| Intel > 18.0.0 |&check;|&check;| &check;         |  
+| Compiler Suite                    | Linux | macOS | Windows (64bit) |  
+|----------------------------------:|:-----:|:-----:|:---------------:|  
+| GNU Compiler Collection > 7.1     |&check;|&check;| &cross;         |  
+| Intel Parallel Studio > 18.0.0    |&check;|&check;| &check;         |  
 
 For more information and quick-start in the programming language of your choice, visit the [ParaMonte library homepage](https://www.cdslab.org/paramonte).  
 
