### 2020-11-09 12207ff
minor enhancements
