### 2020-11-09 fa88443
minor enhancements
