+++ developer-scripts/travis/before_install.linux.sh
@@ -1,25 +1,0 @@
-#!/usr/bin/env bash
-
-set -o errexit
-set -o verbose
-set -o pipefail
-set -o nounset
-set -o errtrace
-
-__file=developer-scripts/travis/install.linux.sh
-
-# Error tracing
-# requires `set -o errtrace`
-__caf_err_report() {
-    error_code=${?}
-    echo "Error (code=${error_code}) in ${__file} in function ${1} on line ${2}." >&2
-    false
-    return ${error_code}
-}
-# Always provide an error backtrace
-trap '__caf_err_report "${FUNCNAME:-.}" ${LINENO}' ERR
-
-[[ -d "${CACHE}/bin" ]] || mkdir -p "${CACHE}/bin"
-${FC} --version
-${CC} --version
-${CXX} --version
