+++ CHANGES.md
@@ -48,8 +48,9 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
   
 | Compiler Suite                    | Linux | macOS | Windows (64bit) |  
 |----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.1     |&check;|&check;| &cross;         |  
-| Intel Parallel Studio > 18.0.0    |&check;|&check;| &check;         |  
+| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
+| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅    | ❌    | ❌              |  
 
 **Compiler / MPI library used for this binary release**  
 
@@ -91,11 +92,12 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 +   minor typo fixes
 
 **Compiler support**  
-  
+
 | Compiler Suite                    | Linux | macOS | Windows (64bit) |  
 |----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.1     |&check;|&check;| &cross;         |  
-| Intel Parallel Studio > 18.0.0    |&check;|&check;| &check;         |  
+| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
+| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅    | ❌    | ❌              |  
 
 **Compiler / MPI library used for this binary release**  
 
@@ -123,11 +125,12 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 +   The fully-deterministic restart functionality is now functional also when chainFileFormat="verbose" in ParaDRAM simulations.
 
 **Compiler support**  
-  
+
 | Compiler Suite                    | Linux | macOS | Windows (64bit) |  
 |----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.1     |&check;|&check;| &cross;         |  
-| Intel Parallel Studio > 18.0.0    |&check;|&check;| &check;         |  
+| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
+| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅    | ❌    | ❌              |  
 
 **Compiler / MPI library used for this binary release**  
 
@@ -148,11 +151,12 @@ This is the first public release of the ParaMonte library.
 +   ParaMonte simulation-output visualization via the ParaMonte Python interface.  
 
 **Compiler support**  
-  
+
 | Compiler Suite                    | Linux | macOS | Windows (64bit) |  
 |----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.1     |&check;|&check;| &cross;         |  
-| Intel Parallel Studio > 18.0.0    |&check;|&check;| &check;         |  
+| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
+| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅    | ❌    | ❌              |  
 
 **Compiler / MPI library used for this binary release**  
 
