+++ CHANGES.md
@@ -46,11 +46,11 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 **Compiler support**  
   
-| Compiler Suite                    | Linux | macOS | Windows (64bit) |  
-|----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
-| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
-| Microsoft C/C++ Compiler > 16.0.0 | ❌    | ❌    | ✅              | 
+| Compiler Suite                    | Windows (64bit) | Linux | macOS |  
+|----------------------------------:|:---------------:|:-----:|:-----:|  
+| GNU Compiler Collection > 7.0     | ✅              | ✅    | ✅    |  
+| Intel Parallel Studio > 18.0.0    | ✅              | ✅    | ✅    |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅              | ❌    | ❌    |  
 
 **Compiler / MPI library used for this binary release**  
 
@@ -93,11 +93,11 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 **Compiler support**  
 
-| Compiler Suite                    | Linux | macOS | Windows (64bit) |  
-|----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
-| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
-| Microsoft C/C++ Compiler > 16.0.0 | ❌    | ❌    | ✅              | 
+| Compiler Suite                    | Windows (64bit) | Linux | macOS |  
+|----------------------------------:|:---------------:|:-----:|:-----:|  
+| GNU Compiler Collection > 7.0     | ✅              | ✅    | ✅    |  
+| Intel Parallel Studio > 18.0.0    | ✅              | ✅    | ✅    |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅              | ❌    | ❌    |  
 
 **Compiler / MPI library used for this binary release**  
 
@@ -126,11 +126,11 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 **Compiler support**  
 
-| Compiler Suite                    | Linux | macOS | Windows (64bit) |  
-|----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
-| Intel Parallel Studio > 18.0.0    | ✅    | ✅    | ✅              |  
-| Microsoft C/C++ Compiler > 16.0.0 | ❌    | ❌    | ✅              |  
+| Compiler Suite                    | Windows (64bit) | Linux | macOS |  
+|----------------------------------:|:---------------:|:-----:|:-----:|  
+| GNU Compiler Collection > 7.0     | ✅              | ✅    | ✅    |  
+| Intel Parallel Studio > 18.0.0    | ✅              | ✅    | ✅    |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅              | ❌    | ❌    |  
 
 **Compiler / MPI library used for this binary release**  
 
@@ -152,11 +152,11 @@ This is the first public release of the ParaMonte library.
 
 **Compiler support**  
 
-| Compiler Suite                    | Linux | macOS | Windows (64bit) |  
-|----------------------------------:|:-----:|:-----:|:---------------:|  
-| GNU Compiler Collection > 7.0     | ✅    | ✅    | ✅              |  
[…]
