### 2020-11-09 98a80ff
minor enhancements
