+++ .travis.yml
@@ -48,7 +48,7 @@ matrix:
       sudo: false
       dist: xenial
       env:
-        - GCC=8
+        - GCC=9
       cache:
         apt: true
         directories:
@@ -61,26 +61,23 @@ matrix:
               key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
           packages: &gcc8pkgs
             - cmake
-            - gcc-8
-            - g++-8
-            - gfortran-8
+            - gcc-9
+            - g++-9
+            - gfortran-9
     -
       <<: *ubuntu
       env:
-        - GCC=7
+        - GCC=9
       addons:
         apt:
           sources: *linuxsrcs
           packages: &gcc7pkgs
             - cmake
-            - gcc-7
-            - g++-7
-            - gfortran-7
+            - gcc-9
+            - g++-9
+            - gfortran-9
     -
       <<: *ubuntu
-      env:
-        - BUILD_TYPE="InstallScript"
-          GCC=8
       addons:
         apt:
           sources: *linuxsrcs
@@ -103,7 +100,7 @@ matrix:
 
 install:
   - pwd
-  - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none
+  - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none --yes-to-all
 
 #after_success:
 #  - bash <(curl -s https://codecov.io/bash)
