+++ .travis.yml
@@ -43,52 +43,52 @@ matrix:
             - cmake
             - openmpi
           update: true
-    - &ubuntu
-      os: linux
-      sudo: false
-      dist: xenial
-      cache:
-        apt: true
-        directories:
-          - "$CACHE"
-      addons:
-        apt:
-          sources: &linuxsrcs
-            - ubuntu-toolchain-r-test
-            - sourceline: 'deb https://apt.kitware.com/ubuntu/ xenial main'
-              key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
-          packages: &gcc8pkgs
-            - cmake
-            - gcc
-            - g++
-            - gfortran
-    -
-      <<: *ubuntu
-      env:
-        - GCC=9
-      addons:
-        apt:
-          sources: *linuxsrcs
-          packages: &gcc7pkgs
-            - cmake
-            - gcc
-            - g++
-            - gfortran
-    -
-      <<: *ubuntu
-      addons:
-        apt:
-          sources: *linuxsrcs
-          packages:
-            - gcc
-            - g++
-            - gfortran
-    - &linux_allowed_failure
-      <<: *ubuntu
-      addons:
-        apt:
-          sources: *linuxsrcs
-          packages: *gcc8pkgs
+#    - &ubuntu
+#      os: linux
+#      sudo: false
+#      dist: xenial
+#      cache:
+#        apt: true
+#        directories:
+#          - "$CACHE"
+#      addons:
+#        apt:
+#          sources: &linuxsrcs
+#            - ubuntu-toolchain-r-test
+#            - sourceline: 'deb https://apt.kitware.com/ubuntu/ xenial main'
+#              key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
+#          packages: &gcc8pkgs
+#            - cmake
+#            - gcc
+#            - g++
+#            - gfortran
+#    -
+#      <<: *ubuntu
+#      env:
+#        - GCC=9
+#      addons:
+#        apt:
+#          sources: *linuxsrcs
+#          packages: &gcc7pkgs
+#            - cmake
+#            - gcc
+#            - g++
+#            - gfortran
+#    -
+#      <<: *ubuntu
+#      addons:
+#        apt:
+#          sources: *linuxsrcs
+#          packages:
+#            - gcc
+#            - g++
+#            - gfortran
+#    - &linux_allowed_failure
+#      <<: *ubuntu
+#      addons:
+#        apt:
+#          sources: *linuxsrcs
+#          packages: *gcc8pkgs
 
   allow_failures:
     - *linux_allowed_failure
@@ -100,7 +100,7 @@ install:
   - pwd
   #- sudo apt-get update
   #- sudo apt-get install gcc gfortran
-  - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none --yes-to-all 
+  - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none
 
 #after_success:
 #  - bash <(curl -s https://codecov.io/bash)
