### 2020-11-09 e7f7183
minor enhancements
