### 2020-11-09 9b8be58
minor enhancements
