### 2020-11-09 85e7991
minor enhancements
