### 2020-11-09 097d42d
minor enhancements
