+++ .travis.yml.success.yml
@@ -3,6 +3,19 @@ sudo: false
 
 matrix:
   include:
+    - os: linux
+      addons:
+        apt:
+          sources:
+            - ubuntu-toolchain-r-test
+          packages:
+            - cmake
+            - gfortran-8
+
+      script:
+        - cmake --version
+        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
+
     - os: osx
       addons:
         homebrew:
@@ -14,6 +27,7 @@ matrix:
       before_install:
         - brew link gcc
         - brew link openmpi
+
       script:
         - cmake --version
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
