### 2020-11-09 d233608
minor enhancements
