### 2020-11-09 efc8b1d
minor enhancements
