+++ .travis.yml
@@ -12,9 +12,11 @@ matrix:
             - cmake
             - openmpi
       before_install:
-        - brew install gcc
+        - brew install gcc || brew upgrade gcc
+        - brew install cmake || brew upgrade cmake
+        - brew install openmpi || brew upgrade openmpi
         - brew link gcc
-        - brew install openmpi
+        - brew link cmake
         - brew link openmpi
       script:
         - cmake --version
