### 2020-11-09 cab5795
minor enhancements
