### 2020-11-09 9c431f6
minor enhancements
