+++ .travis.yml
@@ -1,6 +1,17 @@
 language: c
 sudo: false
 
+cache:
+  directories:
+    - $HOME/Library/Caches/Homebrew
+    - /usr/local/Homebrew
+
+before_cache:
+  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
+# Credit https://discourse.brew.sh/t/best-practice-for-homebrew-on-travis-brew-update-is-5min-to-build-time/5215/9
+# Cache only .git files under "/usr/local/Homebrew" so "brew update" does not take 5min every build
+  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then find /usr/local/Homebrew \! -regex ".+\.git.+" -delete; fi
+
 matrix:
   include:
 
@@ -19,7 +30,7 @@ matrix:
     - os: osx
       before_install:
         - wget "https://github.com/cdslaborg/paramonte/releases/download/v1.4.0/libparamonte_c_darwin_x64_gnu_release_dynamic_heap.tar.gz"
-        - ls libparamonte*.tar.gz | xargs -i tar xvzf {}
+        - tar xvzf libparamonte_c_darwin_x64_gnu_release_dynamic_heap.tar.gz
         - cd libparamonte_c_darwin_x64_gnu_release_dynamic_heap
         - ./build.sh
       script:
