### 2020-11-09 bef391b
minor enhancements
