+++ .travis.yml
@@ -6,11 +6,11 @@ cache:
     - $HOME/Library/Caches/Homebrew
     - /usr/local/Homebrew
 
-before_cache:
-  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
-# Credit https://discourse.brew.sh/t/best-practice-for-homebrew-on-travis-brew-update-is-5min-to-build-time/5215/9
-# Cache only .git files under "/usr/local/Homebrew" so "brew update" does not take 5min every build
-  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then find /usr/local/Homebrew \! -regex ".+\.git.+" -delete; fi
+#before_cache:
+#  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
+## Credit https://discourse.brew.sh/t/best-practice-for-homebrew-on-travis-brew-update-is-5min-to-build-time/5215/9
+## Cache only .git files under "/usr/local/Homebrew" so "brew update" does not take 5min every build
+#  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then find /usr/local/Homebrew \! -regex ".+\.git.+" -delete; fi
 
 matrix:
   include:
@@ -27,10 +27,10 @@ matrix:
         - cmake --version
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
 
-    - os: osx
-      script:
-        - wget "https://github.com/cdslaborg/paramonte/releases/download/v1.4.0/libparamonte_c_darwin_x64_gnu_release_dynamic_heap.tar.gz"
-        - tar xvzf libparamonte_c_darwin_x64_gnu_release_dynamic_heap.tar.gz
-        - cd libparamonte_c_darwin_x64_gnu_release_dynamic_heap
-        - ./build.sh
-        - ./run.sh
+#    - os: osx
+#      script:
+#        - wget "https://github.com/cdslaborg/paramonte/releases/download/v1.4.0/libparamonte_c_darwin_x64_gnu_release_dynamic_heap.tar.gz"
+#        - tar xvzf libparamonte_c_darwin_x64_gnu_release_dynamic_heap.tar.gz
+#        - cd libparamonte_c_darwin_x64_gnu_release_dynamic_heap
+#        - ./build.sh
+#        - ./run.sh
