### 2020-11-09 d9694b4
minor enhancements
