### 2020-11-09 f800b2d
minor enhancements
