+++ .travis.yml
@@ -1,11 +1,11 @@
 language: c
 sudo: false
 
-#before_cache:
-#  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
-## Credit https://discourse.brew.sh/t/best-practice-for-homebrew-on-travis-brew-update-is-5min-to-build-time/5215/9
-## Cache only .git files under "/usr/local/Homebrew" so "brew update" does not take 5min every build
-#  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then find /usr/local/Homebrew \! -regex ".+\.git.+" -delete; fi
+before_cache:
+  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
+# Credit https://discourse.brew.sh/t/best-practice-for-homebrew-on-travis-brew-update-is-5min-to-build-time/5215/9
+# Cache only .git files under "/usr/local/Homebrew" so "brew update" does not take 5min every build
+  - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then find /usr/local/Homebrew \! -regex ".+\.git.+" -delete; fi
 
 cache:
   directories:
@@ -17,11 +17,11 @@ matrix:
 
     - os: osx
       before_install:
-        #- brew install gcc || brew upgrade gcc
-        #- brew install cmake || brew upgrade cmake
+        - brew install gcc || brew upgrade gcc
+        - brew install cmake || brew upgrade cmake
         - brew install openmpi || brew upgrade openmpi
-        #- brew link gcc
-        #- brew link cmake
+        - brew link gcc
+        - brew link cmake
         - brew link openmpi
       script:
         - cmake --version
