### 2020-11-09 8e3e717
minor enhancements
