### 2020-11-09 5b7f755
minor enhancements
