### 2020-11-09 9def2f4
minor enhancements
