+++ .travis.yml
@@ -31,13 +31,24 @@ matrix:
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
 
     - os: osx
+      addons:
+        homebrew:
+          #update: true
+          packages:
+            - gcc
+            - cmake
+            - openmpi
       before_install:
-        - brew install gcc || brew upgrade gcc
-        - brew install cmake || brew upgrade cmake
-        - brew install openmpi || brew upgrade openmpi
         - brew link gcc
         - brew link cmake
         - brew link openmpi
+      #before_install:
+      #  - brew install gcc || brew upgrade gcc
+      #  - brew install cmake || brew upgrade cmake
+      #  - brew install openmpi || brew upgrade openmpi
+      #  - brew link gcc
+      #  - brew link cmake
+      #  - brew link openmpi
       script:
         - cmake --version
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
