### 2020-11-09 4546c65
minor enhancements
