### 2020-11-09 18debc4
minor enhancements
