+++ .travis.yml
@@ -37,11 +37,11 @@ matrix:
           packages:
             - gcc
             - cmake
-            - openmpi
+            - mpich
       before_install:
         - brew link gcc
         - brew link cmake
-        - brew link openmpi
+        - brew link mpich
       #before_install:
       #  - brew install gcc || brew upgrade gcc
       #  - brew install cmake || brew upgrade cmake
