### 2020-11-09 bc83bb4
minor enhancements
