+++ .travis.yml
@@ -37,18 +37,18 @@ matrix:
           packages:
             - gcc
             - cmake
-            - mpich
+            - openmpi
       before_install:
         - brew link gcc
         - brew link cmake
-        - brew link mpich
-      #before_install:
+        - brew link openmpi
+      before_install:
+        - brew install openmpi || brew upgrade openmpi
+        - brew link openmpi
       #  - brew install gcc || brew upgrade gcc
       #  - brew install cmake || brew upgrade cmake
-      #  - brew install openmpi || brew upgrade openmpi
       #  - brew link gcc
       #  - brew link cmake
-      #  - brew link openmpi
       script:
         - cmake --version
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
