### 2020-11-09 40e7b8b
minor enhancements
