+++ .travis.yml
@@ -28,6 +28,10 @@ matrix:
         - alias gfortran=gfortran-8
         - export FC=gfortran-8
         - cmake --version
+        - gfortran --version
+        - gfortran-8 --version
+        - which gfortran
+        - which gfortran-8
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
 
     - os: osx
