### 2020-11-09 16f91e7
minor enhancements
