+++ .travis.yml
@@ -37,18 +37,17 @@ matrix:
     - os: osx
       addons:
         homebrew:
-          #update: true
           packages:
-            - gcc
-            - cmake
-            - openmpi
+          - gcc
+          - cmake
+          - openmpi
       before_install:
         - brew link gcc
         - brew link cmake
         - brew link openmpi
-      before_install:
-        - brew install openmpi || brew upgrade openmpi
-        - brew link openmpi
+      #before_install:
+      #  - brew install openmpi || brew upgrade openmpi
+      #  - brew link openmpi
       #  - brew install gcc || brew upgrade gcc
       #  - brew install cmake || brew upgrade cmake
       #  - brew link gcc
