### 2020-11-09 0dff66b
minor enhancements
