### 2020-11-09 c65dac2
minor enhancements
