### 2020-11-09 142d79c
minor enhancements
