### 2020-11-09 5b5919b
minor enhancements
