### 2020-11-09 d991cfe
minor enhancements
