### 2020-11-09 234f98d
minor enhancements
