### 2020-11-09 ad0c096
minor enhancements
