### 2020-11-09 590ea9e
minor enhancements
