+++ .travis.yml
@@ -46,9 +46,8 @@ matrix:
         #- brew install openmpi || brew upgrade openmpi
         - brew link gcc
         - brew link cmake
-        #- brew link openmpi
+        - brew link openmpi
       script:
-        - which mpiexec
         - cmake --version
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none -s gnu
 
