### 2020-11-09 bc0e38b
minor enhancements
