+++ .travis.yml
@@ -1,5 +1,5 @@
 language: c
-sudo: false
+sudo: true
 
 before_cache:
   - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
