### 2020-11-09 1ad9937
minor enhancements
