+++ .travis.yml
@@ -40,14 +40,14 @@ matrix:
             - gcc
             - cmake
             - openmpi
-      before_install:
+      #before_install:
         #- brew install gcc || brew upgrade gcc
         #- brew install cmake || brew upgrade cmake
         #- brew install openmpi || brew upgrade openmpi
-        - brew link gcc
-        - brew link cmake
         #- brew link openmpi
       script:
+        - brew link gcc
+        - brew link cmake
         - cmake --version
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none mpi" -s gnu
 
