### 2020-11-09 ebc402f
minor enhancements
