+++ .travis.yml
@@ -14,32 +14,49 @@ cache:
 
 matrix:
   include:
-
-    - os: linux
-      addons:
-        apt:
-          sources:
-            - ubuntu-toolchain-r-test
-          packages:
-            - mpich
-            - gfortran-8
-      script:
-        - chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
-        - export FC=gfortran-8
-        - cmake --version
-        - gfortran-8 --version
-        - which mpiexec
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none --fortran $(which gfortran-8) --mpiexec $(which mpiexec) --yes-to-all
-
-    - name: "darwin"
+    - &macOS
       os: osx
-      osx_image: xcode10
+      osx_image: xcode10.2
+      env:
+        - BUILD_TYPE="InstallScript"
+          GCC=9
+          CC=gcc-9
+          CXX=g++-9
+          FC=gfortran-9
       addons:
         homebrew:
           packages:
             - gcc
-            - cmake
-            - openmpi
+            - mpich
+            - shellcheck
+          update: true
+
+#    - os: linux
+#      addons:
+#        apt:
+#          sources:
+#            - ubuntu-toolchain-r-test
+#          packages:
+#            - mpich
+#            - gfortran-8
+#      script:
+#        - chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
+#        - export FC=gfortran-8
+#        - cmake --version
+#        - gfortran-8 --version
+#        - which mpiexec
+#        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none --fortran $(which gfortran-8) --mpiexec $(which mpiexec) --yes-to-all
+#
+#    - name: "darwin"
+#      os: osx
+#      osx_image: xcode10
+#      addons:
+#        homebrew:
+#          packages:
+#            - gcc
+#            - cmake
+#            - openmpi
+#            - update: true
       #before_install:
         #- brew install gcc || brew upgrade gcc
         #- brew install cmake || brew upgrade cmake
