### 2020-11-09 b556e29
minor enhancements
