### 2020-11-10 a5c5e3b
GNU GCC 10 MPICH compatibility. The library be now compiled with GNU Gfortran 10 and MPICH library. The mpi_bcast of MPICH breaks the code because of argument-mismatch. For now, "-std=legacy" has been added to gfortran flags and "-pedantic" has been removed to bypass the MPICH outdated interface.
