### 2020-11-10 19e3f00
minor enhancements
