### 2020-11-10 254f2f9
minor enhancements
