### 2020-11-10 25dbc3c
minor build enhancements
