### 2020-11-10 59f2d0e
minor build enhancements
