### 2020-11-10 6c135da
minor build enhancements
