### 2020-11-10 5a32eda
minor build enhancements
