### 2020-11-11 a6da0d8
minor build enhancements
