### 2020-11-11 f7262a6
minor enhancements
