### 2020-11-11 ef880b9
minor enhancements
