### 2020-11-11 1de471d
minor enhancements
