### 2020-11-11 c26a21c
minor enhancements
