### 2020-11-11 cdslaborg/paramonte#3 @milancurcic
Building from source fails on x86-64 Ubuntu 18.04
This is paramonte checked out from master at commit ebc402f4a0dfb1df4327e6bf92f04d870dffbdb1.

I attach the install output: [out.txt](https://github.com/cdslaborg/paramonte/files/5525028/out.txt)
