### 2020-11-11 cdslaborg/paramonte#4 @milancurcic
License agreement terms
Paramonte is MIT licensed.

However, ACKNOWLEDGMENT.md states:

> As per the ParaMonte library license agreement terms, if you use any parts of
this library for any purposes, kindly acknowledge the use of ParaMonte in your
work (education/research/industry/development/...) by citing the ParaMonte
library's main publications as listed here:

The MIT license doesn't require acknowledgment, only that the copyright statement and the license body is included in the copy.

So I think the first paragraph in the ACKNOWLEDGMENT.md should be reworded to something like:

> If you use ParaMonte, please acknowledge it by citing the ParaMonte library's main publications

to ask for acknowledgment but not make it sound like it's a legal requirement.
