### 2020-11-11 cdslaborg/paramonte#5 @milancurcic
Provide binary releases with OpenMPI
OpenMPI is a popular implementation beside MPICH. It would be nice if there were parallel builds of Paramonte with OpenMPI.
