### 2020-11-11 #2741 @milancurcic
The items I checked off are okay. Few remain unchecked because of issues with installation.

I tested serial binary releases for Fortran with both GNU and Intel. Both worked as expected and I was able to build and run the example program.

I wasn't able to use the parallel releases because they require MPICH. I tend to work with OpenMPI so it'd be nice if binary releases with OpenMPI were available. I opened an [issue](https://github.com/cdslaborg/paramonte/issues/5) for this.

I tried building from source which failed. I opened an [issue](https://github.com/cdslaborg/paramonte/issues/3) for this.

There was also a slight issue with wording in [ACKNOWLEDGMENT.md](https://github.com/cdslaborg/paramonte/blob/master/ACKNOWLEDGMENT.md) which make it sound like the license requires citing the software. I opened an [issue](https://github.com/cdslaborg/paramonte/issues/4) for this.

I confirm and re-iterate points 1 and 3 by @williamfgc as important. Points 2, 5, and 8 I consider nice to have but not essential. Point 6 I couldn't confirm--I had no problem finding the API docs and found them thorough.

I suggest tackling these before proceeding with the review.
