+++ .travis.yml
@@ -15,66 +15,66 @@ cache:
 matrix:
   include:
 
-    - &macOS
-      os: osx
-      osx_image: xcode11
-      env:
-          GCC=9
-          CC=gcc-9
-          CXX=g++-9
-          FC=gfortran-9
-      addons:
-        homebrew:
-          packages:
-            - gcc@9
-            - cmake
-            - openmpi
-            - shellcheck
-          update: true
-      before_install:
-        - brew install gcc@9 || brew upgrade gcc
-        - brew install cmake || brew upgrade cmake
-        - brew install openmpi || brew upgrade openmpi
-        - brew link gcc
-        - brew link cmake
-        - brew link openmpi
-      script:
-        - cmake --version
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none mpi" #-s gnu
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-
-    - &macOS
-      os: osx
-      osx_image: xcode12
-      addons:
-        homebrew:
-          packages:
-            - cmake
-            - mpich
-            - gcc@10
-            - shellcheck
-          update: true
-      before_install:
-        - brew install gcc@10 || brew upgrade gcc
-        - brew install cmake || brew upgrade cmake
-        - brew install mpich || brew upgrade mpich
-        - brew link gcc
-        - brew link cmake
-        - brew link mpich
-      script:
-        - cmake --version
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none mpi" #-s gnu
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
[…]
