### 2020-11-11 030fa95
minor enhancements
