### 2020-11-11 e61ec90
MPI test script added
