### 2020-11-12 8fb28a0
minor enhancements
