### 2020-11-12 5b9de1e
minor enhancements
