### 2020-11-12 ac8ce02
minor enhancements
