### 2020-11-12 089fb03
minor enhancements
