### 2020-11-12 8660600
minor enhancements
