### 2020-11-12 b1b81a2
minor enhancements
