+++ example/buildParaMonteExample.sh
@@ -237,6 +237,8 @@ do
             do
                 if [ -d "${Fortran_COMPILER_LIB_DIR}" ]; then
                     libgfortranPathList=$(find "${Fortran_COMPILER_LIB_DIR}" -name "libgfortran.so*")
+                    echo >&2 "-- ParaMonteExample${LANG_NAME} - searching for the ParaMonte library dll dependency file at: ${Fortran_COMPILER_LIB_DIR}"
+                    echo >&2 "-- ParaMonteExample${LANG_NAME} - detected shared files: ${libgfortranPathList}"
                     for libgfortranPath in ${libgfortranPathList}
                     do
                         echo >&2
@@ -246,6 +248,8 @@ do
                         yes | \cp -rf "${libgfortranPath}/"libgfortran.so.* "${ParaMonteExample_LIB_DIR_CURRENT}/" && {
                             copySucceeded=true
                             break
+                        } || {
+                            echo >&2 "-- ParaMonteExample${LANG_NAME} - the ParaMonte library dll dependency file copy failed at: ${libgfortranPath}"
                         }
                     done
                 fi
