### 2020-11-12 39fb29a
minor enhancements
