### 2020-11-12 1353bf0
minor enhancements
