### 2020-11-12 909ebed
minor enhancements
