### 2020-11-12 e1c6f01
minor enhancements
