### 2020-11-12 f4f14c5
minor enhancements
