### 2020-11-12 be481cd
minor enhancements
