+++ example/buildParaMonteExample.sh
@@ -237,6 +237,7 @@ do
             for Fortran_COMPILER_LIB_DIR in ${Fortran_COMPILER_LIB_DIR_LIST//:/ }
             do
                 if [ -d "${Fortran_COMPILER_LIB_DIR}" ]; then
+
                     libgfortranPathList=$(find "${Fortran_COMPILER_LIB_DIR}" -name "libgfortran.so*")
                     #echo >&2 "-- ParaMonteExample${LANG_NAME} - searching for the ParaMonte library dll dependency file at: ${Fortran_COMPILER_LIB_DIR}"
                     #echo >&2 "-- ParaMonteExample${LANG_NAME} - detected shared files: ${libgfortranPathList}"
@@ -253,6 +254,7 @@ do
                             echo >&2 "-- ParaMonteExample${LANG_NAME} - the ParaMonte library dll dependency file copy failed at: ${libgfortranPath}"
                         }
                     done
+
                 fi
                 if [ "${copySucceeded}" = "true" ]; then break; fi
             done
