### 2020-11-12 5bdb3d0
minor enhancements
