### 2020-11-12 c84628b
minor enhancements
