### 2020-11-12 efea3cf
minor enhancements
