### 2020-11-12 0be779f
minor enhancements
