+++ .travis.yml
@@ -95,18 +95,19 @@ matrix:
             - g++
             - cmake
             - mpich
-            - gfortran
+            - gfortran-8
       script:
         - sudo apt update
         - sudo apt install gcc
         - sudo apt install g++
         - sudo apt install cmake
         - sudo apt install mpich
-        - sudo apt install gfortran
+        - sudo apt install gfortran-10
         #- chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
         - cmake --version
         - gfortran --version
-        - echo >&2 "The MPIEXEC binday path: $(which mpiexec) || echo >&2 "MPI library not found...skipping"
+        - gfortran-10 --version
+        #- echo >&2 "The MPIEXEC binday path: $(which mpiexec) || echo >&2 "MPI library not found...skipping"
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none" --yes-to-all
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
