### 2020-11-12 2fc63a9
minor enhancements
