### 2020-11-12 742885d
minor enhancements
