+++ .travis.yml
@@ -90,7 +90,7 @@ matrix:
             - ubuntu-toolchain-r-test
             - sourceline: 'deb https://apt.kitware.com/ubuntu/ xenial main'
               key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
-          packages: &gcc8pkgs
+          packages: &gcc10pkgs
             - gcc
             - g++
             - cmake
