### 2020-11-13 b7a977a
minor enhancements
