+++ .travis.yml
@@ -96,39 +96,39 @@ matrix:
 #        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
 #        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
 
-    - &ubuntu
-      os: linux
-      sudo: true
-      dist: xenial
-      cache:
-        apt: true
-        directories:
-          - "$CACHE"
-      addons:
-        apt:
-          sources: &linuxsrcs
-            - ubuntu-toolchain-r-test
-            - sourceline: 'deb https://apt.kitware.com/ubuntu/ xenial main'
-              key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
-          packages: &gcc10pkgs
-            - gcc
-            - g++
-            - cmake
-            - mpich
-            - gfortran-8
-      script:
-        #- chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
-        - cmake --version
-        - gfortran --version
-        - gfortran-8 --version
-        #- echo >&2 "The MPIEXEC binday path: $(which mpiexec) || echo >&2 "MPI library not found...skipping"
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none" --yes-to-all
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        #- cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        #- cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        #- cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
+#    - &ubuntu
+#      os: linux
+#      sudo: true
+#      dist: xenial
+#      cache:
+#        apt: true
+#        directories:
+#          - "$CACHE"
+#      addons:
+#        apt:
+#          sources: &linuxsrcs
+#            - ubuntu-toolchain-r-test
+#            - sourceline: 'deb https://apt.kitware.com/ubuntu/ xenial main'
+#              key_url: 'https://apt.kitware.com/keys/kitware-archive-latest.asc'
+#          packages: &gcc10pkgs
+#            - gcc
+#            - g++
+#            - cmake
+#            - mpich
+#            - gfortran-8
+#      script:
+#        #- chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
+#        - cmake --version
+#        - gfortran --version
+#        - gfortran-8 --version
[…]
