### 2020-11-13 523a7e1
minor enhancements
