+++ .travis.yml
@@ -26,11 +26,12 @@ matrix:
             - gfortran-9
       script:
         - chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
-        - cmake --version
-        - gfortran --version
-        - gfortran-9 --version
+        - export PATH="/usr/local/cmake-3.16.3/bin:${PATH}"
+        - cmake --version || echo "cmake not found..."
+        - gfortran --version && FCFLAG="-f gfortran" || echo "gfortran not found..."
+        - gfortran-9 --version && FCFLAG="-f gfortran-9" || echo "gfortran-9 not found..."
         #- echo >&2 "The MPIEXEC binday path: $(which mpiexec) || echo >&2 "MPI library not found...skipping"
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none" --yes-to-all
+        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none" -s gnu --yes-to-all ${FCFLAG}
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
