### 2020-11-13 c2c545b
minor enhancements
