### 2020-11-13 e9e48a8
minor enhancements
