### 2020-11-14 f95307c
minor enhancements
