+++ example/build.sh
@@ -169,6 +169,22 @@ PMLIB_FULL_NAME=${PMLIB_FULL_PATH##*/}
 PMLIB_BASE_NAME=${PMLIB_FULL_NAME%.*}
 
 ####################################################################################################################################
+# determine the GNU compiler list
+####################################################################################################################################
+
+if [[ "$PMLIB_FULL_NAME" =~ .*"_fortran_".* ]]; then
+    cpattern="gfortran"
+elif [[ "$PMLIB_FULL_NAME" =~ .*"_cpp_".* ]]; then
+    cpattern="g++"
+elif [[ "$PMLIB_FULL_NAME" =~ .*"_c_".* ]]; then
+    cpattern="gcc"
+    clist="gcc"
+fi
+if ! [[ "$PMLIB_FULL_NAME" =~ .*"_c_".* ]]; then
+    clist=$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/${cpattern}*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)
+fi
+
+####################################################################################################################################
 # get ParaMonte's compiler suite
 ####################################################################################################################################
 
@@ -185,14 +201,14 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"_fortran_".* ]]; then
     if [[ "$PMLIB_FULL_NAME" =~ .*"_intel_".* ]]; then
         PM_COMPILER_SUITE=intel
         #COMPILER_LIST="ifort gfortran"
-        declare -a COMPILER_LIST=("ifort" "gfortran")
+        declare -a COMPILER_LIST=("ifort" "${clist}")
     fi
     if [[ "$PMLIB_FULL_NAME" =~ .*"_gnu_".* ]]; then
         PM_COMPILER_SUITE=gnu
         #COMPILER_LIST="gfortran ifort"
-        declare -a COMPILER_LIST=("gfortran" "ifort")
+        declare -a COMPILER_LIST=("${clist}" "ifort")
     fi
-    if [ "$PLATFORM" = "mingw" ] || [ "$PLATFORM" = "cygwin" ]; then declare -a COMPILER_LIST=("gfortran"); fi
+    if [ "$PLATFORM" = "mingw" ] || [ "$PLATFORM" = "cygwin" ]; then declare -a COMPILER_LIST=("${clist}"); fi
 fi
 
 if [[ "$PMLIB_FULL_NAME" =~ .*"_c_".* ]]; then
@@ -201,14 +217,14 @@ if [[ "$PMLIB_FULL_NAME" =~ .*"_c_".* ]]; then
     if [[ "$PMLIB_FULL_NAME" =~ .*"_intel_".* ]]; then
         PM_COMPILER_SUITE=intel
         #COMPILER_LIST="icc gcc"
-        declare -a COMPILER_LIST=("icc" "gcc")
+        declare -a COMPILER_LIST=("icc" "${clist}")
     fi
     if [[ "$PMLIB_FULL_NAME" =~ .*"_gnu_".* ]]; then
         PM_COMPILER_SUITE=gnu
         #COMPILER_LIST="gcc icc"
-        declare -a COMPILER_LIST=("gcc" "icc")
+        declare -a COMPILER_LIST=("${clist}" "icc")
     fi
-    if [ "$PLATFORM" = "mingw" ] || [ "$PLATFORM" = "cygwin" ]; then declare -a COMPILER_LIST=("gcc"); fi
+    if [ "$PLATFORM" = "mingw" ] || [ "$PLATFORM" = "cygwin" ]; then declare -a COMPILER_LIST=("${clist}"); fi
 fi
 
 if [[ "$PMLIB_FULL_NAME" =~ .*"_cpp_".* ]]; then
[…]
