### 2020-11-15 5ade9d4
minor enhancements
