+++ example/build.sh
@@ -455,7 +455,7 @@ do
         echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - compiling ParaMonte example with ${COMPILER}"
         echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - ${COMPILER} ${COMPILER_FLAGS} ${SRC_FILES} -c"
 
-        ${COMPILER} ${COMPILER_FLAGS} ${SRC_FILES} -c
+        ${COMPILER} -dumpversion >/dev/null 2>&1 && ${COMPILER} ${COMPILER_FLAGS} ${SRC_FILES} -c
 
         csvLinkerList=${COMPILER}
         LINKER_FLAGS=
@@ -482,7 +482,7 @@ do
             echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - linking ParaMonte example with ${LINKER}"
             echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} ${SRC_FILES//.$SRC_EXT/.o} ${PMLIB_FULL_NAME} -o ${PM_EXAM_EXE_NAME}"
 
-            ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} ${SRC_FILES//.$SRC_EXT/.o} "${PMLIB_FULL_NAME}" -o ${PM_EXAM_EXE_NAME}
+            ${LINKER} -dumpversion >/dev/null 2>&1 && ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} ${SRC_FILES//.$SRC_EXT/.o} "${PMLIB_FULL_NAME}" -o ${PM_EXAM_EXE_NAME}
 
             if [ $? -eq 0 ]; then
 
