### 2020-11-15 1ff39b6
minor enhancements
