### 2020-11-15 d510126
minor enhancements
