### 2020-11-15 e5020e6
minor enhancements
