### 2020-11-15 41dbc0b
minor enhancements
