+++ .travis.yml
@@ -30,7 +30,7 @@ matrix:
         - cmake --version || echo "cmake not found..."
         - gfortran --version && FCFLAG="-f $(which gfortran)" || echo "gfortran not found..."
         - gfortran-9 --version && FCFLAG="-f $(which gfortran-9)" || echo "gfortran-9 not found..."
-        - echo >&2 "The MPIEXEC binary path: $(which mpiexec) || echo >&2 "MPI library not found...skipping"
+        - echo >&2 "The MPIEXEC binary path: $(which mpiexec)" || echo >&2 "MPI library not found...skipping"
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none mpi" -s gnu --yes-to-all ${FCFLAG}
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
