### 2020-11-15 af98c28
minor enhancements
