### 2020-11-15 fcf4ea4
minor enhancements
