### 2020-11-15 0e696f9
minor enhancements
