### 2020-11-15 b084300
minor enhancements
