### 2020-11-15 b8b4da3
minor enhancements
