### 2020-11-15 f72abd7
minor enhancements
