### 2020-11-15 ae69786
minor enhancements
