### 2020-11-15 92c630c
minor enhancements
