### 2020-11-15 8dc0691
minor enhancements
