+++ .travis.yml
@@ -23,11 +23,12 @@ matrix:
           packages:
             - cmake
             - gfortran-9
+            - mpich
+            - libmpich-dev
             #- libblas-dev
             #- liblapack-dev
             #- libopenmpi-dev
             #- openmpi-bin
-            - mpich
       script:
         #- chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
         #- export PATH="/usr/local/cmake-3.16.3/bin:${PATH}"
