+++ .travis.yml
@@ -24,7 +24,7 @@ matrix:
             - cmake
             - gfortran-9
             - mpich
-            - libmpich-dev
+            #- libmpich-dev
             #- libblas-dev
             #- liblapack-dev
             #- libopenmpi-dev
@@ -32,10 +32,11 @@ matrix:
       script:
         #- chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
         #- export PATH="/usr/local/cmake-3.16.3/bin:${PATH}"
-        #- echo >&2 "The MPIEXEC binary path: $(which mpiexec)" || echo >&2 "MPI library not found...skipping"
         - cmake --version || echo "cmake not found..."
         - gfortran --version && FCFLAG="-f $(which gfortran)" || echo "gfortran not found..."
         - gfortran-9 --version && FCFLAG="-f $(which gfortran-9)" || echo "gfortran-9 not found..."
+        - sudo apy install mpich
+        - echo >&2 "The MPIEXEC binary path: $(which mpiexec)" || echo >&2 "MPI library not found...skipping"
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "mpi" --yes-to-all --silent ${FCFLAG}
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
