### 2020-11-15 f8b7cdb
minor enhancements
