### 2020-11-15 0613145
minor enhancements
