### 2020-11-15 1b7b7fd
minor enhancements
