### 2020-11-15 474fdb9
minor enhancements
