### 2020-11-15 1769611
minor enhancements
