### 2020-11-15 8abe7cb
minor enhancements
