### 2020-11-15 65ca305
minor enhancements
