+++ .travis.yml
@@ -42,6 +42,9 @@ matrix:
         - echo $(which mpif90)
         - echo $(which mpifort)
         - echo $(which mpiexec)
+        - mpiexecFlag="$(command -v mpiexec)"
+        - if ! [ -z "${mpiexecFlag}" ]; then mpiexecFlag="--mpiexec ${mpiexecFlag}"; else echo "mpiexec not found"; fi
+
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "mpi" --yes-to-all ${FCFLAG}
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
