### 2020-11-15 aaee2dc
minor enhancements
