### 2020-11-15 7200eee
minor enhancements
