### 2020-11-15 65f61da
minor enhancements
