### 2020-11-15 b91458a
minor enhancements
