+++ .travis.yml
@@ -47,6 +47,9 @@ matrix:
         - if [ -f "${mpiexecFlag}" ]; then mpiexecFlag="--mpiexec ${mpiexecFlag}"; else unset mpiexecFlag; echo "mpiexec not found..."; fi
         - echo "FCFLAG = ${FCFLAG}"
         - echo "mpiexecFlag = ${mpiexecFlag}"
+        - ls /usr/bin/
+        - echo "$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/*mpifort*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)"
+        - echo "$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/*mpicc*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)"
         - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "mpi" --yes-to-all ${FCFLAG} ${mpiexecFlag}
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_c_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
         - cd /Users/travis/build/cdslaborg/paramonte/bin/libparamonte_cpp_linux_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
