### 2020-11-15 3c25934
minor enhancements
