+++ .travis.yml
@@ -23,7 +23,7 @@ matrix:
           packages:
             - cmake
             - gfortran-9
-            - openmpi-bin
+            #- openmpi-bin
             #- mpich
             #- libmpich-dev
             #- libblas-dev
@@ -34,23 +34,23 @@ matrix:
         #- chmod +x ./.travis/install_cmake.sh && ./.travis/install_cmake.sh
         #- export PATH="/usr/local/cmake-3.16.3/bin:${PATH}"
         #- echo "The MPIEXEC binary path: $(which mpiexec)" || echo "MPI library not found...skipping"
+        #- echo $(which mpicc)
+        #- echo $(which mpif90)
+        #- echo $(which mpifort)
+        #- echo $(which mpiexec)
+        #- echo $(which mpirun)
+        #- mpiexecFlag="$(command -v mpirun)"
+        #- if [ -f "${mpiexecFlag}" ]; then mpiexecFlag="--mpiexec ${mpiexecFlag}"; else unset mpiexecFlag; echo "mpiexec not found..."; fi
+        #- echo "FCFLAG = ${FCFLAG}"
+        #- echo "mpiexecFlag = ${mpiexecFlag}"
+        #- ls /usr/bin/
+        #- echo "$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/*mpifort*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)"
+        #- echo "$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/*mpicc*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)"
+        #- sudo apt install openmpi-bin
         - cmake --version || echo "cmake not found..."
         - gfortran --version && FCFLAG="-f $(which gfortran)" || echo "gfortran not found..."
         - gfortran-9 --version && FCFLAG="-f $(which gfortran-9)" || echo "gfortran-9 not found..."
-        - sudo apt install openmpi-bin
-        - echo $(which mpicc)
-        - echo $(which mpif90)
-        - echo $(which mpifort)
-        - echo $(which mpiexec)
-        - echo $(which mpirun)
-        - mpiexecFlag="$(command -v mpirun)"
-        - if [ -f "${mpiexecFlag}" ]; then mpiexecFlag="--mpiexec ${mpiexecFlag}"; else unset mpiexecFlag; echo "mpiexec not found..."; fi
-        - echo "FCFLAG = ${FCFLAG}"
-        - echo "mpiexecFlag = ${mpiexecFlag}"
-        - ls /usr/bin/
-        - echo "$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/*mpifort*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)"
-        - echo "$(( IFS=:; for p in $PATH; do unset lsout; lsout=$(ls -dm "$p"/*mpicc*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi; done ) 2>/dev/null)"
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "mpi" --yes-to-all ${FCFLAG} ${mpiexecFlag}
+        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par none" --yes-to-all ${FCFLAG} ${mpiexecFlag}
[…]
