### 2020-11-15 538eabe
minor enhancements
