### 2020-11-15 c44f3e1
minor enhancements
