+++ .travis.yml
@@ -65,40 +65,6 @@ matrix:
 
     - &macOS
       os: osx
-      osx_image: xcode11
-      env:
-          GCC=9
-          CC=gcc-9
-          CXX=g++-9
-          FC=gfortran-9
-      addons:
-        homebrew:
-          packages:
-            - gcc@9
-            - cmake
-            - openmpi
-            - shellcheck
-          update: true
-      before_install:
-        - brew install gcc@9 || brew upgrade gcc
-        - brew install cmake || brew upgrade cmake
-        - brew install openmpi || brew upgrade openmpi
-        - brew link gcc
-        - brew link cmake
-        - brew link openmpi
-      script:
-        - cmake --version
-        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none mpi" #-s gnu
-        - echo $(pwd)
-        - cd $(pwd)/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd $(pwd)/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd $(pwd)/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
-        - cd $(pwd)/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        - cd $(pwd)/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-        - cd $(pwd)/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
-
-    - &macOS
-      os: osx
       osx_image: xcode12
       addons:
         homebrew:
@@ -125,6 +91,40 @@ matrix:
         - cd $(pwd)/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
         - cd $(pwd)/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
         - cd $(pwd)/bin/libparamonte_fortran_darwin_x64_gnu_release_dynamic_heap_mpi && ./build.sh && ./run.sh
+
+#    - &macOS
+#      os: osx
+#      osx_image: xcode11
+#      env:
+#          GCC=9
+#          CC=gcc-9
+#          CXX=g++-9
+#          FC=gfortran-9
+#      addons:
+#        homebrew:
+#          packages:
+#            - gcc@9
+#            - cmake
+#            - openmpi
+#            - shellcheck
+#          update: true
+#      before_install:
+#        - brew install gcc@9 || brew upgrade gcc
+#        - brew install cmake || brew upgrade cmake
+#        - brew install openmpi || brew upgrade openmpi
+#        - brew link gcc
+#        - brew link cmake
+#        - brew link openmpi
+#      script:
+#        - cmake --version
+#        - ./install.sh --lang "C C++ Fortran" --build release --lib dynamic --mem heap --par "none mpi" #-s gnu
+#        - echo $(pwd)
+#        - cd $(pwd)/bin/libparamonte_c_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
+#        - cd $(pwd)/bin/libparamonte_cpp_darwin_x64_gnu_release_dynamic_heap && ./build.sh && ./run.sh
[…]
