### 2020-11-15 c605269
minor enhancements
