### 2020-11-15 48041fb
minor enhancements
