### 2020-11-16 3d5b78d
minor enhancements
