+++ auxil/getCompilerVersion.f90
@@ -43,7 +43,7 @@ program getCompilerVersion
     use iso_fortran_env, only: compiler_version, IK => int32
     implicit none
     integer(IK) , parameter     :: INTEL_VERSION(3) = [18_IK,0_IK,0_IK]
-    integer(IK) , parameter     :: GNU_VERSION(3) = [7_IK,3_IK,0_IK]
+    integer(IK) , parameter     :: GNU_VERSION(3) = [7_IK,1_IK,0_IK]
     integer(IK) , allocatable   :: MinVersion(:)
     integer(IK)                 :: fileunit, startindex, endindex
     logical                     :: isIntel = .false.
