+++ buildParaMonte.sh
@@ -314,10 +314,10 @@ cat << EndOfMessage
         -e | --heap_enabled     : the ParaMonte library heap array allocation enabled?: true, false
         -t | --test_enabled     : the ParaMonte library test run enabled?: true, false
         -x | --exam_enabled     : the ParaMonte library examples run enabled?: true, false
+        -S | --shared_enabled   : release the shared library dependencies in the binary release of the ParaMonte library: true, false
         -f | --fortran          : path to Fortran compiler. If provided, the ParaMonte library will be built via the specified compiler.
         -M | --mpiexec          : path to mpiexec routine. If provided, it will be used to find the MPI library.
         -F | --fresh            : enables a fresh installation of all of the prerequisites of ParaMonte library. Applicable only to GNU compiler suite.
-        -S | --silent           : silences all messages from the installation of all of the prerequisites of ParaMonte library.
         -y | --yes-to-all       : if a fresh installation of all of the prerequisites is needed, automatically answer yes to all permission requests.
         -B | --bootstrap        : enables robust bootstrap build when building the required GCC version with an old GCC version. Applicable only to GNU compiler suite.
         -n | --nproc            : the default number of processes (coarray images) on which the ParaMonte examples/tests (if any) will be run: positive integer
@@ -350,12 +350,12 @@ unset INTERFACE_LANGUAGE
 unset FOR_COARRAY_NUM_IMAGES
 unset Fortran_COMPILER_PATH_USER
 ParaMonteExample_RUN_ENABLED=true
+localInstallationEnabled=false
 ParaMonteTest_RUN_ENABLED=true
 freshInstallEnabled=false
 YES_TO_ALL_DISABLED=true
-RELEASE_ENABLED=false
 DRYRUN_ENABLED=false
-silent_flag=""
+shared_enabled=true
 CLEAN=false
 
 chmod a+x ./configParaMonte.sh
@@ -372,7 +372,7 @@ while [ "$1" != "" ]; do
                                 INTERFACE_LANGUAGE=$1
                                 ;;
         -s | --compiler_suite ) shift
-                                PMCS=$1; export PMCS
+                                PMCS=$1; export PMCS;
                                 ;;
         -b | --build )          shift
                                 BTYPE=$1; export BTYPE
@@ -398,6 +398,9 @@ while [ "$1" != "" ]; do
         -x | --exam_enabled )   shift
                                 ParaMonteExample_RUN_ENABLED=$1; export ParaMonteExample_RUN_ENABLED
                                 ;;
+        -S | --shared_enabled ) shift
+                                shared_enabled=$1; export shared_enabled
+                                ;;
         -f | --fortran )        shift
                                 Fortran_COMPILER_PATH_USER=$1; export Fortran_COMPILER_PATH_USER
                                 ;;
@@ -409,11 +412,9 @@ while [ "$1" != "" ]; do
                                 ;;
         -F | --fresh )          freshInstallEnabled=true; export freshInstallEnabled
                                 ;;
-        -S | --silent )         silent_flag=">/dev/null"
[…]
