### 2020-11-16 7f2c2cd
minor enhancements
