+++ src/interface/MATLAB/CHANGES.md
@@ -5,6 +5,13 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 ## **Version 2.x.x**  
 
+### Version  2.2.1 -- November 15, 2020  
+
+**Minor enhancements**  
+
++   Minor enhancements to the Kernel library 
+    build scripts and dependencies management.
+
 ### Version  2.2.0 -- October 29, 2020  
 
 **Enhancements**  
