### 2020-11-16 8ac9b5b
minor enhancements
