### 2020-11-16 e367e4c
minor enhancements
