### 2020-11-16 3baacd0
minor enhancements
