+++ .gitattributes.txt
@@ -1,12 +1,10 @@
-# Prevent dev-ops files from making it into the release archives
-.travis.yml export-ignore
-.pullapprove.yml export-ignore
-.gitattributes export-ignore
-.gitignore export-ignore
-codecov.yml export-ignore
+# Prevent development files from making it into the release archives
 *.enc export-ignore
 .github export-ignore
-makefile export-ignore
-Makefile export-ignore
-make.inc* export-ignore
+.gitignore export-ignore
+.travis.yml export-ignore
 .Dockerfiles export-ignore
+.gitattributes export-ignore
+.pullapprove.yml export-ignore
+**/.travis/* export-ignore
+codecov.yml export-ignore
