### 2020-11-16 f234afb
minor enhancements
