### 2020-11-16 ffce4a5
minor enhancements
