### 2020-11-16 806cbcd
minor enhancements
