### 2020-11-16 e2782a6
minor enhancements
