+++ example/build.sh
@@ -458,161 +458,174 @@ do
         echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - compiling ParaMonte example with ${COMPILER}"
         echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - ${COMPILER} ${COMPILER_FLAGS} ${SRC_FILES} -c"
 
-        ${COMPILER} -dumpversion >/dev/null 2>&1 && ${COMPILER} ${COMPILER_FLAGS} ${SRC_FILES} -c
+        ${COMPILER} -dumpversion >/dev/null 2>&1 && ${COMPILER} ${COMPILER_FLAGS} ${SRC_FILES} -c && \
+        {
 
-        csvLinkerList=${COMPILER}
-        LINKER_FLAGS=
-        if ([ "${EXAMPLE_LANGUAGE}" = "C" ] || [ "${EXAMPLE_LANGUAGE}" = "C++" ]) && [ "${PM_LIB_TYPE}" = "static" ]; then
-            if [ "${PM_COMPILER_SUITE}" = "intel" ]; then
-                csvLinkerList="ifort"
-                LINKER_FLAGS="-nofor_main"
-                #if [ "${MPI_ENABLED}" = "true" ]; then
-                #    LINKER="mpiifort" # xxx point of weakness: assumes intel mpi to have been installed
-                #fi
+            csvLinkerList=${COMPILER}
+            LINKER_FLAGS=
+            if ([ "${EXAMPLE_LANGUAGE}" = "C" ] || [ "${EXAMPLE_LANGUAGE}" = "C++" ]) && [ "${PM_LIB_TYPE}" = "static" ]; then
+                if [ "${PM_COMPILER_SUITE}" = "intel" ]; then
+                    csvLinkerList="ifort"
+                    LINKER_FLAGS="-nofor_main"
+                    #if [ "${MPI_ENABLED}" = "true" ]; then
+                    #    LINKER="mpiifort" # xxx point of weakness: assumes intel mpi to have been installed
+                    #fi
+                fi
+                if [ "${PM_COMPILER_SUITE}" = "gnu" ]; then
+                    csvLinkerList="gfortran"
+                    #csvLinkerList="${clist}"
+                    #if [ "${MPI_ENABLED}" = "true" ]; then
+                    #    LINKER="mpifort"
+                    #fi
+                fi
             fi
-            if [ "${PM_COMPILER_SUITE}" = "gnu" ]; then
-                csvLinkerList="gfortran"
-                #csvLinkerList="${clist}"
-                #if [ "${MPI_ENABLED}" = "true" ]; then
-                #    LINKER="mpifort"
-                #fi
-            fi
-        fi
 
-        for LINKER in $(echo ${csvLinkerList} | sed "s/,/ /g")
-        do
+            for LINKER in $(echo ${csvLinkerList} | sed "s/,/ /g")
+            do
+
+                echo >&2
+                echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - linking ParaMonte example with ${LINKER}"
+                echo >&2 "-- ParaMonteExample${EXAMPLE_LANGUAGE} - ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} ${SRC_FILES//.$SRC_EXT/.o} ${PMLIB_FULL_NAME} -o ${PM_EXAM_EXE_NAME}"
+
+                ${LINKER} -dumpversion >/dev/null 2>&1 && \
+                ${LINKER} ${COMPILER_FLAGS} ${LINKER_FLAGS} ${SRC_FILES//.$SRC_EXT/.o} "${PMLIB_FULL_NAME}" -o ${PM_EXAM_EXE_NAME}
[…]
