### 2020-11-16 2dafaf0
minor enhancements
