+++ .travis.yml
@@ -34,6 +34,7 @@ matrix:
         - echo "pmRootDir = ${pmRootDir}"
         - echo "pmBinDir = ${pmBinDir}"
         - if ! [ -d "${pmBinDir}" ]; then mkdir -p ${pmRootDir}/bin/; fi
+        - ls "$(dpkg -L mpich)"
         - ls "$(dpkg -L openmpi)"
         - ls "$(dpkg -L openmpi-bin)"
         - ls "$(dpkg -L openmpi-dev)"
