### 2020-11-16 5a0435f
minor enhancements
