### 2020-11-16 178e6b2
minor enhancements
