### 2020-11-16 a02165a
minor enhancements
