### 2020-11-16 bed69ba
minor enhancements
