### 2020-11-16 3c914f9
minor enhancements
