+++ example/buildParaMonteExample.sh
@@ -282,62 +282,75 @@ do
 
             #### copy MPI shared library files
 
-            #if ! [ -z ${MPIEXEC_PATH+x} ] && [ "${MPI_ENABLED}" = "true" ]; then
-            #
-            #    echo >&2
-            #    echo >&2 "-- ParaMonteExample${LANG_NAME} - copying the the mpiexec executable file..."
-            #    echo >&2 "-- ParaMonteExample${LANG_NAME} - from: ${MPIEXEC_PATH}*"
-            #    echo >&2 "-- ParaMonteExample${LANG_NAME} -   to: ${ParaMonteExample_LIB_DIR_CURRENT}/"
-            #    yes | \cp -rf "${MPIEXEC_PATH}"* "${ParaMonteExample_LIB_DIR_CURRENT}/"
-            #    MPIEXEC_BIN_DIR=$(dirname "${MPIEXEC_PATH}")
-            #    MPIEXEC_ROOT_DIR="${MPIEXEC_BIN_DIR}"/..
-            #    MPIEXEC_LIB_DIR_LIST="${MPIEXEC_ROOT_DIR}/lib:${MPIEXEC_ROOT_DIR}/lib64"
-            #    keyList="libmpi:libmpifort"
-            #    for key in ${keyList//:/ }
-            #    do
-            #        for MPIEXEC_LIB_DIR in ${MPIEXEC_LIB_DIR_LIST//:/ }
-            #        do
-            #            if [ -d "${MPIEXEC_LIB_DIR}" ]; then
-            #                find "${MPIEXEC_LIB_DIR}" -name "${key}.so"* -print0 |
-            #                while IFS= read -r -d '' libMpiPath; do
-            #                    echo >&2
-            #                    echo >&2 "-- ParaMonteExample${LANG_NAME} - copying the ParaMonte library dll dependency file..."
-            #                    echo >&2 "-- ParaMonteExample${LANG_NAME} - from: ${libMpiPath}"
-            #                    echo >&2 "-- ParaMonteExample${LANG_NAME} -   to: ${ParaMonteExample_LIB_DIR_CURRENT}/"
-            #                    #yes | \cp -rf "${libMpiPath}" "${ParaMonteExample_LIB_DIR_CURRENT}/" && {
-            #                    yes | \cp -rf "${libMpiPath}" "${ParaMonteExample_LIB_DIR_CURRENT}/"
-            #                done
-            #            fi
-            #        done
-            #    done
-            #
-            #    #### copy numa shared files
-            #
-            #    FILE_LIST="libnuma.so:libpciaccess.so"
-            #    FILE_DIR_LIST="/usr/lib64"
-            #    for FILE in ${FILE_LIST//:/ }
-            #    do
-            #        for FILE_DIR in ${FILE_DIR_LIST//:/ }
-            #        do
-            #            if [ -d "${FILE_DIR}" ]; then
-            #                flist=$(( IFS=:; unset lsout; lsout=$(ls -dm "${FILE_DIR}/${FILE}"*); if ! [[ -z "${lsout// }" ]]; then echo "${lsout}, "; fi) 2>/dev/null)
-            #                for fpath in $(echo $flist | sed "s/,/ /g"); do
-            #                    echo >&2
-            #                    echo >&2 "-- ParaMonteExample${LANG_NAME} - copying the ParaMonte library dll dependency file..."
-            #                    echo >&2 "-- ParaMonteExample${LANG_NAME} - from: ${fpath}"
-            #                    echo >&2 "-- ParaMonteExample${LANG_NAME} -   to: ${ParaMonteExample_LIB_DIR_CURRENT}/"
[…]
