### 2020-11-16 ef78f1c
minor enhancements
