+++ example/buildParaMonteExample.sh
@@ -228,6 +228,8 @@ do
 
     if [ "${shared_enabled}" = "true" ]; then
 
+        #### GNU
+
         if [ "${PMCS}" = "gnu" ] && ! [ "${isMacOS}" = "true" ] && ! [ "${CAF_ENABLED}" = "true" ]; then # caf does not have lib dependency
 
             #### copy Fortran compiler shared library files
@@ -352,6 +354,36 @@ do
 
             fi
 
+        fi
+
+        #### intel
+
+        if [ "${PMCS}" = "intel" ] && ! [ -z ${MPIEXEC_PATH+x} ] && [ "${MPI_ENABLED}" = "true" ]; then
+            #echo >&2
+            #echo >&2 "-- ParaMonteExample${LANG_NAME} - copying the the mpiexec executable file..."
+            #echo >&2 "-- ParaMonteExample${LANG_NAME} - from: ${MPIEXEC_PATH}*"
+            #echo >&2 "-- ParaMonteExample${LANG_NAME} -   to: ${ParaMonteExample_LIB_DIR_CURRENT}/"
+            #yes | \cp -rf "${MPIEXEC_PATH}"* "${ParaMonteExample_LIB_DIR_CURRENT}/"
+            MPI_BIN_DIR=$(dirname "${MPIEXEC_PATH}")
+            MPI_ROOT_DIR="${MPI_BIN_DIR}"/..
+            MPI_LIB_DIR_LIST="${MPI_ROOT_DIR}/lib:${MPI_ROOT_DIR}/lib64"
+            # copy all so files
+            for MPI_LIB_DIR in ${MPI_LIB_DIR_LIST//:/ }
+            do
+                echo >&2
+                echo >&2 "-- ParaMonteExample${LANG_NAME} - copying the ParaMonte library dll dependency file..."
+                echo >&2 "-- ParaMonteExample${LANG_NAME} - from: ${MPI_LIB_DIR}/libmpifort.so*"
+                echo >&2 "-- ParaMonteExample${LANG_NAME} - from: ${MPI_LIB_DIR}/libmpi.so*"
+                echo >&2 "-- ParaMonteExample${LANG_NAME} -   to: ${ParaMonteExample_LIB_DIR_CURRENT}/"
+                thisIntelBuild="${BTYPE}"; if [ "${BTYPE}" = "testing" ]; then thisIntelBuild="release"; fi
+                if [ -d "${MPI_LIB_DIR}" ]; then
+                    cp -rf "${MPI_LIB_DIR}/"libmpifort.so* "${ParaMonteExample_LIB_DIR_CURRENT}/" && \
+                    cp -rf "${MPI_LIB_DIR}/${thisIntelBuild}_mt/"libmpi.so* "${ParaMonteExample_LIB_DIR_CURRENT}/" && \
+                    break || {
+                        echo >&2 "-- ParaMonteExample${LANG_NAME} - copy failed. skipping..."
+                    }
+                fi
+            done
         fi
 
     fi
