+++ src/kernel/System_mod.f90
@@ -55,12 +55,11 @@ module System_mod
 
     !> The `RandomFileName_type` class.
     type :: RandomFileName_type
-        character(:), allocatable       :: path !< The full path to the randomly-generated unique file name.
-        character(:), allocatable       :: dir  !< The directory within which is the unique new file is supposed to be generated.
-        character(:), allocatable       :: key  !< The optionally user-specified file prefix for the unique file name.
-        character(:), allocatable       :: ext  !< The optionally user-specified file extension.
-        type(Err_type)                  :: Err  !< An object of class [Err_type](@ref err_mod::err_type) indicating whether
-                                                !! any error has occurred during the file name generation.
+        character(:), allocatable       :: path                     !< The full path to the randomly-generated unique file name.
+        character(:), allocatable       :: dir                      !< The directory within which is the unique new file is supposed to be generated.
+        character(:), allocatable       :: key                      !< The optionally user-specified file prefix for the unique file name.
+        character(:), allocatable       :: ext                      !< The optionally user-specified file extension.
+        type(Err_type)                  :: Err                      !< An object of class [Err_type](@ref err_mod::err_type).
     end type RandomFileName_type
 
     !> The `RandomFileName_type` constructor.
@@ -70,11 +69,11 @@ module System_mod
 
     !> The `SystemInfo_type` class.
     type :: SystemInfo_type
-        integer(IK)                     :: nRecord  !< The number of elements of the vector `List`.
-        type(CharVec_type), allocatable :: List(:)  !< An array of length `nRecord` of strings, each element of which represents
-                                                    !! one line in the output system information.
-        type(Err_type)                  :: Err      !< An object of class [Err_type](@ref err_mod::err_type) indicating whether
-                                                    !! any error has occurred during information collection.
+        integer(IK)                     :: nRecord                  !< The number of elements of the vector `List`.
+        type(CharVec_type), allocatable :: List(:)                  !< An array of length `nRecord` of strings, each element of which represents
+                                                                    !! one line in the output system information.
+        type(Err_type)                  :: Err                      !< An object of class [Err_type](@ref err_mod::err_type) indicating whether
+                                                                    !! any error has occurred during information collection.
     contains
         procedure, nopass               :: get => getSystemInfo
     end type SystemInfo_type
@@ -84,15 +83,38 @@ module System_mod
         module procedure                :: constructSystemInfo
     end interface SystemInfo_type
 
+    !> The Shell name type.
+    type, private :: ShellName_type
+        character(:), allocatable       :: current                  !< The name of the current runtime shell.
+        character(:), allocatable       :: default                  !< The name of the default runtime shell.
+    end type ShellName_type
+
+    !> The `Shell_type` class.
+    type :: Shell_type
[…]
