### 2020-11-19 47e0372
minor enhancements
