### 2020-11-19 c97acef
minor enhancements
