### 2020-11-19 41fc726
minor enhancements
