### 2020-11-19 fbde753
minor enhancements
