### 2020-11-19 c18a244
minor enhancements
