### 2020-11-19 77d229b
minor enhancements
