+++ src/test/Test_BandSpectrum_mod.f90
@@ -123,19 +123,19 @@ contains
         Test = Test_type(moduleName=MODULE_NAME)
         call Test%run(test_getEbreak, "test_getEbreak()")
         call Test%run(test_getPhotonFluence_1, "test_getPhotonFluence_1")
-        !call Test%run(test_getPhotonFluence_2, "test_getPhotonFluence_2")
-        !call Test%run(test_getPhotonFluence_3, "test_getPhotonFluence_3")
-        !call Test%run(test_getPhotonFluence_4, "test_getPhotonFluence_4")
-        !call Test%run(test_getEnergyFluence_1, "test_getEnergyFluence_1")
-        !call Test%run(test_getEnergyFluence_2, "test_getEnergyFluence_2")
-        !call Test%run(test_getEnergyFluence_3, "test_getEnergyFluence_3")
-        !call Test%run(test_getEnergyFluence_4, "test_getEnergyFluence_4")
-        !call Test%run(test_getPhotonFluenceFromEnergyFluence_1, "test_getPhotonFluenceFromEnergyFluence_1")
-        !call Test%run(test_getPhotonFluenceFromEnergyFluence_2, "test_getPhotonFluenceFromEnergyFluence_2")
-        !call Test%run(test_getPhotonFluenceFromEnergyFluence_3, "test_getPhotonFluenceFromEnergyFluence_3")
-        !call Test%run(test_getPhotonFluenceFromEnergyFluence_4, "test_getPhotonFluenceFromEnergyFluence_4")
-        !call Test%run(test_getPhotonFluenceFromEnergyFluence_5, "test_getPhotonFluenceFromEnergyFluence_5")
-        !call Test%finalize()
+        call Test%run(test_getPhotonFluence_2, "test_getPhotonFluence_2")
+        call Test%run(test_getPhotonFluence_3, "test_getPhotonFluence_3")
+        call Test%run(test_getPhotonFluence_4, "test_getPhotonFluence_4")
+        call Test%run(test_getEnergyFluence_1, "test_getEnergyFluence_1")
+        call Test%run(test_getEnergyFluence_2, "test_getEnergyFluence_2")
+        call Test%run(test_getEnergyFluence_3, "test_getEnergyFluence_3")
+        call Test%run(test_getEnergyFluence_4, "test_getEnergyFluence_4")
+        call Test%run(test_getPhotonFluenceFromEnergyFluence_1, "test_getPhotonFluenceFromEnergyFluence_1")
+        call Test%run(test_getPhotonFluenceFromEnergyFluence_2, "test_getPhotonFluenceFromEnergyFluence_2")
+        call Test%run(test_getPhotonFluenceFromEnergyFluence_3, "test_getPhotonFluenceFromEnergyFluence_3")
+        call Test%run(test_getPhotonFluenceFromEnergyFluence_4, "test_getPhotonFluenceFromEnergyFluence_4")
+        call Test%run(test_getPhotonFluenceFromEnergyFluence_5, "test_getPhotonFluenceFromEnergyFluence_5")
+        call Test%finalize()
     end subroutine test_BandSpectrum
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
