### 2020-11-19 4cf1239
minor enhancements
