### 2020-11-19 bd1d5b2
minor enhancements
