+++ src/kernel/CorrCoef_mod.f90
@@ -66,6 +66,7 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
+    !> \brief
     !> Return the Spearman correlation coefficient of the two input data arrays.
     !> 
     !> @param[in] ndata             :   The length of the two input data arrays.
