### 2020-11-19 787428d
minor enhancements
