### 2020-11-19 f4f82f3
minor enhancements
