### 2020-11-26 807d55c
kernel tests now enabled by default.
