### 2020-11-26 0af2327
minor enhancements
