### 2020-11-26 f76c0fb
minor enhancements
