### 2020-11-26 dcffa04
minor enhancements
