### 2020-11-26 9a938e1
minor enhancements
