+++ src/kernel/System_mod.f90
@@ -91,6 +91,7 @@ module System_mod
 
     !> The `Shell_type` class.
     type :: Shell_type
+        logical                         :: isSh         = .false.   !< The logical value indicating whether the shell is Unix sh.
         logical                         :: isCMD        = .false.   !< The logical value indicating whether the shell is Windows CMD.
         logical                         :: isZsh        = .false.   !< The logical value indicating whether the shell is Unix zsh.
         logical                         :: isCsh        = .false.   !< The logical value indicating whether the shell is Unix csh.
@@ -160,6 +161,7 @@ module System_mod
     ! cache the OS query result to speed up code
 
     logical                             :: mv_osCacheEnabled = .false.
+    logical                             :: mv_shCacheEnabled = .false.
     type(OS_type)                       :: mv_OS
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -192,7 +194,7 @@ contains
     !> Query all attributes of the [OS_type](@ref os_type) class: `name`, `slash`, `isWindows`, `Err`.
     !>
     !> \param[out]  OS                  :   An object of class [OS_type](@ref os_type).
-    !> \param[in]   shellQueryEnabled   :   A logical variable indicating if the type and name of the current 
+    !> \param[in]   shellQueryEnabled   :   A logical variable indicating if the type and name of the current
     !>                                      runtime shell should be queried or not (optional, default = `.false.`).
     subroutine queryOS(OS, shellQueryEnabled)
 #if defined DLL_ENABLED && !defined CFI_ENABLED
@@ -211,20 +213,34 @@ contains
         character(:)    , allocatable           :: osname
 #endif
 
+        shellQueryEnabledDefault = .false.
+        if (present(shellQueryEnabled)) shellQueryEnabledDefault = shellQueryEnabled
+        OS%Err%occurred = .false.
+        OS%Err%msg = ""
+
         if (mv_osCacheEnabled) then
+
             OS%name         = mv_OS%name
             OS%slash        = mv_OS%slash
             OS%isWindows    = mv_OS%isWindows
             OS%isDarwin     = mv_OS%isDarwin
             OS%isLinux      = mv_OS%isLinux
-            OS%Shell        = mv_OS%Shell
-            return
-        end if
 
-        shellQueryEnabledDefault = .false.
-        if (present(shellQueryEnabled)) shellQueryEnabledDefault = shellQueryEnabled
-        OS%Err%occurred = .false.
-        OS%Err%msg = ""
+            if (mv_shCacheEnabled) then
+                OS%Shell    = mv_OS%Shell
+            else
+                mv_shCacheEnabled = .true.
+                call OS%Shell%query(OS%isWindows)
+                if (OS%Shell%Err%occurred) then
+                    OS%Err = OS%Shell%Err
+                    return
+                end if
+                mv_OS%Shell = OS%Shell
+            end if
+
+            return
+
+        end if
 
 #if defined OS_IS_WINDOWS
 
@@ -373,7 +389,6 @@ contains
 
 #endif
 
-        if (shellQueryEnabledDefault) call OS%Shell%query(OS%isWindows)
 
         mv_osCacheEnabled = .true.
         mv_OS%name      = OS%name
@@ -381,7 +396,22 @@ contains
         mv_OS%isWindows = OS%isWindows
         mv_OS%isDarwin  = OS%isDarwin
         mv_OS%isLinux   = OS%isLinux
[…]
