### 2020-11-26 5bcac13
minor enhancements
