+++ src/kernel/Parallelism_mod.f90
@@ -53,6 +53,21 @@ module Parallelism_mod
 
     character(*), parameter :: MODULE_NAME = "@Parallelism_mod"
 
+    !> The [Image_type](@ref image_type) class.
+    type                            :: Image_type
+        integer(IK)                 :: id                   !<  The ID of the runtime parallel process: 1, 2, 3, ...
+        integer(IK)                 :: count                !<  The total number of runtime parallel processes available.
+        logical                     :: isFirst = .false.    !<  A logical flag indicating whether the current process is ID #1.
+        logical                     :: isNotFirst = .false. !<  A logical flag indicating whether the current process is NOT ID #1.
+        logical                     :: isLeader = .false.   !<  A logical flag indicating whether the current process is a leader. This must be defined by the user.
+        logical                     :: isRooter = .false.   !<  A logical flag indicating whether the current process is a follower. This must be defined by the user.
+        character(:), allocatable   :: name                 !<  The name of the current process as a string with the format `"@process(id)"`.
+    contains
+        procedure, nopass           :: sync => syncImages
+        procedure, pass             :: query => queryImage
+        procedure, nopass           :: finalize => finalizeImages
+    end type Image_type
+
     type, private :: Maximum_type
         real(RK)                    :: value
         integer(IK)                 :: nproc
@@ -96,12 +111,109 @@ module Parallelism_mod
         module procedure :: constructForkJoin
     end interface ForkJoin_type
 
-    !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 contains
 
-    !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
+    !> \brief
+    !> Return the statistics of the parallel processors available, depending on the type of parallelism requested.
+    !> This is a dynamic member of the [Image_type](@ref image_type) class.
+    !>
+    !> @param[inout]    Image       :   An object of class [Image_type](@ref image_type).
+    !>                                  On output, all properties of `Image` will be reset,
+    !>                                  except the attributes `isLeader` and `isRooter`.
+    !>
+    !> \warning
+    !> This routine must not contain any synchronization statements under any circumstances.
+    subroutine queryImage(Image)
+
+        use Statistics_mod, only: fitGeoCyclicLogPDF
[…]
