+++ src/kernel/tests/Test_Cosmology_mod.f90
@@ -68,8 +68,8 @@ contains
         call Test%run(test_ldiswickram, "test_ldiswickram")
         call Test%run(test_getLogLumDisWicMpc, "test_getLogLumDisWicMpc")
 #if defined CODECOV_ENABLED
-        call Test%run(test_getLookBackTime_1, "test_getLookBackTime_1") ! The internal function passing as actual argument causes segfault with Gfortran (any version) on Linux subsystem for Windows.
-        call Test%run(test_getLookBackTime_2, "test_getLookBackTime_2") ! The internal function passing as actual argument causes segfault with Gfortran (any version) on Linux subsystem for Windows.
+        call Test%run(test_getLookBackTime_1, "test_getLookBackTime_1") ! The internal function passing as actual argument causes segfault with Gfortran (any version) on Windows subsystem for Linux.
+        call Test%run(test_getLookBackTime_2, "test_getLookBackTime_2") ! The internal function passing as actual argument causes segfault with Gfortran (any version) on Windows subsystem for Linux.
 #endif
         call Test%run(test_getUniverseAgeDerivative, "test_getLookBackTime_2")
         call Test%finalize()
