### 2020-11-27 b5f370f
minor enhancements
