+++ src/kernel/tests/Test_mod.f90
@@ -145,7 +145,7 @@ contains
         use String_mod, only: num2str
         use Err_mod, only: abort
         implicit none
-        integer(IK)                 :: percentageTestPassed
+        real(RK)                    :: percentageTestPassed
         integer(IK)                 :: ntotal
         character(:), allocatable   :: color
         character(:), allocatable   :: msg
@@ -153,12 +153,12 @@ contains
 
         ntotal = mv_npass + mv_nfail
 
-        percentageTestPassed = nint( 100_IK * mv_npass / real(ntotal, kind=RK) )
+        percentageTestPassed = 100_IK * mv_npass / real(ntotal, kind=RK)
 
         if (mv_Image%isFirst) then
 
-            color = "green"; if (percentageTestPassed==0_IK) color = "red"
-            msg = style( num2str(percentageTestPassed,"(g0)")//"% of "//num2str(ntotal)//" tests passed. ", "bright", color)
+            color = "green"; if (nint(percentageTestPassed,kind=IK)==0_IK) color = "red"
+            msg = style( num2str(percentageTestPassed,"(f0.2)")//"% of "//num2str(ntotal)//" tests passed. ", "bright", color)
 
             color = "green"; if (mv_nfail>0_IK) color = "red"
             msg = msg // style( num2str(mv_nfail,"(g0)")//" tests failed. ", "bright", color)
