### 2020-11-27 2fbe3cc
minor enhancements
