+++ src/kernel/ParaDXXX_mod@Setup_smod.inc.f90
@@ -1698,7 +1698,7 @@ contains
                                                                " and "//num2str(imageMinProbKS)//"." )
                             end if
 #if defined CAF_ENABLED || MPI_ENABLED
-                            call execute_command_line(" ")
+                            call execute_command_line(" ", cmdstat = self%Err%stat)
                             flush(output_unit)
 #if defined CAF_ENABLED
                             sync all
