### 2020-11-27 b6ea2f3
minor enhancements
