+++ buildParaMonte.sh
@@ -538,7 +538,7 @@ fi
 openCoarraysVersion="2.9.0"
 gnuVersionOpenCoarrays="10.1.0"
 mpichVersionOpenCoarrays="3.2"
-gnuVersionParaMonteCompatible="7.0.0"
+gnuVersionParaMonteCompatible="8.4.0"
 cmakeVersionParaMonteCompatible="3.14.0"
 intelVersionParaMonteCompatible="18.0.0"
 
@@ -831,91 +831,95 @@ fi
 #### identify the MPI wrappers
 ####################################################################################################################################
 
-gnuCMpiWrapperList="mpicc"
-gnuFortranMpiWrapperList="mpifort:mpif90"
+if [ "${MPI_ENABLED}" = "true" ]; then
 
-intelCMpiWrapperList="mpiicc"
-intelFortranMpiWrapperList="mpiifort"
+    gnuCMpiWrapperList="mpicc"
+    gnuFortranMpiWrapperList="mpifort:mpif90"
 
-for SUITE in $SUITE_LIST
-do
+    intelCMpiWrapperList="mpiicc"
+    intelFortranMpiWrapperList="mpiifort"
 
-    echo >&2
-    echo >&2 "-- ${BUILD_NAME}MPI - checking for ${SUITE} MPI wrappers and libraries presence..."
-    echo >&2
-
-    for LANG in $LANG_LIST
+    for SUITE in $SUITE_LIST
     do
 
-        suiteLangMpiWrapperName="${SUITE}${LANG}MpiWrapperName"
-        suiteLangMpiWrapperList="${SUITE}${LANG}MpiWrapperList"
-        suiteLangMpiWrapperPath="${SUITE}${LANG}MpiWrapperPath"
+        echo >&2
+        echo >&2 "-- ${BUILD_NAME}MPI - checking for ${SUITE} MPI wrappers and libraries presence..."
+        echo >&2
 
-        searchFailed=false
-        for mpiWrapperName in ${!suiteLangMpiWrapperList//:/ }
+        for LANG in $LANG_LIST
         do
 
-            eval ${suiteLangMpiWrapperName}="${mpiWrapperName}"
+            suiteLangMpiWrapperName="${SUITE}${LANG}MpiWrapperName"
+            suiteLangMpiWrapperList="${SUITE}${LANG}MpiWrapperList"
+            suiteLangMpiWrapperPath="${SUITE}${LANG}MpiWrapperPath"
 
-            if eval "command -v ${!suiteLangMpiWrapperName} >/dev/null 2>&1"; then
+            searchFailed=false
+            for mpiWrapperName in ${!suiteLangMpiWrapperList//:/ }
+            do
 
-                eval "unset ${suiteLangMpiWrapperPath}"
-                eval ${suiteLangMpiWrapperPath}='$(command -v ${!suiteLangMpiWrapperName})'
+                eval ${suiteLangMpiWrapperName}="${mpiWrapperName}"
 
-                # ensure the MPI library is not intel
+                if eval "command -v ${!suiteLangMpiWrapperName} >/dev/null 2>&1"; then
 
-                if [[ "${SUITE}" = "gnu" && "${!suiteLangMpiWrapperPath}" =~ .*"intel".* ]]; then
+                    eval "unset ${suiteLangMpiWrapperPath}"
+                    eval ${suiteLangMpiWrapperPath}='$(command -v ${!suiteLangMpiWrapperName})'
 
-                    searchFailed=true
+                    # ensure the MPI library is not intel
 
-                else
+                    if [[ "${SUITE}" = "gnu" && "${!suiteLangMpiWrapperPath}" =~ .*"intel".* ]]; then
 
[…]
