### 2020-11-27 cb83488
minor enhancements
