### 2020-11-27 0d2364c
minor enhancements
