### 2020-11-27 912c518
minor enhancements
