+++ src/kernel/tests/Test_StarFormation_mod.f90
@@ -101,7 +101,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 9.87062241049176_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -132,7 +132,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 11.5870199718159_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -163,7 +163,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 13.9081968356527_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -194,7 +194,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 7.62065413866848_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -225,7 +225,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 7.87332272998867_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -256,7 +256,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 6.85265527251550_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -287,7 +287,7 @@ contains
         real(RK), parameter :: zplus1 = 1.e1_RK
         real(RK), parameter :: logzplus1 = log(zplus1)
         real(RK), parameter :: logRate_ref = 6.81074639841109_RK
-        real(RK), parameter :: tolerance = 1.e-12_RK
+        real(RK), parameter :: tolerance = 1.e-10_RK
         real(RK)            :: difference
         real(RK)            :: logRate
 
@@ -315,7 +315,7 @@ contains
 
         logical :: assertion
         real(RK)    , parameter :: z = 1.e-2_RK
-        real(RK)    , parameter :: tolerance = 1.e-12_RK
+        real(RK)    , parameter :: tolerance = 1.e-10_RK
         real(RK)    , parameter :: binaryMergerRateS15_ref = .000000000000000_RK
         real(RK)                :: binaryMergerRateS15
[…]
