### 2020-11-27 913f36f
minor enhancements
