### 2020-11-27 087515d
minor enhancements
