+++ src/kernel/ParaDXXXProposal_mod.inc.f90
@@ -304,11 +304,10 @@ contains
 
         block
             use Matrix_mod, only: getCholeskyFactor
-            !real(RK), allocatable :: CholeskyLower(:,:) ! dummy variable to avoid copy in / copy out
-            !CholeskyLower = comv_CholDiagLower(1:ndim,1:ndim,0)
-            !call getCholeskyFactor( ndim, CholeskyLower, comv_CholDiagLower(1:ndim,0,0) )
-            !comv_CholDiagLower(1:ndim,1:ndim,0) = CholeskyLower
-            call getCholeskyFactor( ndim, comv_CholDiagLower(:,:,0), comv_CholDiagLower(1:ndim,0,0) )
+            real(RK), allocatable :: CholeskyLower(:,:) ! dummy variable to avoid copy in / copy out
+            CholeskyLower = comv_CholDiagLower(1:ndim,1:ndim,0)
+            call getCholeskyFactor( ndim, CholeskyLower, comv_CholDiagLower(1:ndim,0,0) )
+            comv_CholDiagLower(1:ndim,1:ndim,0) = CholeskyLower
             call getInvCovMat()
         end block
         if (comv_CholDiagLower(1,0,0)<0._RK) then
@@ -668,8 +667,7 @@ contains
 
                 ! ensure the old Cholesky factorization can be recovered
 
-                !call getCholeskyFactor( nd, comv_CholDiagLower(1:nd,1:nd,0), comv_CholDiagLower(1:nd,0,0) )
-                call getCholeskyFactor( nd, comv_CholDiagLower(:,:,0), comv_CholDiagLower(1:nd,0,0) ) ! avoid temporary array creation
+                call getCholeskyFactor( nd, comv_CholDiagLower(1:nd,1:nd,0), comv_CholDiagLower(1:nd,0,0) )
                 if (comv_CholDiagLower(1,0,0)<0._RK) then
                     write(mc_logFileUnit,"(A)")
                     write(mc_logFileUnit,"(A)") "Singular covariance matrix detected:"
