+++ CHANGES.md
@@ -5,6 +5,39 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 ## **Version 1.x.x**  
 
+### **Version 1.4.2** -- November 15, 2020
+
+**Enhancements**  
+
++   Minor enhancements to the ParaMonte C/C++/Fortran example build scripts `build.sh` and `build.bat`.  
++   The default build settings are now limited to `heap` memory allocation with `dynamic` library builds
+    for only serial and MPI parallelization for all languages. The `stack` memory allocation results in 
+    a ~10% gain in the efficiency of the code. The benefits of stack-memory builds are marginal and are 
+    often problematic, in particular, for usage with non-compiled languages. However, users can still 
+    build the library with `stack` memory allocation by specifying the appropriate build flags with
+    the `install.sh` on Unix or `install.bat` script on Windows systems. For further information, 
+    see the installation guidelines on the ParaMonte documentation website.
+
+    order that is automatically inferred at the compilation time. Also, the dependencies 
+    on the MPI compiler wrappers is now removed as the MPI libraries are not required to 
+    build the ParaMonte examples, even in cases of parallel ParaMonte example builds.
+
+**Compiler support**  
+  
+| Compiler Suite                    | Windows (64bit) | Linux | macOS |  
+|----------------------------------:|:---------------:|:-----:|:-----:|  
+| GNU Compiler Collection > 8.4     | ✅              | ✅    | ✅    |  
+| Intel Parallel Studio > 18.0.0    | ✅              | ✅    | ✅    |  
+| Microsoft C/C++ Compiler > 16.0.0 | ✅              | ❌    | ❌    |  
+
+**Compiler / MPI library used for this binary release**  
+
++   **Windows**: Intel Parallel Studio Version 19.0.4.245 Build 20190417 / Intel(R) MPI Library 2019 Update 4 for Windows  
++   **Linux**: Intel Parallel Studio Version 18.0.2 20180210 / Intel(R) MPI Library for Linux OS, Version 2018 Update 2 Build 20180125  
++   **Linux**: GNU 10.1.0 / MPICH 3.2  
++   **macOS**: Intel Parallel Studio Version 19.1.0.216 20200306  
++   **macOS**: GNU 10.2.0 / Open-MPI 4.0.5  
+
 ### **Version 1.4.1** -- November 15, 2020
 
 **Enhancements**  
