+++ .travis.yml
@@ -1,6 +1,11 @@
 language: c
 sudo: true
 
+branches:
+  only:
+    - master
+    - dev
+
 before_cache:
   - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then brew cleanup; fi
   - if [ "${TRAVIS_OS_NAME}" = "osx" ]; then find /usr/local/Homebrew \! -regex ".+\.git.+" -delete; fi
