### 2020-11-27 afc1407
minor enhancements
