+++ src/kernel/ParaMonte_mod.f90
@@ -317,7 +317,13 @@ contains
 #if defined PARAMONTE_VERSION
         self%version  = "Version " // PARAMONTE_VERSION
 #else
-        self%version  = "Unknown Version"
+        ! WARNING: The following ParaMonte library version tag as it will be replaced by the version 
+        ! WARNING: that is generated by the preprocessing build script of the ParaMonte library.
+        ! WARNING: This is superior to the above method of using the compiler Preprocessor
+        ! WARNING: since CMAKE triggers a complete lengthy rebuild of the library when the
+        ! WARNING: only the library version has changed.
+        self%version  = "Unknown ParaMonte Version"
+#include "ParaMonte_mod@version@kernel.inc.f90"
 #endif
 
         ! setup general processor / coarray image variables
