+++ buildParaMonte.sh
@@ -69,6 +69,7 @@ workingDir="$(pwd)"
 sourceFileDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 ParaMonte_ROOT_DIR="${sourceFileDir}"; export ParaMonte_ROOT_DIR
 ParaMonteKernel_SRC_DIR="${ParaMonte_ROOT_DIR}/src/kernel"; export ParaMonteKernel_SRC_DIR
+ParaMonteKernelVersion_SRC_INC_PATH="${ParaMonteKernel_SRC_DIR}/ParaMonte_mod@version@kernel.inc.f90"; export ParaMonteKernel_SRC_DIR
 ParaMonteTest_SRC_DIR="${ParaMonte_ROOT_DIR}/src/kernel/tests"; export ParaMonteTest_SRC_DIR
 #export ParaMonte_ROOT_DIR="${ParaMonte_ROOT_DIR:-${PWD%/}}"
 
@@ -261,9 +262,17 @@ export ARCHITECTURE
 
 read -r ParaMonteVersion < .VERSION
 
+unset fppParaMonteVersion
 unset FPP_PARAMONTE_VERSION_FLAG
 if ! [ -z ${ParaMonteVersion+x} ]; then
     export ParaMonteVersion
+    # NOTE: uncomment the following line, if you wish to insert the ParaMonte version to the source code via Cmake via the compiler preprocessor.
+    # fppParaMonteVersion="${ParaMonteVersion}"; export fppParaMonteVersion
+    if [ -z ${fppParaMonteVersion+x} ]; then
+        echo "! WARNING: DO NOT CHANGE THE CONTENTS OF THIS FILE MANUALLY." > "${ParaMonteKernelVersion_SRC_INC_PATH}"
+        echo "! WARNING: This file is auto-generated by the ParaMonte build scripts." >> "${ParaMonteKernelVersion_SRC_INC_PATH}"
+        echo "        self%version  = \"${ParaMonteVersion}\"" >> "${ParaMonteKernelVersion_SRC_INC_PATH}"
+    fi
 fi
 
 #echo "$(cat ./auxil/.ParaMonteBanner)"
