+++ buildParaMonte.sh
@@ -265,14 +265,18 @@ read -r ParaMonteVersion < .VERSION
 unset fppParaMonteVersion
 unset FPP_PARAMONTE_VERSION_FLAG
 if ! [ -z ${ParaMonteVersion+x} ]; then
+
     export ParaMonteVersion
+
     # NOTE: uncomment the following line, if you wish to insert the ParaMonte version to the source code via Cmake via the compiler preprocessor.
     # fppParaMonteVersion="${ParaMonteVersion}"; export fppParaMonteVersion
-    if [ -z ${fppParaMonteVersion+x} ]; then
-        echo "! WARNING: DO NOT CHANGE THE CONTENTS OF THIS FILE MANUALLY." > "${ParaMonteKernelVersion_SRC_INC_PATH}"
-        echo "! WARNING: This file is auto-generated by the ParaMonte build scripts." >> "${ParaMonteKernelVersion_SRC_INC_PATH}"
-        echo "        self%version  = \"${ParaMonteVersion}\"" >> "${ParaMonteKernelVersion_SRC_INC_PATH}"
-    fi
+
+    # write the version source include file.
+
+    echo "! WARNING - DO NOT CHANGE THE CONTENTS OF THIS FILE MANUALLY." > "${ParaMonteKernelVersion_SRC_INC_PATH}"
+    echo "! WARNING - This file is auto-generated by the ParaMonte build scripts." >> "${ParaMonteKernelVersion_SRC_INC_PATH}"
+    echo "        self%version = \"${ParaMonteVersion}\"" >> "${ParaMonteKernelVersion_SRC_INC_PATH}"
+
 fi
 
 #echo "$(cat ./auxil/.ParaMonteBanner)"
