### 2020-11-27 507a501
minor enhancements
