### 2020-11-28 796cd3c
minor enhancements
