+++ src/kernel/ParaMonte_mod@version@kernel.inc.f90
@@ -1,3 +1,3 @@
 ! WARNING - DO NOT CHANGE THE CONTENTS OF THIS FILE MANUALLY.
 ! WARNING - This file is auto-generated by the ParaMonte build scripts.
-        self%version = "1.4.2"
+self%version = "1.4.2"
