+++ src/kernel/tests/Test_ParaMonte_mod.f90
@@ -69,7 +69,6 @@ contains
         implicit none
         call test_ParaDRAM()
         call test_ParaDISE()
-        call 
     end subroutine test_ParaMonte
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
