+++ src/kernel/tests/Test_ParaDRAM_mod.f90
@@ -43,19 +43,18 @@
 !>  \brief This module contains tests of the module [ParaMonte_mod](@ref paramonte_mod).
 !>  @author Amir Shahmoradi
 
-module Test_ParaMonte_mod
+module Test_ParaDRAM_mod
 
     use Constants_mod, only: IK, RK
     use Test_mod, only: Test_type
+    use ParaDRAM_mod
 
+    !use Statistics_mod, only: paradramPrintEnabled
     implicit none
-
-    ! Standard MultiVariate Normal (SMVN) specifications: https://en.wikipedia.org/wiki/Multivariate_normal_distribution
-    integer(IK) , parameter :: NDIM = 1_IK                                              ! number of dimensions of the distribution
-    real(RK)    , parameter :: LOG_SMVN_COEF = NDIM*log(1._RK/sqrt(2._RK*acos(-1._RK))) ! log(1/sqrt(2*Pi)^ndim)
+    !paradramPrintEnabled = .true.
 
     private
-    public :: test_ParaMonte
+    public :: test_ParaDRAM
 
     type(Test_type) :: Test
 
@@ -65,109 +64,109 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-    subroutine test_ParaMonte()
-        implicit none
-        call test_ParaDRAM()
-        call test_ParaDISE()
-    end subroutine test_ParaMonte
-
-!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
     subroutine test_ParaDRAM()
-
-        use ParaDRAM_mod, only: MODULE_NAME
         implicit none
-
         Test = Test_type(moduleName=MODULE_NAME)
-
-        call test_runParaDRAM()
+        call Test%run(test_runSampler_1, "test_runSampler_1")
+        call Test%run(test_runSampler_2, "test_runSampler_2")
+        call Test%run(test_runSampler_3, "test_runSampler_3")
         call Test%finalize()
-
     end subroutine test_ParaDRAM
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-    subroutine test_ParaDISE()
-
-        use ParaDISE_mod, only: MODULE_NAME
+    !> \brief
+    !> Test the ParaDRAM sampler with no input arguments or input file.
+    function test_runSampler_1() result(assertion)
         implicit none
-
-        Test = Test_type(moduleName=MODULE_NAME)
-
-        call test_runParaDISE()
-        call Test%finalize()
-
-    end subroutine test_ParaDISE
-
-!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-    subroutine test_runParaDRAM()
-
[…]
