+++ cmake/Modules/SetParaMonteFlags.cmake
@@ -613,6 +613,14 @@ set(FPP_FLAGS
 #add_definitions(${FPP_FLAGS})
 
 #:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+#: set code coverage preprocessor flag
+#:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+
+if (CODECOV_ENABLED)
+    set(FPP_FLAGS "${FPP_FLAGS}" -DCODECOV_ENABLED )
+endif()
+
+#:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 #: set up coarray flags
 #:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
