+++ install.sh
@@ -513,7 +513,7 @@ if [ "${LANG_LIST}" = "matlab" ] || [ "${LANG_LIST}" = "python" ]; then
 fi
 
 if ! [ "${codecov_flag}" = "" ]; then
-    PARALLELISM_LIST="mpi"
+    PARALLELISM_LIST="none"
     BTYPE_LIST="release"
     LTYPE_LIST="static"
     LANG_LIST="fortran"
