+++ src/kernel/ParaMonteChainFileContents_mod.f90
@@ -939,4 +939,4 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-end module ParaMonteChainFileContents_mod
+end module ParaMonteChainFileContents_mod ! LCOV_EXCL_LINE
