+++ src/kernel/FileContents_mod.f90
@@ -130,10 +130,12 @@ contains
         end if
 
         call getNumRecordInFile(path,numRecord,Err)
+        ! LCOV_EXCL_START
         if (Err%occurred) then
             Err%msg = PROCEDURE_NAME // Err%msg
             return
         end if
+        ! LCOV_EXCL_STOP
 
         allocate(Contents(numRecord))
 
@@ -142,6 +144,7 @@ contains
             read(fileUnit,"(A)",iostat=Err%stat) record
             if (Err%stat==0) then
                 Contents(irecord)%record = trim(adjustl(record))
+                ! LCOV_EXCL_START
             elseif (is_iostat_end(iostat)) then
                 Err%occurred = .true.
                 Err%msg =   PROCEDURE_NAME // ": End-of-file error occurred while expecting " // &
@@ -157,15 +160,18 @@ contains
                 Err%msg =   PROCEDURE_NAME // ": Unknown error occurred while reading line number " // &
                             num2str(irecord) // " from file='" // path // "'."
                 return
+                ! LCOV_EXCL_STOP
             end if
         end do
 
         close(fileUnit,iostat=Err%stat,status=closeStatus)
+        ! LCOV_EXCL_START
         if (Err%stat>0) then
             Err%occurred = .true.
             Err%msg = PROCEDURE_NAME // "Error occurred while attempting to close or delete the open file='" // path // "'."
             return
         end if
+        ! LCOV_EXCL_STOP
 
     end subroutine getFileContents
 
@@ -203,31 +209,39 @@ contains
         ! GFortran 7.3 bug: If file is not open, compiler assumes internal file if `number` is specified, causing runtime error.
 
         inquire( file=filePath, exist=fileExists, opened=fileIsOpen, number=fileUnit, iostat=Err%stat )
+        ! LCOV_EXCL_START
         if (Err%stat/=0) then
             Err%occurred = .true.
             Err%msg = PROCEDURE_NAME // ": Error occurred while inquiring the status of file='" // filePath // "'."
             return
         end if
+        ! LCOV_EXCL_STOP
 
+        ! LCOV_EXCL_START
         if (.not.fileExists) then
             Err%occurred = .true.
             Err%msg = PROCEDURE_NAME // ": The input file='" // filePath // "' does not exist."
             return
         end if
+        ! LCOV_EXCL_STOP
 
         if (fileIsOpen) close(unit=fileUnit,iostat=Err%stat)
+        ! LCOV_EXCL_START
         if (Err%stat>0) then
             Err%occurred = .true.
             Err%msg = PROCEDURE_NAME // ": Error occurred while attempting to close the open input file='" // filePath // "'."
             return
         end if
+        ! LCOV_EXCL_STOP
 
         open(newunit=fileUnit,file=filePath,status="old",iostat=Err%stat)
+        ! LCOV_EXCL_START
         if (Err%stat>0) then
             Err%occurred = .true.
             Err%msg = PROCEDURE_NAME // ": Error occurred while opening input file='" // filePath // "'."
             return
         end if
+        ! LCOV_EXCL_STOP
 
         numRecord = 0_IK
         do
@@ -241,23 +255,27 @@ contains
                 cycle
             elseif(is_iostat_end(iostat)) then
                 exit
+            ! LCOV_EXCL_START
             else
                 Err%occurred = .true.
                 Err%stat = iostat
                 Err%msg = PROCEDURE_NAME // ": Error occurred while reading input file='" // filePath // "'."
                 return
             end if
+            ! LCOV_EXCL_STOP
[…]
