+++ src/kernel/SpecBase_InterfaceType_mod.f90
@@ -147,4 +147,4 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-end module SpecBase_InterfaceType_mod
+end module SpecBase_InterfaceType_mod ! LCOV_EXCL_LINE
