+++ src/kernel/SpecMCMC_ScaleFactor_mod.f90
@@ -217,4 +217,4 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-end module SpecMCMC_ScaleFactor_mod
+end module SpecMCMC_ScaleFactor_mod ! LCOV_EXCL_LINE
