+++ src/kernel/SpecBase_MpiFinalizeRequested_mod.f90
@@ -113,4 +113,4 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-end module SpecBase_MpiFinalizeRequested_mod
+end module SpecBase_MpiFinalizeRequested_mod ! LCOV_EXCL_LINE
