+++ src/kernel/ParaMonte_mod.f90
@@ -1188,4 +1188,4 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-end module ParaMonte_mod
+end module ParaMonte_mod ! LCOV_EXCL_LINE
