### 2020-11-28 06b1618
minor enhancements
