### 2020-11-28 a22c488
minor enhancements
