+++ src/kernel/tests/Test_Batse_mod.f90
@@ -272,19 +272,19 @@ contains
         implicit none
         logical                 :: assertion
         real(RK), parameter     :: tolerance = sqrt(epsilon(1._RK))
-        real(RK), parameter     :: logEffectivePeakPhotonFlux_ref = real(THRESH_ERFC_AMP,RK)
+        real(RK), parameter     :: logEffectivePeakPhotonFlux_ref = -real(THRESH_ERFC_AMP,RK)
         real(RK)                :: logEffectivePeakPhotonFlux
         real(RK)                :: difference
         logEffectivePeakPhotonFlux = getLogEffectivePeakPhotonFlux(0._RK,real(THRESH_ERFC_AVG,RK))
         difference = abs(logEffectivePeakPhotonFlux - logEffectivePeakPhotonFlux_ref)
         assertion = difference < tolerance
-        if (Test%isDebugMode .and. .not. assertion) then
+        !if (Test%isDebugMode .and. .not. assertion) then
             write(Test%outputUnit,"(*(g0))")
             write(Test%outputUnit,"(*(g0))") "logEffectivePeakPhotonFlux_ref    = ", logEffectivePeakPhotonFlux_ref
             write(Test%outputUnit,"(*(g0))") "logEffectivePeakPhotonFlux        = ", logEffectivePeakPhotonFlux
             write(Test%outputUnit,"(*(g0))") "difference                        = ", difference
             write(Test%outputUnit,"(*(g0))")
-        end if
+        !end if
     end function test_getLogEffectivePeakPhotonFlux_SPR_1
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -294,19 +294,19 @@ contains
         implicit none
         logical                 :: assertion
         real(RK), parameter     :: tolerance = sqrt(epsilon(1._RK))
-        real(RK), parameter     :: logEffectivePeakPhotonFlux_ref = real(THRESH_ERFC_AMP,RK)
+        real(RK), parameter     :: logEffectivePeakPhotonFlux_ref = -real(THRESH_ERFC_AMP,RK)
         real(RK)                :: logEffectivePeakPhotonFlux
         real(RK)                :: difference
         logEffectivePeakPhotonFlux = getLogEffectivePeakPhotonFlux(0._RK,real(THRESH_ERFC_AVG,RK))
         difference = abs(logEffectivePeakPhotonFlux - logEffectivePeakPhotonFlux_ref)
         assertion = difference < tolerance
-        if (Test%isDebugMode .and. .not. assertion) then
+        !if (Test%isDebugMode .and. .not. assertion) then
             write(Test%outputUnit,"(*(g0))")
             write(Test%outputUnit,"(*(g0))") "logEffectivePeakPhotonFlux_ref    = ", logEffectivePeakPhotonFlux_ref
             write(Test%outputUnit,"(*(g0))") "logEffectivePeakPhotonFlux        = ", logEffectivePeakPhotonFlux
             write(Test%outputUnit,"(*(g0))") "difference                        = ", difference
             write(Test%outputUnit,"(*(g0))")
-        end if
+        !end if
     end function test_getLogEffectivePeakPhotonFlux_DPR_1
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
