### 2020-11-29 ab9db01
minor enhancements
