### 2020-11-29 fdf8b8e
minor enhancements
