+++ src/kernel/System_mod.f90
@@ -195,7 +195,7 @@ contains
     !>
     !> \param[out]  OS                  :   An object of class [OS_type](@ref os_type).
     !> \param[in]   shellQueryEnabled   :   A logical variable indicating if the type and name of the current
-    !>                                      runtime shell should be queried or not (optional, default = `.false.`).
+    !>                                      runtime shell should be queried or not (optional, default = `.true.`).
     subroutine queryOS(OS, shellQueryEnabled)
 #if defined DLL_ENABLED && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: queryOS
@@ -213,7 +213,7 @@ contains
         character(:)    , allocatable           :: osname
 #endif
 
-        shellQueryEnabledDefault = .false.
+        shellQueryEnabledDefault = .true.
         if (present(shellQueryEnabled)) shellQueryEnabledDefault = shellQueryEnabled
         OS%Err%occurred = .false.
         OS%Err%msg = ""
@@ -231,10 +231,12 @@ contains
             else
                 mv_shCacheEnabled = .true.
                 call OS%Shell%query(OS%isWindows)
+                ! LCOV_EXCL_START
                 if (OS%Shell%Err%occurred) then
                     OS%Err = OS%Shell%Err
                     return
                 end if
+                ! LCOV_EXCL_STOP
                 mv_OS%Shell = OS%Shell
             end if
 
@@ -264,11 +266,13 @@ contains
 
         if (allocated(OS%name)) deallocate(OS%name); allocate( character(MAX_OS_NAME_LEN) :: OS%name )
         call getEnvVar( name="OS", value=OS%name, Err=OS%Err )
+        ! LCOV_EXCL_START
         if (OS%Err%occurred) then
             OS%Err%msg = PROCEDURE_NAME // ": Error occurred while querying OS type." // NLC // OS%Err%msg
             OS%name = ""
             return
         end if
+        ! LCOV_EXCL_STOP
 
         OS%name = trim(adjustl(OS%name))
 
@@ -292,11 +296,13 @@ contains
 
             if (allocated(OS%name)) deallocate(OS%name); allocate( character(MAX_OS_NAME_LEN) :: OS%name )
             call getEnvVar( name="OSTYPE", value=OS%name, Err=OS%Err )
+            ! LCOV_EXCL_START
             if (OS%Err%occurred) then
                 OS%Err%msg = PROCEDURE_NAME // ": Error occurred while querying OS type." // NLC // OS%Err%msg
                 OS%name = ""
                 return
             end if
+            ! LCOV_EXCL_STOP
 
             OS%name = trim(adjustl(OS%name))
             osname = getLowerCase(OS%name)
@@ -322,30 +328,37 @@ contains
                     integer                     :: fileUnit
                     type(RandomFileName_type)   :: RFN
                     RFN = RandomFileName_type(key="queryOS")
+                    ! LCOV_EXCL_START
                     if (RFN%Err%occurred) then
                         OS%Err = RFN%Err
                         OS%Err%msg = PROCEDURE_NAME // ": Error occurred while inquiring OS type." // NLC // OS%Err%msg
                         OS%name = ""
                         return
                     end if
+                    ! LCOV_EXCL_STOP
 
                     call executeCmd( command="uname > "//RFN%path, Err=OS%Err )
+                    ! LCOV_EXCL_START
                     if (OS%Err%occurred) then
                         OS%Err%msg = PROCEDURE_NAME // ": Error occurred while executing command 'uname > "// RFN%path // "'." // NLC // OS%Err%msg
                         OS%name = ""
                         return
                     end if
[…]
