### 2020-11-29 c062180
minor enhancements
