+++ buildParaMonte.sh
@@ -1985,9 +1985,9 @@ echo >&2 "-- ${BUILD_NAME} - all generated build files will be stored at: ${Para
 
 # set object/module/lib files directories
 
-ParaMonte_OBJ_DIR=${ParaMonte_BLD_DIR}/obj; export ParaMonte_OBJ_DIR
-ParaMonte_MOD_DIR=${ParaMonte_BLD_DIR}/mod; export ParaMonte_MOD_DIR
-ParaMonte_LIB_DIR=${ParaMonte_BLD_DIR}/lib; export ParaMonte_LIB_DIR
+ParaMonte_OBJ_DIR="${ParaMonte_BLD_DIR}/obj"; export ParaMonte_OBJ_DIR
+ParaMonte_MOD_DIR="${ParaMonte_BLD_DIR}/mod"; export ParaMonte_MOD_DIR
+ParaMonte_LIB_DIR="${ParaMonte_BLD_DIR}/lib"; export ParaMonte_LIB_DIR
 
 echo >&2 "-- ${BUILD_NAME} - ParaMonte object files directory: ${ParaMonte_OBJ_DIR}"
 echo >&2 "-- ${BUILD_NAME} - ParaMonte module files directory: ${ParaMonte_MOD_DIR}"
@@ -1995,7 +1995,7 @@ echo >&2 "-- ${BUILD_NAME} - ParaMonte library files directory: ${ParaMonte_LIB_
 
 # make bin directory
 
-ParaMonte_BIN_DIR=${ParaMonte_ROOT_DIR}/bin
+ParaMonte_BIN_DIR="${ParaMonte_ROOT_DIR}/bin"
 echo >&2 "-- ${BUILD_NAME} - ParaMonte binaries directory: ${ParaMonte_BIN_DIR}"
 if [[ -d "${ParaMonte_BIN_DIR}" ]]; then
     echo >&2 "-- ${BUILD_NAME} - ParaMonte binaries directory already exists. skipping..."
@@ -2009,14 +2009,14 @@ export ParaMonte_BIN_DIR
 # set ParaMonte library source directories
 ####################################################################################################################################
 
-         ParaMonteExample_SRC_DIR=${ParaMonte_ROOT_DIR}/example
-       ParaMonteInterface_SRC_DIR=${ParaMonte_ROOT_DIR}/src/interface
-      ParaMonteInterfaceC_SRC_DIR=${ParaMonteInterface_SRC_DIR}/C
- ParaMonteInterfaceMATLAB_SRC_DIR=${ParaMonteInterface_SRC_DIR}/MATLAB
- ParaMonteInterfacePython_SRC_DIR=${ParaMonteInterface_SRC_DIR}/Python
-ParaMonteInterfaceFortran_SRC_DIR=${ParaMonteInterface_SRC_DIR}/Fortran
-      ParaMonteMATLABTest_SRC_DIR=${ParaMonteInterfaceMATLAB_SRC_DIR}/test
-      ParaMontePythonTest_SRC_DIR=${ParaMonteInterfacePython_SRC_DIR}/test
+         ParaMonteExample_SRC_DIR="${ParaMonte_ROOT_DIR}/example"
+       ParaMonteInterface_SRC_DIR="${ParaMonte_ROOT_DIR}/src/interface"
+      ParaMonteInterfaceC_SRC_DIR="${ParaMonteInterface_SRC_DIR}/C"
+ ParaMonteInterfaceMATLAB_SRC_DIR="${ParaMonteInterface_SRC_DIR}/MATLAB"
+ ParaMonteInterfacePython_SRC_DIR="${ParaMonteInterface_SRC_DIR}/Python"
[…]
