### 2020-11-29 7c2a6b1
minor enhancements
