+++ src/kernel/tests/Test_Misc_mod.f90
@@ -106,6 +106,7 @@ contains
         Vector2 = Vector2_ref
         call swap_IK(Vector1,Vector2)
         assertion = all(Vector1==Vector2_ref) .and. all(Vector2==Vector1_ref)
+        ! LCOV_EXCL_START
         if (Test%isDebugMode .and. .not. assertion) then
             write(Test%outputUnit,"(*(g0,:,' '))")
             write(Test%outputUnit,"(*(g0,:,' '))") "Vector1_ref =", Vector1_ref
@@ -114,6 +115,7 @@ contains
             write(Test%outputUnit,"(*(g0,:,' '))") "Vector1     =", Vector1
             write(Test%outputUnit,"(*(g0,:,' '))")
         end if
+        ! LCOV_EXCL_STOP
     end function test_swap_IK_1
 
     !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -132,6 +134,7 @@ contains
         Vector2 = Vector2_ref
         call swap_SPI(Vector1,Vector2)
         assertion = all(Vector1==Vector2_ref) .and. all(Vector2==Vector1_ref)
+        ! LCOV_EXCL_START
         if (Test%isDebugMode .and. .not. assertion) then
             write(Test%outputUnit,"(*(g0,:,' '))")
             write(Test%outputUnit,"(*(g0,:,' '))") "Vector1_ref =", Vector1_ref
@@ -140,6 +143,7 @@ contains
             write(Test%outputUnit,"(*(g0,:,' '))") "Vector1     =", Vector1
             write(Test%outputUnit,"(*(g0,:,' '))")
         end if
+        ! LCOV_EXCL_STOP
     end function test_swap_SPI_1
 
     !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -158,6 +162,7 @@ contains
         Vector2 = Vector2_ref
         call swap_DPI(Vector1,Vector2)
         assertion = all(Vector1==Vector2_ref) .and. all(Vector2==Vector1_ref)
+        ! LCOV_EXCL_START
         if (Test%isDebugMode .and. .not. assertion) then
             write(Test%outputUnit,"(*(g0,:,' '))")
             write(Test%outputUnit,"(*(g0,:,' '))") "Vector1_ref =", Vector1_ref
@@ -166,6 +171,7 @@ contains
             write(Test%outputUnit,"(*(g0,:,' '))") "Vector1     =", Vector1
             write(Test%outputUnit,"(*(g0,:,' '))")
         end if
+        ! LCOV_EXCL_STOP
     end function test_swap_DPI_1
 
     !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -184,6 +190,7 @@ contains
         Vector2 = Vector2_ref
         call swap_RK(Vector1,Vector2)
         assertion = all(Vector1==Vector2_ref) .and. all(Vector2==Vector1_ref)
+        ! LCOV_EXCL_START
[…]
