+++ src/kernel/Path_mod.f90
@@ -130,7 +130,7 @@ contains
     !> \return
     !> `Path` : An object of class [Path_type](@ref path_type) containing the path properties and methods.
     function constructPath(inputPath,OS) result(Path)
-#if defined DLL_ENABLED && !defined CFI_ENABLED
+#if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: constructPath
 #endif
         use System_mod, only: OS_type
@@ -154,7 +154,7 @@ contains
     !> \warning
     !> On output, do not forget to check the value `Path%%Err%%occurred` before using the output `Path`.
     subroutine queryPath(Path,inputPath,OS)
-#if defined DLL_ENABLED && !defined CFI_ENABLED
+#if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: queryPath
 #endif
         use Err_mod, only: Err_type
@@ -244,7 +244,7 @@ contains
     !> This routine strictly assumes that there is no dangling `\` in the input Linux path, and if there is,
     !> then either it is used to escape the special shell characters, or otherwise, the path is a Windows path.
     pure subroutine winifyPath(inputPath,outputPath,Err)!,ignoreWindowsReservedChars)
-#if defined DLL_ENABLED && !defined CFI_ENABLED
+#if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: winifyPath
 #endif
         use Err_mod, only: Err_type
@@ -341,7 +341,7 @@ contains
     !> \param[in]       inputPath   :   The input path. If provided, it will overwrite `Path%original`.
     !> \param[out]      outputPath  :   The output modified path which conforms to the rules of the Unix OS.
     pure subroutine linifyPath(inputPath,outputPath)
-#if defined DLL_ENABLED && !defined CFI_ENABLED
+#if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: linifyPath
 #endif
         use Constants_mod, only: IK
@@ -391,7 +391,7 @@ contains
     !> \param[out]  slashOS :   The slash separator character of length 1, used by the OS to separate segments of the paths.
     !> \param[out]  Err     :   An object of class [Err_type](@ref err_mod::err_type) containing error handling tools.
     subroutine getSlashOS(slashOS,Err)
-#if defined DLL_ENABLED && !defined CFI_ENABLED
+#if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: getSlashOS
 #endif
         use Err_mod, only: Err_type
@@ -433,7 +433,7 @@ contains
     !> \param[out]      outputPath  :   The output modified path which conforms to the rules of the current OS.
     !> \param[out]      Err         :   An object of class [Err_type](@ref err_mod::err_type) containing error handling tools.
     subroutine modifyPath(inputPath,outputPath,Err)
[…]
