### 2020-11-30 8054c19
minor enhancements
