### 2020-11-30 3fc0023
minor enhancements
