### 2020-11-30 2f7a2bb
minor enhancements
