### 2020-11-30 d2c1100
minor enhancements
