### 2020-11-30 949eed3
minor enhancements
