### 2020-11-30 e5cbb74
minor enhancements
