+++ src/kernel/System_mod.f90
@@ -98,6 +98,7 @@ module System_mod
         logical                         :: isBash       = .false.   !< The logical value indicating whether the shell is Unix Bash.
         logical                         :: isPowerShell = .false.   !< The logical value indicating whether the shell is Windows PowerShell.
         logical                         :: isUnix       = .false.   !< The logical value indicating whether the shell is Unix-like.
+        character(:), allocatable       :: slash                    !< The path separator character in the current shell (Windows Shell: "\", Unix-like: "/").
         character(:), allocatable       :: name                     !< The name of or path to the current shell.
         type(Err_type)                  :: Err                      !< An object of class [Err_type](@ref err_mod::err_type) indicating
                                                                     !! whether error has occurred during the query.
@@ -160,9 +161,9 @@ module System_mod
 
     ! cache the OS query result to speed up code
 
-    logical                             :: mv_osCacheEnabled = .false.
-    logical                             :: mv_shCacheEnabled = .false.
-    type(OS_type)                       :: mv_OS
+    logical      , private              :: mv_osCacheActivated = .false.
+    logical      , private              :: mv_shCacheActivated = .false.
+    type(OS_type), private              :: mv_OS
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -218,7 +219,7 @@ contains
         OS%Err%occurred = .false.
         OS%Err%msg = ""
 
-        if (mv_osCacheEnabled) then
+        if (mv_osCacheActivated) then
 
             OS%name         = mv_OS%name
             OS%slash        = mv_OS%slash
@@ -226,13 +227,13 @@ contains
             OS%isDarwin     = mv_OS%isDarwin
             OS%isLinux      = mv_OS%isLinux
 
-            if (mv_shCacheEnabled) then
+            if (mv_shCacheActivated) then
                 OS%Shell    = mv_OS%Shell
             else
-                mv_shCacheEnabled = .true.
+                mv_shCacheActivated = .true.
                 call OS%Shell%query(OS%isWindows)
-                ! LCOV_EXCL_START
                 if (OS%Shell%Err%occurred) then
+                ! LCOV_EXCL_START
                     OS%Err = OS%Shell%Err
                     return
                 end if
@@ -266,8 +267,8 @@ contains
 
         if (allocated(OS%name)) deallocate(OS%name); allocate( character(MAX_OS_NAME_LEN) :: OS%name )
         call getEnvVar( name="OS", value=OS%name, Err=OS%Err )
-        ! LCOV_EXCL_START
         if (OS%Err%occurred) then
+        ! LCOV_EXCL_START
             OS%Err%msg = PROCEDURE_NAME // ": Error occurred while querying OS type." // NLC // OS%Err%msg
             OS%name = ""
             return
@@ -296,8 +297,8 @@ contains
 
             if (allocated(OS%name)) deallocate(OS%name); allocate( character(MAX_OS_NAME_LEN) :: OS%name )
             call getEnvVar( name="OSTYPE", value=OS%name, Err=OS%Err )
-            ! LCOV_EXCL_START
             if (OS%Err%occurred) then
+            ! LCOV_EXCL_START
                 OS%Err%msg = PROCEDURE_NAME // ": Error occurred while querying OS type." // NLC // OS%Err%msg
                 OS%name = ""
                 return
@@ -328,8 +329,8 @@ contains
[…]
