+++ src/kernel/tests/Test_Math_mod.f90
@@ -74,22 +74,28 @@ contains
         call Test%run(test_getLowerGamma_2, "test_getLowerGamma_2")
         call Test%run(test_getLowerGamma_3, "test_getLowerGamma_3")
         call Test%run(test_getLowerGamma_4, "test_getLowerGamma_4")
+        call Test%run(test_getLowerGamma_5, "test_getLowerGamma_5")
         call Test%run(test_getUpperGamma_1, "test_getUpperGamma_1")
         call Test%run(test_getUpperGamma_2, "test_getUpperGamma_2")
         call Test%run(test_getUpperGamma_3, "test_getUpperGamma_3")
         call Test%run(test_getUpperGamma_4, "test_getUpperGamma_4")
+        call Test%run(test_getUpperGamma_5, "test_getUpperGamma_5")
+        call Test%run(test_getGammaSeries_1, "test_getGammaSeries_1")
         call Test%run(test_getLogSubExp_RK_1, "test_getLogSubExp_RK_1")
         call Test%run(test_getLogSumExp_RK_1, "test_getLogSumExp_RK_1")
         call Test%run(test_getLogSumExp_RK_2, "test_getLogSumExp_RK_2")
         call Test%run(test_getLogSumExp_CK_1, "test_getLogSumExp_CK_1")
         call Test%run(test_getLogSumExp_CK_2, "test_getLogSumExp_CK_2")
         call Test%run(test_getLogFactorial_1, "test_getLogFactorial_1")
+        call Test%run(test_getGammaHalfInt_1, "test_getGammaHalfInt_1")
+        call Test%run(test_getGammaContFrac_1, "test_getGammaContFrac_1")
         call Test%run(test_getLogEllVolCoef_1, "test_getLogEllVolCoef_1")
         call Test%run(test_getLogEllVolCoef_2, "test_getLogEllVolCoef_2")
         call Test%run(test_getLogEggBoxSD_RK_1, "test_getLogEggBoxSD_RK_1")
         call Test%run(test_getLogEggBoxSD_CK_1, "test_getLogEggBoxSD_CK_1")
         call Test%run(test_getLogVolUnitBall_1, "test_getLogVolUnitBall_1")
         call Test%run(test_getLogVolUnitBall_2, "test_getLogVolUnitBall_2")
+        call Test%run(test_getLogGammaHalfInt_1, "test_getLogGammaHalfInt_1")
         call Test%run(test_getLogVolEllipsoid_1, "test_getLogVolEllipsoid_1")
         call Test%run(test_getLogVolEllipsoids_1, "test_getLogVolEllipsoids_1")
         call Test%run(test_getCumSumReverse_IK_1, "test_getCumSumReverse_IK_1")
@@ -219,8 +225,8 @@ contains
         if (allocated(Difference)) deallocate(Difference); allocate(Difference, mold = CumSum)
         Difference = abs(CumSum - CumSum_ref)
         assertion  = all(Difference < tolerance)
-        ! LCOV_EXCL_START
         if (Test%isDebugMode .and. .not. assertion) then
+        ! LCOV_EXCL_START
             write(Test%outputUnit,"(*(g0,:,' '))")
             write(Test%outputUnit,"(*(g0,:,' '))") "CumSum_ref    = ", CumSum_ref
             write(Test%outputUnit,"(*(g0,:,' '))") "CumSum        = ", CumSum
@@ -238,17 +244,18 @@ contains
         logical                     :: assertion
         integer(IK)                 :: i
[…]
