### 2020-12-02 ff5196f
minor enhancements
