### 2020-12-03 f4e21c7
minor enhancements
