### 2020-12-03 bc71d16
minor enhancements
