+++ src/kernel/tests/Test_ParaDRAM_mod@SpecBase_mod.inc.f90
@@ -57,6 +57,7 @@
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
                             , chainFileFormat = "nonsense" &
+                            , outputFileName = "test_SpecBase_ChainFileFormat_type_1" &
                             )
         assertion = assertion .and. PD%Err%occurred
 #endif
@@ -76,6 +77,7 @@
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
                             , inputFile = "&ParaDRAM chainFileFormat = 'nonsense' /" &
+                            , outputFileName = "test_SpecBase_ChainFileFormat_type_2" &
                             )
         assertion = assertion .and. PD%Err%occurred
 #endif
@@ -95,8 +97,9 @@
 #if defined CODECOV_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
-                            , domainLowerLimitVec = domainLowerLimitVec &
                             , mpiFinalizeRequested = .false. &
+                            , domainLowerLimitVec = domainLowerLimitVec &
+                            , outputFileName = "test_SpecBase_DomainLowerLimitVec_type_1" &
                             )
         assertion = PD%Err%occurred
 #endif
@@ -117,8 +120,9 @@
 #if defined CODECOV_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
-                            , inputFile = "&ParaDRAM domainLowerLimitVec = "//num2str(domainLowerLimitVec)//" /" &
                             , mpiFinalizeRequested = .false. &
+                            , inputFile = "&ParaDRAM domainLowerLimitVec = "//num2str(domainLowerLimitVec)//" /" &
+                            , outputFileName = "test_SpecBase_DomainLowerLimitVec_type_2" &
                             )
         assertion = PD%Err%occurred
 #endif
@@ -138,8 +142,9 @@
 #if defined CODECOV_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
-                            , domainUpperLimitVec = domainUpperLimitVec &
                             , mpiFinalizeRequested = .false. &
+                            , domainUpperLimitVec = domainUpperLimitVec &
+                            , outputFileName = "test_SpecBase_DomainUpperLimitVec_type_1" &
                             )
         assertion = PD%Err%occurred
 #endif
@@ -160,8 +165,9 @@
 #if defined CODECOV_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
-                            , inputFile = "&ParaDRAM domainUpperLimitVec = "//num2str(domainUpperLimitVec)//" /" &
                             , mpiFinalizeRequested = .false. &
+                            , inputFile = "&ParaDRAM domainUpperLimitVec = "//num2str(domainUpperLimitVec)//" /" &
+                            , outputFileName = "test_SpecBase_DomainUpperLimitVec_type_2" &
                             )
         assertion = PD%Err%occurred
 #endif
@@ -183,9 +189,10 @@
 #if defined CODECOV_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
                             , domainLowerLimitVec = domainLowerLimitVec &
                             , domainUpperLimitVec = domainUpperLimitVec &
-                            , mpiFinalizeRequested = .false. &
+                            , outputFileName = "test_SpecBase_DomainUpperLimitVec_type_3" &
                             )
[…]
