+++ src/kernel/SpecBase_ChainFileFormat_mod.f90
@@ -145,15 +145,17 @@ contains
 #endif
         use String_mod, only: getLowerCase
         implicit none
-        class(ChainFileFormat_type), intent(inout)    :: ChainFileFormatObj
-        character(*), intent(in)                      :: chainFileFormat
+        class(ChainFileFormat_type), intent(inout)  :: ChainFileFormatObj
+        character(*), intent(in)                    :: chainFileFormat
+        character(:), allocatable                   :: lowerCaseVal
         ChainFileFormatObj%val = trim(adjustl(chainFileFormat))
         if ( ChainFileFormatObj%val==trim(adjustl(ChainFileFormatObj%null)) ) then
             ChainFileFormatObj%val = trim(adjustl(ChainFileFormatObj%def))
         end if
-        if (getLowerCase(ChainFileFormatObj%val)==getLowerCase(ChainFileFormatObj%compact)) ChainFileFormatObj%iscompact = .true.
-        if (getLowerCase(ChainFileFormatObj%val)==getLowerCase(ChainFileFormatObj%verbose)) ChainFileFormatObj%isverbose = .true.
-        if (getLowerCase(ChainFileFormatObj%val)==getLowerCase(ChainFileFormatObj%binary)) ChainFileFormatObj%isBinary = .true.
+        lowerCaseVal = getLowerCase(ChainFileFormatObj%val)
+        ChainFileFormatObj%iscompact = lowerCaseVal == getLowerCase(ChainFileFormatObj%compact)
+        ChainFileFormatObj%isverbose = lowerCaseVal == getLowerCase(ChainFileFormatObj%verbose)
+        ChainFileFormatObj%isBinary  = lowerCaseVal == getLowerCase(ChainFileFormatObj%binary)
     end subroutine setChainFileFormat
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -165,7 +167,7 @@ contains
         use Err_mod, only: Err_type
         use String_mod, only: num2str
         implicit none
-        class(ChainFileFormat_type), intent(in)   :: ChainFileFormat
+        class(ChainFileFormat_type), intent(in)     :: ChainFileFormat
         character(*), intent(in)                    :: methodName
         type(Err_type), intent(inout)               :: Err
         character(*), parameter                     :: PROCEDURE_NAME = "@checkForSanity()"
