### 2020-12-03 e80f4f9
minor enhancements
