+++ src/kernel/tests/Test_ParaDRAM_mod@SpecBase_mod.inc.f90
@@ -101,7 +101,7 @@
                             , domainLowerLimitVec = domainLowerLimitVec &
                             , outputFileName = "test_SpecBase_DomainLowerLimitVec_type_1" &
                             )
-        assertion = PD%Err%occurred
+        assertion = assertion .and. PD%Err%occurred
 #endif
     end function test_SpecBase_DomainLowerLimitVec_type_1
 
@@ -124,7 +124,7 @@
                             , inputFile = "&ParaDRAM domainLowerLimitVec = "//num2str(domainLowerLimitVec)//" /" &
                             , outputFileName = "test_SpecBase_DomainLowerLimitVec_type_2" &
                             )
-        assertion = PD%Err%occurred
+        assertion = assertion .and. PD%Err%occurred
 #endif
     end function test_SpecBase_DomainLowerLimitVec_type_2
 
@@ -146,7 +146,7 @@
                             , domainUpperLimitVec = domainUpperLimitVec &
                             , outputFileName = "test_SpecBase_DomainUpperLimitVec_type_1" &
                             )
-        assertion = PD%Err%occurred
+        assertion = assertion .and. PD%Err%occurred
 #endif
     end function test_SpecBase_DomainUpperLimitVec_type_1
 
@@ -169,7 +169,7 @@
                             , inputFile = "&ParaDRAM domainUpperLimitVec = "//num2str(domainUpperLimitVec)//" /" &
                             , outputFileName = "test_SpecBase_DomainUpperLimitVec_type_2" &
                             )
-        assertion = PD%Err%occurred
+        assertion = assertion .and. PD%Err%occurred
 #endif
     end function test_SpecBase_DomainUpperLimitVec_type_2
 
@@ -179,11 +179,11 @@
     !> Test the ParaDRAM sampler with an input value for `domainUpperLimitVec` 
     !> that is smaller than the input value for `domainLowerLimitVec`.
     function test_SpecBase_DomainUpperLimitVec_type_3() result(assertion)
-        use Constants_mod, only: RK, HUGE_RK
+        use Constants_mod, only: RK
         implicit none
         logical             :: assertion
-        real(RK), parameter :: domainLowerLimitVec(*) = [+2._RK] ! NOTE: HUGE_RK is the null value.
-        real(RK), parameter :: domainUpperLimitVec(*) = [-2._RK] ! NOTE: HUGE_RK is the null value.
+        real(RK), parameter :: domainLowerLimitVec(*) = [+2._RK]
+        real(RK), parameter :: domainUpperLimitVec(*) = [-2._RK]
         type(ParaDRAM_type) :: PD
         assertion = .true.
 #if defined CODECOV_ENABLED
@@ -194,7 +194,7 @@
                             , domainUpperLimitVec = domainUpperLimitVec &
                             , outputFileName = "test_SpecBase_DomainUpperLimitVec_type_3" &
                             )
-        assertion = PD%Err%occurred
+        assertion = assertion .and. PD%Err%occurred
 #endif
     end function test_SpecBase_DomainUpperLimitVec_type_3
 
@@ -204,11 +204,11 @@
     !> Test the ParaDRAM sampler with an input value for `domainUpperLimitVec` 
     !> that is equal to the input value for `domainLowerLimitVec`.
     function test_SpecBase_DomainUpperLimitVec_type_4() result(assertion)
[…]
