### 2020-12-03 c6bfa0a
minor enhancements
