+++ src/kernel/tests/Test_ParaDRAM_mod@SpecBase_mod.inc.f90
@@ -758,7 +758,7 @@
                             , outputFileName = "test_SpecBase_RandomSeed_type_1" &
                             , randomSeed = -12345_IK &
                             )
-        assertion = assertion .and. PD%Err%occurred
+        assertion = assertion .and. .not. PD%Err%occurred
 #endif
     end function test_SpecBase_RandomSeed_type_1
 
@@ -779,9 +779,178 @@
                             , outputFileName = "test_SpecBase_RandomSeed_type_2" &
                             , inputFile = "&ParaDRAM randomSeed = 12345 /" &
                             )
-        assertion = assertion .and. PD%Err%occurred
+        assertion = assertion .and. .not. PD%Err%occurred
 #endif
     end function test_SpecBase_RandomSeed_type_2
+
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+    !> \brief
+    !> Test whether the ParaDRAM sampler appropriately considers a binary restart file format.
+    function test_SpecBase_RestartFileFormat_type_1() result(assertion)
+        use Constants_mod, only: IK, RK
+        implicit none
+        logical             :: assertion
+        type(ParaDRAM_type) :: PD
+        assertion = .true.
+#if defined CODECOV_ENABLED
+        call PD%runSampler  ( ndim = 1_IK &
+                            , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
+                            , outputFileName = "test_SpecBase_RestartFileFormat_type_1" &
+                            , restartFileFormat = "BINARY" &
+                            )
+        assertion = assertion .and. .not. PD%Err%occurred
+#endif
+    end function test_SpecBase_RestartFileFormat_type_1
+
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+    !> \brief
+    !> Test whether the ParaDRAM sampler appropriately considers an ascii restart file format.
+    function test_SpecBase_RestartFileFormat_type_2() result(assertion)
+        use Constants_mod, only: IK, RK
+        implicit none
+        logical             :: assertion
+        type(ParaDRAM_type) :: PD
+        assertion = .true.
+#if defined CODECOV_ENABLED
+        call PD%runSampler  ( ndim = 1_IK &
+                            , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
+                            , outputFileName = "test_SpecBase_RestartFileFormat_type_2" &
+                            , inputFile = "&ParaDRAM restartFileFormat = 'asCII' /" &
+                            )
+        assertion = assertion .and. .not. PD%Err%occurred
+#endif
+    end function test_SpecBase_RestartFileFormat_type_2
+
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+    !> \brief
[…]
