### 2020-12-04 eb01641
minor enhancements
