+++ src/kernel/SpecBase_TargetAcceptanceRate_mod.f90
@@ -113,7 +113,7 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-    subroutine setTargetAcceptanceRate(TargetAcceptanceRateObj,TargetAcceptanceRate)
+    subroutine setTargetAcceptanceRate(TargetAcceptanceRateObj,targetAcceptanceRate)
 #if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: setTargetAcceptanceRate
 #endif
@@ -122,19 +122,13 @@ contains
         class(TargetAcceptanceRate_type), intent(inout)     :: TargetAcceptanceRateObj
         real(RK), intent(in)                                :: TargetAcceptanceRate(2)
         logical                                             :: lowerLimitSet, upperLimitSet
-        TargetAcceptanceRateObj%Val = TargetAcceptanceRate
-        lowerLimitSet = TargetAcceptanceRateObj%Val(1)/=TargetAcceptanceRateObj%null
-        upperLimitSet = TargetAcceptanceRateObj%Val(2)/=TargetAcceptanceRateObj%null
-        if      ( lowerLimitSet .and. (.not. upperLimitSet) ) then
-            TargetAcceptanceRateObj%Val(2) = TargetAcceptanceRateObj%Val(1)
-        elseif  ( upperLimitSet .and. (.not. lowerLimitSet) ) then
-            TargetAcceptanceRateObj%Val(1) = TargetAcceptanceRateObj%Val(2)
-        elseif  ( .not. (lowerLimitSet .or.  upperLimitSet) ) then
-            TargetAcceptanceRateObj%Val = TargetAcceptanceRateObj%Def
-            TargetAcceptanceRateObj%scalingRequested = .false.
-        elseif ( all(TargetAcceptanceRateObj%Val==TargetAcceptanceRateObj%Def) ) then
-            TargetAcceptanceRateObj%scalingRequested = .false.
-        end if
+        TargetAcceptanceRateObj%Val = targetAcceptanceRate
+        lowerLimitSet = TargetAcceptanceRateObj%Val(1) /= TargetAcceptanceRateObj%null
+        upperLimitSet = TargetAcceptanceRateObj%Val(2) /= TargetAcceptanceRateObj%null
+        if (lowerLimitSet .and. .not. upperLimitSet) TargetAcceptanceRateObj%Val(2) = TargetAcceptanceRateObj%Val(1)
+        if (upperLimitSet .and. .not. lowerLimitSet) TargetAcceptanceRateObj%Val(1) = TargetAcceptanceRateObj%Val(2)
+        if (.not.(lowerLimitSet .or. upperLimitSet)) TargetAcceptanceRateObj%Val(:) = TargetAcceptanceRateObj%Def
+        TargetAcceptanceRateObj%scalingRequested = any(TargetAcceptanceRateObj%Val /= TargetAcceptanceRateObj%Def)
     end subroutine setTargetAcceptanceRate
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
[…]

[…]
