### 2020-12-04 daf0af7
minor enhancements
