+++ src/kernel/tests/Test_ParaDRAM_mod@SpecMCMC_mod.inc.f90
@@ -56,7 +56,7 @@
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
-                            , outputFileName = MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ChainSize_type_1" &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ChainSize_type_1" &
                             , chainSize = 0_IK &
                             )
         assertion = assertion .and. PD%Err%occurred
@@ -76,7 +76,7 @@
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
-                            , outputFileName = MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ChainSize_type_2" &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ChainSize_type_2" &
                             , inputFile = "&ParaDRAM chainSize = 1 /" &
                             )
         assertion = assertion .and. PD%Err%occurred
@@ -96,7 +96,7 @@
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
-                            , outputFileName = MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ProposalModel_type_1" &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ProposalModel_type_1" &
                             , proposalModel = "UniForm" &
                             )
         assertion = assertion .and. .not. PD%Err%occurred
@@ -116,7 +116,7 @@
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
-                            , outputFileName = MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ProposalModel_type_2" &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ProposalModel_type_2" &
                             , inputFile = "&ParaDRAM proposalModel = 'UNIFORM' /" &
                             )
         assertion = assertion .and. .not. PD%Err%occurred
@@ -136,7 +136,7 @@
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
-                            , outputFileName = MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ProposalModel_type_3" &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecMCMC/test_SpecMCMC_ProposalModel_type_3" &
                             , proposalModel = "Normal" &
                             )
         assertion = assertion .and. .not. PD%Err%occurred
@@ -156,7 +156,7 @@
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
[…]

[…]
