### 2020-12-04 9e25e3f
minor enhancements
