+++ src/kernel/SpecMCMC_mod.f90
@@ -154,18 +154,18 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-    subroutine setFromInputFile( SpecMCMC, Err, nd, domainLowerLimitVec, domainUpperLimitVec )
+    subroutine setFromInputFile(SpecMCMC, SpecBase, Err) ! domainLowerLimitVec, domainUpperLimitVec )
 #if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: setFromInputFile
 #endif
-
+        use SpecBase_mod, only: SpecBase_type
         use Constants_mod, only: IK, RK
         use Err_mod, only: Err_type
+
         implicit none
+
         class(SpecMCMC_type), intent(inout)     :: SpecMCMC
-        integer(IK) , intent(in)                :: nd
-        real(RK)    , intent(in)                :: domainLowerLimitVec(nd)
-        real(RK)    , intent(in)                :: domainUpperLimitVec(nd)
+        type(SpecBase_type), intent(in)         :: SpecBase
         type(Err_type), intent(out)             :: Err
 
         Err%occurred = .false.
@@ -180,25 +180,16 @@ contains
         call SpecMCMC%SampleRefinementCount                 %set(sampleRefinementCount)
         call SpecMCMC%SampleRefinementMethod                %set(sampleRefinementMethod)
         call SpecMCMC%RandomStartPointRequested             %set(randomStartPointRequested)
-        call SpecMCMC%RandomStartPointDomainLowerLimitVec   %set(randomStartPointDomainLowerLimitVec, domainLowerLimitVec)
-        call SpecMCMC%RandomStartPointDomainUpperLimitVec   %set(randomStartPointDomainUpperLimitVec, domainUpperLimitVec)
-        call SpecMCMC%startPointVec                         %set(startPointVec &
-                                                                ,randomStartPointDomainLowerLimitVec    = SpecMCMC%RandomStartPointDomainLowerLimitVec%val  &
-                                                                ,randomStartPointDomainUpperLimitVec    = SpecMCMC%RandomStartPointDomainUpperLimitVec%val  &
-                                                                ,randomStartPointRequested              = SpecMCMC%RandomStartPointRequested%val            &
-                                                                ,domainLowerLimitVec                    = domainLowerLimitVec                               &
-                                                                ,domainUpperLimitVec                    = domainUpperLimitVec                               )
-
-
-        deallocate(randomStartPointDomainLowerLimitVec)
-        deallocate(randomStartPointDomainUpperLimitVec)
-        deallocate(startPointVec)
+        call SpecMCMC%RandomStartPointDomainLowerLimitVec   %set(SpecBase, randomStartPointDomainLowerLimitVec)
+        call SpecMCMC%RandomStartPointDomainUpperLimitVec   %set(SpecBase, randomStartPointDomainUpperLimitVec)
+        call SpecMCMC%startPointVec                         %set(startPointVec)
 
     end subroutine setFromInputFile
 
[…]
