+++ src/kernel/String_mod.f90
@@ -80,7 +80,7 @@ module String_mod
         type(CharVec_type), allocatable   :: Parts(:)       !< The string parts.
         integer(IK)                       :: nPart = 0_IK   !< The number of parts in the string.
     contains
-        procedure, nopass :: splitStr, replaceStr, getLowerCase, getUpperCase, isInteger, isDigit
+        procedure, nopass :: split, replaceStr, getLowerCase, getUpperCase, isInteger, isDigit
         procedure, nopass :: str2int, str2real, str2int32, str2int64, str2real32, str2real64
         procedure, nopass :: pad => padString
     end type String_type
@@ -156,6 +156,97 @@ contains
     !> \brief
     !> Split the input string string with the input `substitute` in the input `string` and return the result.
     !>
+    !> \param[in]       string  :   The input string.
+    !> \param[in]       delim   :   The delimiter to be used to split the input string.
+    !> \param[out]      npart   :   The number of substrings resulting from splitting the string (**optional**).
+    !>
+    !> \return
+    !> `Parts` : An allocatable array of type [CharVec_type](@ref jaggedarray_mod::charvec_type)
+    !> containing the split parts of the input string.
+    !>
+    !> \remark
+    !> When `delim = ""`, this routine returns a jagged array of strings, each element of which is one
+    !> character of the input string.
+    !>
+    !> \remark
+    !> This procedure is a static method of the class [String_type](@ref string_type).
+    !>
+    !> \remark
+    !> This procedure should be preferred over the the legacy implementation [splitStr()](@ref splitstr)
+    !> that is kept only for legacy support. This routine is significantly faster than the legacy implementation
+    !> and its semantic behavior is identical to the Python3's `split()` string method, except when `delim` is empty.
+    !>
+    !> \author
+    ! Amir Shahmoradi, Friday Dec 4, 2020, 11:40 PM, Dallas, TX
+    function split(string,delim,npart) result(Parts)
+#if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
+        !DEC$ ATTRIBUTES DLLEXPORT :: split
+#endif
+        implicit none
+        character(len=*)    , intent(in)            :: string, delim
+        integer(IK)         , intent(out), optional :: npart
+        character(len=:)    , allocatable           :: dummyStr
+        type(CharVec_type)  , allocatable           :: Parts(:)
+        integer(IK)         , allocatable           :: PartEnd(:)
+        integer(IK)         , allocatable           :: PartBegin(:)
+        integer(IK)                                 :: dlmlenMinusOne
+        integer(IK)                                 :: strlen, dlmlen, npartMax, ipart, ibeg, iend, i
+        logical                                     :: npartIsPresent
+
+        dlmlen = len(delim)
+        strlen = len(string)
+        npartIsPresent = present(npart)
+
+        ! if dlm is empty, return the whole string split character by character
+
+        if (dlmlen==0_IK) then
+            allocate(Parts(strlen))
+            do ipart = 1, strlen
+                Parts(ipart)%record = string(ipart:ipart)
+            end do
+            if (npartIsPresent) npart = strlen
+            return
+        end if
+
+        npartMax = 1_IK + strlen / dlmlen ! There can be at most strlen + 1 splits
+        allocate(PartBegin(npartMax), PartEnd(npartMax)) ! This will contain the beginning and the ends of the splits.
+        dlmlenMinusOne = dlmlen - 1_IK
+
+        ibeg = 0_IK
+        ipart = 1_IK
+        PartBegin(ipart) = 1_IK
+        loopParseString: do
+
+            ibeg = ibeg + 1_IK
[…]
