### 2020-12-05 beaf171
minor enhancements
