+++ src/kernel/tests/Test_File_mod.f90
@@ -224,15 +224,15 @@ contains
         assertion = assertion .and. File%Err%occurred   .eqv. .false.
         assertion = assertion .and. File%Err%msg        == ""
 
-        ! LCOV_EXCL_START
         if (Test%isDebugMode .and. .not. assertion) then
+        ! LCOV_EXCL_START
             write(Test%outputUnit,"(*(g0))")
             write(Test%outputUnit,"(*(g0))")   "File%Path%original  : ", File%Path%original
             write(Test%outputUnit,"(*(g0))")   "File%Path%modified  : ", File%Path%modified
             write(Test%outputUnit,"(*(g0))")   "File%Path%dir       : ", File%Path%dir
             write(Test%outputUnit,"(*(g0))")   "File%Path%name      : ", File%Path%name
             write(Test%outputUnit,"(*(g0))")   "File%Path%ext       : ", File%Path%ext
-            write(Test%outputUnit,"(*(g0))")   "File%Path%slashOS   : ", File%Path%slashOS
+            write(Test%outputUnit,"(*(g0))")   "File%Path%shellSlash: ", File%Path%shellSlash
             write(Test%outputUnit,"(*(g0))")   "File%unit           : ", File%unit
             write(Test%outputUnit,"(*(g0))")   "File%number         : ", File%number
             write(Test%outputUnit,"(*(g0))")   "File%recl           : ", File%recl
@@ -290,18 +290,19 @@ contains
 
         assertion = assertion .and. File%Path%original  == "./test_File_mod/\test_File_mod\-"
 
-        if (File%Path%slashOS=="/") then
+        if (File%Path%shellSlash=="/") then
             assertion = assertion .and. File%Path%modified  == "./test_File_mod/\test_File_mod\-"
             assertion = assertion .and. File%Path%dir       == "./test_File_mod/"
             assertion = assertion .and. File%Path%name      == "\test_File_mod\-"
             assertion = assertion .and. File%Path%ext       == ""
+#if defined OS_IS_WINDOWS
         else
             assertion = assertion .and. File%Path%modified  == ".\test_File_mod\\test_File_mod\-"
             assertion = assertion .and. File%Path%dir       == ".\test_File_mod\\test_File_mod\"
             assertion = assertion .and. File%Path%name      == "-"
             assertion = assertion .and. File%Path%ext       == ""
+#endif
         end if
-
         assertion = assertion .and. File%unit           == 13
         assertion = assertion .and. File%number         == -2147483647
         assertion = assertion .and. File%recl           == 9999
@@ -325,15 +326,15 @@ contains
         assertion = assertion .and. File%Err%occurred   .eqv. .false.
         assertion = assertion .and. File%Err%msg        == ""
 
-        ! LCOV_EXCL_START
         if (Test%isDebugMode .and. .not. assertion) then
+        ! LCOV_EXCL_START
             write(Test%outputUnit,"(*(g0))")
             write(Test%outputUnit,"(*(g0))")   "File%Path%original  : ", File%Path%original
             write(Test%outputUnit,"(*(g0))")   "File%Path%modified  : ", File%Path%modified
[…]
