### 2020-12-06 a8f79e0
minor enhancements
