+++ src/kernel/ParaDXXXProposal_mod.inc.f90
@@ -635,6 +635,7 @@ contains
                 adaptationMeasureComputationNeeded = .true.
                 mv_sampleSizeOld_save = mv_sampleSizeOld_save + sampleSizeCurrent
 
+            ! LCOV_EXCL_START
             else blockPosDefCheck
 
                 adaptationMeasure = 0._RK
@@ -690,6 +691,7 @@ contains
                 end if
 
             end if blockPosDefCheck
+            ! LCOV_EXCL_STOP
 
             !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -773,6 +775,7 @@ contains
             end do
             call getLogSqrtDetPosDefMat(nd,CovMatUpperCurrent,logSqrtDetSum,singularityOccurred)
             if (singularityOccurred) then
+                ! LCOV_EXCL_START
                 write(mc_logFileUnit,"(A)")
                 write(mc_logFileUnit,"(A)") "Singular covariance matrix detected while computing the Adaptation measure:"
                 write(mc_logFileUnit,"(A)")
@@ -792,10 +795,13 @@ contains
                                 &Otherwise, restarting the simulation might resolve the error."
                 call abort( Err = ProposalErr, prefix = mc_methodBrand, newline = "\n", outputUnit = mc_logFileUnit )
                 return
+                ! LCOV_EXCL_STOP
             end if
+
             !adaptationMeasure = 1._RK - exp( 0.5_RK*(mv_logSqrtDetOld_save+logSqrtDetNew) - logSqrtDetSum )
             adaptationMeasure = sqrt( 1._RK - exp( 0.5*(mv_logSqrtDetOld_save + logSqrtDetNew) - logSqrtDetSum ) ) ! totalVariationUpperBound
             mv_logSqrtDetOld_save = logSqrtDetNew
+
 !block
 !integer, save :: counter = 0
 !counter = counter + 1
@@ -812,10 +818,12 @@ contains
 !end block
 
             if (adaptationMeasure<0._RK) then
+                ! LCOV_EXCL_START
                 call warn   ( prefix = mc_methodBrand &
                             , outputUnit = mc_logFileUnit &
                             , msg = mc_negativeTotalVariationMsg//num2str(adaptationMeasure) )
                 adaptationMeasure = 0._RK
+                ! LCOV_EXCL_STOP
             end if
 
             ! update the higher-stage delayed-rejection Cholesky Lower matrices
@@ -915,6 +923,7 @@ contains
     !> \warning
     !> When CAF parallelism is used, this routine must be exclusively called by the rooter images.
     !> When MPI parallelism is used, this routine must be called by all images.
+#if defined CAF_ENABLED || MPI_ENABLED
     subroutine bcastAdaptation()
 #if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: bcastAdaptation
@@ -939,6 +948,7 @@ contains
 #endif
         call getInvCovMat()
     end subroutine bcastAdaptation
+#endif
 
[…]
