### 2020-12-06 8c746c8
minor enhancements
