+++ src/kernel/ParaDISE_mod@Setup_smod.f90
@@ -52,5 +52,4 @@ submodule (ParaDISE_mod) Setup_smod
 #define PARADISE ParaDISE
 #include "ParaDXXX_mod@Setup_smod.inc.f90"
 #undef PARADISE
-end submodule Setup_smod
-
+end submodule Setup_smod ! LCOV_EXCL_LINE
