+++ src/kernel/tests/Test_ParaDRAM_mod@SpecBase_mod.inc.f90
@@ -975,47 +975,3 @@
     end function test_SpecBase_TargetAcceptanceRate_type_6
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-    !> \brief
-    !> Test the ParaDRAM sampler with a wrong internal input namelist group:
-    !> +    Infinity values for `domainLowerLimitVec` and `domainUpperLimitVec`.
-    function test_runSampler_7() result(assertion)
-        use Constants_mod, only: IK, RK
-        implicit none
-        logical             :: assertion
-        type(ParaDRAM_type) :: PD
-        assertion = .true.
-#if defined CODECOV_ENABLED
-        call PD%runSampler  ( ndim = 1_IK &
-                            , getLogFunc = getLogFuncMVN &
-                            , mpiFinalizeRequested = .false. &
-                            , inputFile = "&ParaDXXX randomSeed = 1111 /" &
-                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecBase/test_runSampler_7" &
-                            )
-        assertion = assertion .and. .not. PD%Err%occurred
-#endif
-    end function test_runSampler_7
-
-!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-    !> \brief
-    !> Test the ParaDRAM sampler with a wrong internal input namelist group:
-    !> +    Infinity values for `domainLowerLimitVec` and `domainUpperLimitVec`.
-    function test_runSampler_8() result(assertion)
-        use Constants_mod, only: IK, RK
-        implicit none
-        logical             :: assertion
-        type(ParaDRAM_type) :: PD
-        assertion = .true.
-#if defined CODECOV_ENABLED
-        call PD%runSampler  ( ndim = 1_IK &
-                            , getLogFunc = getLogFuncMVN &
-                            , mpiFinalizeRequested = .false. &
-                            , inputFile = Test%inDir//"/Test_ParaDRAM_mod@test_runSampler_8.in" &
-                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"@SpecBase/test_runSampler_8" &
-                            )
-        assertion = assertion .and. .not. PD%Err%occurred
-#endif
-    end function test_runSampler_8
-
-!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
