+++ src/kernel/ParaDISEProposalAbstract_mod.f90
@@ -52,5 +52,4 @@ module ParaDISEProposalAbstract_mod
 #define PARADISE ParaDISE
 #include "ParaDXXXProposalAbstract_mod.inc.f90"
 #undef ParaDISE
-end module ParaDISEProposalAbstract_mod
-
+end module ParaDISEProposalAbstract_mod ! LCOV_EXCL_LINE
