+++ src/kernel/ParaDISE_mod@Kernel_smod.f90
@@ -52,5 +52,4 @@ submodule (ParaDISE_mod) Kernel_smod
 #define PARADISE ParaDISE
 #include "ParaDXXX_mod@Kernel_smod.inc.f90"
 #undef PARADISE
-end submodule Kernel_smod
-
+end submodule Kernel_smod ! LCOV_EXCL_LINE

[…]
