### 2020-12-06 1989ff6
minor enhancements
