+++ src/kernel/Sort_mod.f90
@@ -41,7 +41,7 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 !>  \brief This module contains procedures for sorting arrays.
-!>  @author Amir Shahmoradi
+!>  \author Amir Shahmoradi
 
 module Sort_mod
 
