+++ src/kernel/DateTime_mod.f90
@@ -41,7 +41,7 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 !>  \brief This module contains classes and procedures for computing the date and time.
-!>  @author Amir Shahmoradi
+!>  \author Amir Shahmoradi
 
 module DateTime_mod
 
