+++ src/kernel/ParaDXXX_mod@Input_smod.inc.f90
@@ -46,7 +46,7 @@
 !> \remark
 !> This module requires preprocessing, prior to compilation.
 !>
-!> @author Amir Shahmoradi
+!> \author Amir Shahmoradi
 
 #if defined PARADRAM
 
