### 2020-12-06 b4a59bb
minor enhancements
