### 2020-12-06 7edec18
minor enhancements
