+++ src/kernel/ParaDISE_mod.f90
@@ -52,5 +52,4 @@ module ParaDISE_mod
 #define PARADISE ParaDISE
 #include "ParaDXXX_mod.inc.f90"
 #undef PARADISE
-end module ParaDISE_mod
-
+end module ParaDISE_mod ! LCOV_EXCL_LINE
