+++ src/kernel/ParaDXXX_mod@Setup_smod.inc.f90
@@ -242,9 +242,11 @@ contains
                                 , inputFile = inputFile &
                                 )
         if (self%Err%occurred) then
+            ! LCOV_EXCL_START
             self%Err%msg = PROCEDURE_NAME // self%Err%msg
             call self%abort( Err = self%Err, prefix = self%brand, newline = NLC, outputUnit = self%LogFile%unit )
             return
+            ! LCOV_EXCL_STOP
         end if
 
 
@@ -266,9 +268,11 @@ contains
 
         call self%getSpecFromInputFile(ndim)
         if (self%Err%occurred) then
+            ! LCOV_EXCL_START
             self%Err%msg = PROCEDURE_NAME // self%Err%msg
             call self%abort( Err = self%Err, prefix = self%brand, newline = "\n", outputUnit = self%LogFile%unit )
             return
+            ! LCOV_EXCL_STOP
         end if
 
         !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -277,53 +281,55 @@ contains
 
         call self%setWarnAboutProcArgHasPriority()
         if (self%procArgNeeded) then
-            call self%SpecBase%setFromInputArgs ( Err                                   = self%Err                              &
-                                                , sampleSize                            = sampleSize                            &
-                                                , randomSeed                            = randomSeed                            &
-                                                , description                           = description                           &
-                                                , outputFileName                        = outputFileName                        &
-                                                , outputDelimiter                       = outputDelimiter                       &
-                                                , chainFileFormat                       = chainFileFormat                       &
-                                                , variableNameList                      = variableNameList                      &
-                                                , restartFileFormat                     = restartFileFormat                     &
-                                                , outputColumnWidth                     = outputColumnWidth                     &
-                                                , overwriteRequested                    = overwriteRequested                    &
-                                                , outputRealPrecision                   = outputRealPrecision                   &
-                                                , silentModeRequested                   = silentModeRequested                   &
-                                                , domainLowerLimitVec                   = domainLowerLimitVec                   &
-                                                , domainUpperLimitVec                   = domainUpperLimitVec                   &
-                                                , parallelizationModel                  = parallelizationModel                  &
-                                                , progressReportPeriod                  = progressReportPeriod                  &
-                                                , targetAcceptanceRate                  = targetAcceptanceRate                  &
-                                                , mpiFinalizeRequested                  = mpiFinalizeRequested                  &
-                                                , maxNumDomainCheckToWarn               = maxNumDomainCheckToWarn               &
-                                                , maxNumDomainCheckToStop               = maxNumDomainCheckToStop               &
+            call self%SpecBase%setFromInputArgs ( Err                                   = self%Err                              & ! LCOV_EXCL_LINE
+                                                , sampleSize                            = sampleSize                            & ! LCOV_EXCL_LINE
+                                                , randomSeed                            = randomSeed                            & ! LCOV_EXCL_LINE
+                                                , description                           = description                           & ! LCOV_EXCL_LINE
+                                                , outputFileName                        = outputFileName                        & ! LCOV_EXCL_LINE
+                                                , outputDelimiter                       = outputDelimiter                       & ! LCOV_EXCL_LINE
[…]
