+++ src/kernel/ParaDISE_mod@Input_smod.f90
@@ -52,5 +52,4 @@ submodule (ParaDISE_mod) Input_smod
 #define PARADISE ParaDISE
 #include "ParaDXXX_mod@Input_smod.inc.f90"
 #undef PARADISE
-end submodule Input_smod
-
+end submodule Input_smod ! LCOV_EXCL_LINE

[…]
