### 2020-12-06 ceba893
minor enhancements
