+++ src/kernel/ParaDXXX_mod@Setup_smod.inc.f90
@@ -887,10 +887,12 @@ contains
                                             , comSecTime = self%Stats%avgCommTimePerFunCall & ! LCOV_EXCL_LINE
                                             )
                     if (ForkJoin%Err%occurred) then
+                        ! LCOV_EXCL_START
                         self%Err = ForkJoin%Err
                         self%Err%msg = PROCEDURE_NAME // self%Err%msg
                         call self%abort( Err = self%Err, prefix = self%brand, newline = NLC, outputUnit = self%LogFile%unit )
                         return
+                        ! LCOV_EXCL_STOP
                     end if
                 end if ! otherwise, use the previously-generated ForkJoin
 
@@ -898,7 +900,7 @@ contains
 
                 write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT)
                 write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT) "stats.parallelism.optimal.absolute.numProcess"
-                write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT)
+                write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT) ! LCOV_EXCL_LINE
                 if (isForkJoinParallelism) then
                     write(self%LogFile%unit,GENERIC_TABBED_FORMAT) ForkJoin%Speedup%Maximum%nproc
                 else
@@ -911,7 +913,7 @@ contains
 
                 write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT)
                 write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT) "stats.parallelism.optimal.absolute.speedup"
-                write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT)
+                write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT) ! LCOV_EXCL_LINE
                 if (isForkJoinParallelism) then
                     write(self%LogFile%unit,GENERIC_TABBED_FORMAT) ForkJoin%Speedup%Maximum%value
                 else
@@ -928,7 +930,7 @@ contains
 
                 write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT)
                 write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT) "stats.parallelism.optimal.absolute.scaling.strong.speedup"
-                write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT)
+                write(self%LogFile%unit,GENERIC_OUTPUT_FORMAT) ! LCOV_EXCL_LINE
                 if (isForkJoinParallelism) then
                     do imageCount = 1, ForkJoin%Speedup%count, 10
                         write(self%LogFile%unit,formatScaling) ForkJoin%Speedup%Scaling(imageCount:min(imageCount+9_IK,ForkJoin%Speedup%count))
@@ -1060,7 +1062,7 @@ contains
                                             , sumWeight = self%Stats%Chain%count & ! LCOV_EXCL_LINE
                                             , nd = ndim & ! LCOV_EXCL_LINE
                                             , Point = self%Chain%State(1:ndim,self%Stats%BurninLoc%compact:self%Chain%count%compact) & ! LCOV_EXCL_LINE
-                                            , Weight = self%Chain%Weight(self%Stats%BurninLoc%compact:self%Chain%count%compact) &
+                                            , Weight = self%Chain%Weight(self%Stats%BurninLoc%compact:self%Chain%count%compact) & ! LCOV_EXCL_LINE
                                             , CovMatUpper = self%Stats%Chain%CovMat & ! LCOV_EXCL_LINE
                                             , Mean = self%Stats%Chain%Mean & ! LCOV_EXCL_LINE
                                             )
@@ -1262,9 +1264,9 @@ contains
 
[…]
