### 2020-12-06 f739b1c
minor enhancements
