+++ src/kernel/ParaMonteChainFileContents_mod.f90
@@ -238,17 +238,21 @@ contains
             elseif (getLowerCase(chainFileForm)=="verbose") then
                 isVerbose = .false.
             else
+                ! LCOV_EXCL_START
                 Err%occurred = .true.
                 Err%msg = PROCEDURE_NAME//": Unrecognized chain file form: "//chainFileForm
                 return
+                ! LCOV_EXCL_STOP
             end if
 
             if (isBinary) then
                 thisForm = "unformatted"
                 if (.not. present(ndim) .or. .not. present(lenHeader) .or. .not. present(delimiter)) then
+                    ! LCOV_EXCL_START
                     Err%occurred = .true.
                     Err%msg = PROCEDURE_NAME//": If the chain file is in binary form, chainSize, lenHeader, delimiter, and ndim must be provided by the user."
                     return
+                    ! LCOV_EXCL_STOP
                 end if
             else
                 thisForm = "formatted"
@@ -256,9 +260,11 @@ contains
 
             if (fileIsOpen) then
                 if (chainFileUnit==-1) then
+                    ! LCOV_EXCL_START
                     Err%occurred = .true.
                     Err%msg = PROCEDURE_NAME//": The file located at: "//chainFilePathTrimmed//NLC//"is open, but no unit is connected to the file."//NLC
                     return
+                    ! LCOV_EXCL_STOP
                 else
                     close(chainFileUnit)
                 end if
@@ -280,22 +286,24 @@ contains
 #endif
                         )
                     if (Err%stat/=0) then
+                        ! LCOV_EXCL_START
                         Err%occurred = .true.
                         Err%msg = PROCEDURE_NAME//": Unable to open the file located at: "//chainFilePathTrimmed//NLC
                         return
+                        ! LCOV_EXCL_STOP
                     end if
                     if (allocated(Record%value)) deallocate(Record%value)
                     allocate( character(lenHeader) :: Record%value )
                     read(chainFileUnit) Record%value
                     block
-                        integer(IK)             :: processID
-                        integer(IK)             :: delRejStage
-                        real(RK)                :: meanAccRate
-                        real(RK)                :: adaptation
-                        integer(IK)             :: burninLoc
-                        integer(IK)             :: weight
-                        real(RK)                :: logFunc
-                        real(RK), allocatable   :: State(:)
+                        integer(IK)             :: processID ! LCOV_EXCL_LINE
+                        integer(IK)             :: delRejStage ! LCOV_EXCL_LINE
+                        real(RK)                :: meanAccRate ! LCOV_EXCL_LINE
+                        real(RK)                :: adaptation ! LCOV_EXCL_LINE
+                        integer(IK)             :: burninLoc ! LCOV_EXCL_LINE
+                        integer(IK)             :: weight ! LCOV_EXCL_LINE
+                        real(RK)                :: logFunc ! LCOV_EXCL_LINE
+                        real(RK), allocatable   :: State(:) ! LCOV_EXCL_LINE
                         allocate(State(ndim))
                         chainSizeDefault = 0_IK
                         loopFindChainSizeDefault: do
@@ -305,6 +313,7 @@ contains
                             elseif (is_iostat_end(Err%stat)) then
                                 exit loopFindChainSizeDefault
                             elseif (is_iostat_eor(Err%stat)) then
+                            ! LCOV_EXCL_START
                                 Err%occurred = .true.
[…]
