### 2020-12-06 eef6b12
minor enhancements
