### 2020-12-07 a39df0b
minor enhancements
