+++ src/kernel/tests/Test_Parallelism_mod.f90
@@ -45,8 +45,8 @@
 
 module Test_Parallelism_mod
 
-    use Parallelism_mod
     use Test_mod, only: Test_type
+    use Parallelism_mod
     implicit none
 
     private
@@ -54,33 +54,53 @@ module Test_Parallelism_mod
 
     type(Test_type) :: Test
 
-    type :: TestFuncRosenBrock1D_type
-        ! MATLAB reference code:
-        ! function f = objectivefcn1(x)
-        !     f = exp(-x^2) * cos(x) *sin(2*x);
-        ! end
-        ! CMD: x0 = [0.25];
-        ! CMD: options = optimset('tolx',1.e-10)
-        ! CMD: [x, f] = fminsearch(@objectivefcn1,x0,options)
-        real(RK) :: XMIN_REF = -0.471354350447655_RK
-        real(RK) :: FMIN_REF = -0.577293243101421_RK
-    contains
-        procedure, nopass :: get => getTestFuncRosenBrock1D
-    end type TestFuncRosenBrock1D_type
-
-    type :: TestFuncRosenBrock2D_type
-        ! MATLAB reference code:
-        ! function f = objectivefcn1(x)
-        !     f = exp(-(x(1)-x(2))^2 - 2*x(1)^2)*cos(x(2))*sin(2*x(2));
-        ! end
-        ! CMD: x0 = [0.25,-0.25];
-        ! CMD: options = optimset('tolx',1.e-10)
-        ! CMD: [x, f] = fminsearch(@objectivefcn1,x0,options)
-        real(RK) :: XMIN_REF(2) = [-0.169552635123540_RK, -0.508657910611664_RK]
-        real(RK) :: FMIN_REF = -0.625285429865811_RK
-    contains
-        procedure, nopass :: get => getTestFuncRosenBrock2D
-    end type TestFuncRosenBrock2D_type
+    real(RK)   , parameter :: comSecTime = 5.E-007_RK
+    real(RK)   , parameter :: parSecTime = .7E-03_RK
+    real(RK)   , parameter :: successProb = 0.234_RK
+    integer(IK), parameter :: lenProcessID = 100_IK
+    integer(IK), parameter :: processCount = 8_IK
+    integer(IK), parameter :: ProcessID(lenProcessID) = [ 1_IK, 3_IK, 7_IK, 6_IK, 6_IK, 1_IK, 2_IK, 8_IK, 8_IK, 2_IK &
+                                                        , 8_IK, 3_IK, 7_IK, 6_IK, 3_IK, 5_IK, 8_IK, 4_IK, 8_IK, 4_IK &
+                                                        , 1_IK, 5_IK, 3_IK, 3_IK, 7_IK, 6_IK, 2_IK, 2_IK, 4_IK, 3_IK &
+                                                        , 2_IK, 2_IK, 4_IK, 4_IK, 7_IK, 3_IK, 8_IK, 4_IK, 6_IK, 4_IK &
+                                                        , 3_IK, 2_IK, 8_IK, 3_IK, 8_IK, 1_IK, 1_IK, 1_IK, 4_IK, 2_IK &
+                                                        , 8_IK, 4_IK, 6_IK, 5_IK, 3_IK, 1_IK, 8_IK, 4_IK, 1_IK, 3_IK &
+                                                        , 4_IK, 3_IK, 4_IK, 8_IK, 3_IK, 6_IK, 2_IK, 7_IK, 8_IK, 4_IK &
+                                                        , 7_IK, 5_IK, 4_IK, 5_IK, 1_IK, 2_IK, 1_IK, 3_IK, 3_IK, 6_IK &
[…]
