+++ src/kernel/tests/main.f90
@@ -87,6 +87,7 @@ block; use Test_Math_mod; call test_Math(); end block
 block; use Test_Matrix_mod; call test_Matrix(); end block
 block; use Test_Misc_mod; call test_Misc(); end block
 block; use Test_Optimization_mod; call test_Optimization(); end block
+block; use Test_Parallelism_mod; call test_Parallelism(); end block
 block; use Test_Path_mod; call test_Path(); end block
 block; use Test_RandomSeed_mod; call test_RandomSeed(); end block
 block; use Test_Sort_mod; call test_Sort(); end block

[…]
