### 2020-12-07 2eb1e98
minor enhancements
