### 2020-12-07 a38480b
minor enhancements
