### 2020-12-07 c866cfe
minor enhancements
