### 2020-12-07 3f178c6
minor enhancements
