### 2020-12-07 b8423d9
minor enhancements
