+++ src/kernel/ParaMCMCRefinedChain_mod.f90
@@ -55,7 +55,7 @@ module ParaMCMCRefinedChain_mod
 
     implicit none
 
-    character(*), parameter :: MODULE_NAME = "@ParaMCMCRefinedChain_mod"
+    character(*), parameter :: MODULE_NAME = "\paramCMCRefinedChain_mod"
 
     !> The `RefinedChain_type` class.
     type                                    :: RefinedChain_type
@@ -93,17 +93,17 @@ contains
     !> Return the refined Markov chain, given the input Markov chain and its specifications.
     !> This procedure is a method of the [RefinedChain_type](@ref refinedchain_type) class.
     !>
-    !> @param[inout] RefinedChain           :   An object of class [RefinedChain_type](@ref refinedchain_type).
-    !> @param[inout] CFC                    :   An object of type [ChainFileContents_type](@ref paramontechainfilecontents_mod::chainfilecontents_type)
+    !> \param[inout] RefinedChain           :   An object of class [RefinedChain_type](@ref refinedchain_type).
+    !> \param[inout] CFC                    :   An object of type [ChainFileContents_type](@ref paramontechainfilecontents_mod::chainfilecontents_type)
     !!                                          containing the Markov chain.
-    !> @param[out]   Err                    :   An object of class [Err_type](@ref err_mod::err_type) indicating whether any error has occurred or not.
-    !> @param[in]    burninLoc              :   The estimated location of burnin point in the Markov chain (**optional**).
+    !> \param[out]   Err                    :   An object of class [Err_type](@ref err_mod::err_type) indicating whether any error has occurred or not.
+    !> \param[in]    burninLoc              :   The estimated location of burnin point in the Markov chain (**optional**).
     !!                                          If not provided, it will be extracted from the components of the input `CFC`.
-    !> @param[in]    refinedChainSize       :   The requested refined sample size (**optional**). If the size of the refined sample is given as input,
+    !> \param[in]    refinedChainSize       :   The requested refined sample size (**optional**). If the size of the refined sample is given as input,
     !!                                          then the requested sample is directly generated based on the input size.
-    !> @param[in]    sampleRefinementCount  :   The maximum number of times the sample can be refined (**optional**, default = `Infinity`).
+    !> \param[in]    sampleRefinementCount  :   The maximum number of times the sample can be refined (**optional**, default = `Infinity`).
     !!                                      :   For example, if set to 1, then only one round of refinement will be performed on the Markov chain.
-    !> @param[in]    sampleRefinementMethod :   The requested method of refining the sample (**optional**, default = "BatchMeans").
+    !> \param[in]    sampleRefinementMethod :   The requested method of refining the sample (**optional**, default = "BatchMeans").
     subroutine getRefinedChain  ( RefinedChain              &
                                 , CFC                       &
                                 , Err                       &
@@ -393,10 +393,10 @@ contains
 
     !> Return the refined vector of weights of the vector of weights of a weighted Markov chain.
     !>
-    !> @param[in]   np                  :   The number of elements of the `Weight` vector.
-    !> @param[in]   Weight              :   The input vector of weights.
-    !> @param[in]   skip                :   The size of the jumps that have to be made through the weighted Markov chain.
-    !> @param[in]   refinedChainSize    :   The requested refined sample size (**optional**). If present, then the refined chain (represented by the
+    !> \param[in]   np                  :   The number of elements of the `Weight` vector.
+    !> \param[in]   Weight              :   The input vector of weights.
[…]
