+++ src/kernel/tests/Test_ParaDXXX_mod@Test_SpecBase_smod.inc.f90
@@ -40,19 +40,23 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-!>  \brief This include file contains tests of the module [SpecBase_mod](@ref specbase_mod) when accessed by the ParaDXXX sampler.
+!>  \brief This include file contains the body of the submodules 
+!>  [ParaDRAM_mod@Test_SpecBase_smod](@ref paradram_mod@test_specbase_smod) and 
+!>  [ParaDISE_mod@@Test_SpecBase_smod](@ref paradise_mod@@test_specbase_smod).
 !>  \author Amir Shahmoradi
+
+contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     !> \brief
     !> Test the ParaDXXX sampler with a wrong input value for `chainFileFormat`.
-    function test_SpecBase_ChainFileFormat_type_1() result(assertion)
+    module function test_SpecBase_ChainFileFormat_type_1() result(assertion)
         implicit none
         logical             :: assertion
         type(ParaDXXX_type) :: PD
         assertion = .true.
-#if defined CODECOV_ENABLED
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
@@ -67,12 +71,12 @@
 
     !> \brief
     !> Test the ParaDXXX sampler with a wrong input value for `chainFileFormat`.
-    function test_SpecBase_ChainFileFormat_type_2() result(assertion)
+    module function test_SpecBase_ChainFileFormat_type_2() result(assertion)
         implicit none
         logical             :: assertion
         type(ParaDXXX_type) :: PD
         assertion = .true.
-#if defined CODECOV_ENABLED
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
@@ -87,12 +91,12 @@
 
     !> \brief
     !> Test the ParaDXXX sampler with a valid input value for `chainFileFormat`.
-    function test_SpecBase_ChainFileFormat_type_3() result(assertion)
+    module function test_SpecBase_ChainFileFormat_type_3() result(assertion)
         implicit none
         logical             :: assertion
         type(ParaDXXX_type) :: PD
         assertion = .true.
-#if defined CODECOV_ENABLED
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
         call PD%runSampler  ( ndim = 1_IK &
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
@@ -107,12 +111,12 @@
 
     !> \brief
[…]
