+++ src/kernel/SpecMCMC_ChainSize_mod.f90
@@ -79,8 +79,8 @@ contains
         implicit none
         character(*), intent(in)    :: methodName
         type(ChainSize_type)        :: ChainSizeObj
-#if defined CODECOV_ENABLED
-        ChainSizeObj%def    = 100_IK
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
+        ChainSizeObj%def    = 300_IK
 #else
         ChainSizeObj%def    = 100000_IK
 #endif
