### 2020-12-09 732994a
minor enhancements
