+++ src/kernel/Parallelism_mod (2).f90
@@ -41,7 +41,7 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 !>  \brief This module contains procedures for computing the parallel performance of the parallel algorithms.
-!>  \author Amir Shahmoradi
+!>  @author Amir Shahmoradi
 
 module Parallelism_mod
 
@@ -53,54 +53,38 @@ module Parallelism_mod
 
     character(*), parameter :: MODULE_NAME = "@Parallelism_mod"
 
-    !> The [Image_type](@ref image_type) class.
-    type                            :: Image_type
-        integer(IK)                 :: id                   !<  The ID of the runtime parallel process: 1, 2, 3, ...
-        integer(IK)                 :: count                !<  The total number of runtime parallel processes available.
-        logical                     :: isFirst = .false.    !<  A logical flag indicating whether the current process is ID #1.
-        logical                     :: isNotFirst = .false. !<  A logical flag indicating whether the current process is NOT ID #1.
-        logical                     :: isLeader = .false.   !<  A logical flag indicating whether the current process is a leader. This must be defined by the user.
-        logical                     :: isRooter = .false.   !<  A logical flag indicating whether the current process is a follower. This must be defined by the user.
-        character(:), allocatable   :: name                 !<  The name of the current process as a string with the format `"@process(id)"`.
-    contains
-        procedure, nopass           :: sync => syncImages
-        procedure, pass             :: query => queryImage
-        procedure, nopass           :: finalize => finalizeImages
-    end type Image_type
-
     type, private :: Maximum_type
-        real(RK)                    :: value            !<  The maximum speedup attained (or attainable).
-        integer(IK)                 :: nproc            !<  The required number of processes for maximum speedup.
+        real(RK)                    :: value
+        integer(IK)                 :: nproc
     end type Maximum_type
 
     type, private :: Speedup_type
-        integer(IK)                 :: count            !<  The size of the `Scaling` vector.
-        real(RK)                    :: current          !<  The speedup given the current available number of processes.
-        type(Maximum_type)          :: Maximum          !<  An object of type [Maximum_type](@ref maximum_type), containing the predicted maximum speedup and process count.
-        real(RK)    , allocatable   :: Scaling(:)       !<  A real vector of length `Speedup_type%count` containing the predicted speedup for a range process counts.
+        integer(IK)                 :: count
+        real(RK)                    :: current
+        type(Maximum_type)          :: Maximum
+        real(RK)    , allocatable   :: Scaling(:)
     end type Speedup_type
 
     type, private :: UniqueProcess_type
-        integer(IK)                 :: count            !<  The sizes of the two vector components `Identity` and `Frequency` of `UniqueProcess_type`.
[…]
