### 2020-12-09 8a98c7e
minor enhancements
