+++ src/kernel/tests/Test_Cosmology_mod.f90
@@ -67,10 +67,8 @@ contains
         call Test%run(test_getlogdvdz, "test_getlogdvdz")
         call Test%run(test_ldiswickram, "test_ldiswickram")
         call Test%run(test_getLogLumDisWicMpc, "test_getLogLumDisWicMpc")
-!#if !defined OS_IS_WSL || !defined CODECOV_ENABLED || defined DLL_ENABLED
         call Test%run(test_getLookBackTime_1, "test_getLookBackTime_1") ! The internal function passing as actual argument causes segfault with Gfortran (any version) on Windows subsystem for Linux.
         call Test%run(test_getLookBackTime_2, "test_getLookBackTime_2") ! The internal function passing as actual argument causes segfault with Gfortran (any version) on Windows subsystem for Linux.
-!#endif
         call Test%run(test_getUniverseAgeDerivative, "test_getUniverseAgeDerivative")
         call Test%finalize()
     end subroutine test_Cosmology
@@ -172,8 +170,6 @@ contains
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-!#if !defined OS_IS_WSL || !defined CODECOV_ENABLED || defined DLL_ENABLED
-
     function test_getLookBackTime_1() result(assertion)
 
         use Constants_mod, only: RK, IK
@@ -235,8 +231,6 @@ contains
         ! LCOV_EXCL_STOP
 
     end function test_getLookBackTime_2
-
-!#endif
 
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
