### 2020-12-09 d537004
minor enhancements
