+++ src/kernel/System_mod.f90
@@ -312,10 +312,13 @@ contains
 
         end if
 
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 #if defined OS_IS_WINDOWS || defined OS_IS_DARWIN || defined OS_IS_LINUX
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
         OS%name = OS_NAME
         OS%slash = OS_PATH_SEPARATOR
+
 #if defined OS_IS_WINDOWS
         OS%isWindows = .true.
 #elif defined OS_IS_DARWIN
@@ -324,7 +327,9 @@ contains
         OS%isLinux = .true.
 #endif
 
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 #else
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
         if (allocated(OS%name)) deallocate(OS%name); allocate( character(MAX_OS_NAME_LEN) :: OS%name )
         call getEnvVar( name="OS", value=OS%name, Err=OS%Err )
@@ -463,7 +468,9 @@ contains
 
         end if blockOS
 
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 #endif
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
         mv_osCacheActivated = .true.
         mv_OS%name      = OS%name
@@ -990,6 +997,8 @@ contains
     !> \param[out]  count       :   The count of elements in the output `List` (**optional**).
     !> \param[in]  cacheFile    :   The path to the external file where the results of the system information query will be stored and kept (**optional**).
     !>                              If no file is specified, the system information will not be stored in an external file.
+    !> \todo
+    !> This code can be improved. See the extensive note in the body of the procedure.
     subroutine getSystemInfo(List,Err,OS,count,cacheFile)
 #if IFORT_ENABLED && defined DLL_ENABLED && (OS_IS_WINDOWS || defined OS_IS_DARWIN) && !defined CFI_ENABLED
         !DEC$ ATTRIBUTES DLLEXPORT :: getSystemInfo
@@ -1079,21 +1088,34 @@ contains
         call executeCmd( command=command // " 2> " // stdErr, Err=Err )
         if (Err%occurred) then
         ! LCOV_EXCL_START
-            Err%msg =   PROCEDURE_NAME // ": Error occurred while attempting to write the system info to external file." // NLC // Err%msg
-            return
+            Err%msg = PROCEDURE_NAME // ": Error occurred while attempting to write the system info to external file." // NLC // Err%msg
+            ! WARNING: XXX TODO
+            ! WARNING: On some platforms, such Windows Subsystem for Linux, the CMD exit status 
+            ! WARNING: might not be returned reliably and therefore, cause `executeCmd()` to return
+            ! WARNING: an error. In such a case, no error for copy file should be really raised.
+            ! WARNING: If the file already exists upon copy action, no error should be raised.
+            ! WARNING: Note that this method may have some vulnerabilities, for example, when  
+            ! WARNING: a file copy is created, but the copy action did not accomplish the
[…]
