### 2020-12-10 9c7fd74
minor enhancements
