### 2020-12-10 dc4cdb1
minor enhancements
