### 2020-12-10 dbd6fb2
minor enhancements
