### 2020-12-10 ad8c46c
minor enhancements
