### 2020-12-10 35f0e9a
minor enhancements
