### 2020-12-10 d1b984f
minor enhancements
