+++ src/kernel/tests/Test_ParaDXXX_mod@Test_runSampler_smod.inc.f90
@@ -726,7 +726,7 @@ contains
 
         if (PD%Image%isLeader) then
 
-            RefinedChain = readRefinedChain( sampleFilePath=PD%SampleFile%Path%original, delimiter = PD%SpecBase%OutputDelimiter%val, ndim = PD%nd%val )
+            RefinedChain = readRefinedChain( sampleFilePath = PD%SampleFile%Path%original, delimiter = PD%SpecBase%OutputDelimiter%val, ndim = PD%nd%val )
 
             assertion = assertion .and. RefinedChain%numRefinement == 0_IK
 
@@ -749,6 +749,9 @@ contains
                     write(*,"(10(g0,:,', '))")
                     write(*,"(10(g0,:,', '))") "shape(PD%RefinedChain%LogFuncState) ", shape(PD%RefinedChain%LogFuncState)
                     write(*,"(10(g0,:,', '))") "shape(RefinedChain%LogFuncState)    ", shape(RefinedChain%LogFuncState)
+                    write(*,"(10(g0,:,', '))")
+                    write(*,"(10(g0,:,', '))") "PD%RefinedChain%LogFuncState        ", PD%RefinedChain%LogFuncState
+                    write(*,"(10(g0,:,', '))") "RefinedChain%LogFuncState           ", RefinedChain%LogFuncState
                     write(*,"(10(g0,:,', '))")
                 end if
                 return
@@ -965,5 +968,96 @@ contains
         ! LCOV_EXCL_STOP
 #endif
     end function test_runSampler_17
+
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+    !> \brief
+    !> Test whether the read method of the `ParaMCMCRefinedChain_type` class can successfully read an external sample file.
+    !> \remark
+    !> This is similar to test #15, except that it also verifies the functionality of the optional argument `tenabled` to `readRefinedChain()`.
+    module function test_runSampler_18() result(assertion)
+        use ParaMCMCRefinedChain_mod, only: readRefinedChain, RefinedChain_type
+        implicit none
+        logical                 :: assertion
+        real(RK)    , parameter :: tolerance = 1.e-10_RK
+        character(*), parameter :: DELIM = "delim"
+        integer(IK) , parameter :: NDIM = 2_IK
+        type(ParaDXXX_type)     :: PD
+        type(RefinedChain_type) :: RefinedChain
+        real(RK), allocatable   :: Difference(:,:)
+        assertion = .true.
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
+
+        call PD%runSampler  ( ndim = NDIM &
+                            , getLogFunc = getLogFuncMVN &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_18" &
+                            , mpiFinalizeRequested = .false. &
+                            , outputRealPrecision = 15_IK &
+                            , outputDelimiter = DELIM &
+                            , sampleSize = 10_IK &
+                            )
+        assertion = assertion .and. .not. PD%Err%occurred
+        if (.not. assertion) return ! LCOV_EXCL_LINE
+
+        if (PD%Image%isLeader) then
+
+            RefinedChain = readRefinedChain ( sampleFilePath = PD%SampleFile%Path%original & ! LCOV_EXCL_LINE
+                                            , delimiter = PD%SpecBase%OutputDelimiter%val & ! LCOV_EXCL_LINE
[…]
