+++ src/kernel/tests/Test_ParaDXXX_mod.inc.f90
@@ -76,6 +76,7 @@
         module function test_runSampler_15() result(assertion); logical :: assertion; end
         module function test_runSampler_16() result(assertion); logical :: assertion; end
         module function test_runSampler_17() result(assertion); logical :: assertion; end
+        module function test_runSampler_18() result(assertion); logical :: assertion; end
     end interface
 
     interface
@@ -248,6 +249,7 @@ contains
         call Test%run(test_runSampler_15, "test_runSampler_15")
         call Test%run(test_runSampler_16, "test_runSampler_16")
         call Test%run(test_runSampler_17, "test_runSampler_17")
+        call Test%run(test_runSampler_18, "test_runSampler_18")
 
         call Test%run(test_SpecBase_RandomSeed_type_1, "test_SpecBase_RandomSeed_type_1")
         call Test%run(test_SpecBase_RandomSeed_type_2, "test_SpecBase_RandomSeed_type_2")
