+++ src/kernel/tests/Test_ParaDXXX_mod@Test_runSampler_smod.inc.f90
@@ -391,11 +391,11 @@ contains
         ! LCOV_EXCL_STOP
 
         if (PD2%Image%isLeader) then
+        ! LCOV_EXCL_START
 
             assertion = assertion .and. all(shape(PD2%RefinedChain%LogFuncState) == shape(PD1%RefinedChain%LogFuncState))
 
             if (.not. assertion) then
-            ! LCOV_EXCL_START
                 if (Test%isDebugMode) then
                     write(*,"(10(g0,:,', '))")
                     write(*,"(10(g0,:,', '))") "shape(PD1%RefinedChain%LogFuncState)", shape(PD1%RefinedChain%LogFuncState)
@@ -404,12 +404,10 @@ contains
                 end if
                 return
             end if
-            ! LCOV_EXCL_STOP
 
             assertion = assertion .and. all( abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState) < 1.e-6_RK ) ! by default, the output precision is only 8 digits
 
             if (.not. assertion) then
-            ! LCOV_EXCL_START
                 if (Test%isDebugMode) then
                     write(Test%outputUnit,"(*(g0,:,' '))")
                     write(Test%outputUnit,"(*(g0,:,' '))")   "process, Difference:", Test%Image%id, abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState)
@@ -417,9 +415,9 @@ contains
                 end if
                 return
             end if
-            ! LCOV_EXCL_STOP
 
         end if
+        ! LCOV_EXCL_STOP
 #endif
     end function test_runSampler_9
 
@@ -490,9 +488,7 @@ contains
         assertion = assertion .and. .not. PD2%Err%occurred
         if (.not. assertion) return ! LCOV_EXCL_LINE
 
-        if (PD2%Image%isLeader) then
-            assertion = assertion .and. all( abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState) < 1.e-6_RK ) ! by default, the output precision is only 8 digits
-        endif
+        if (PD2%Image%isLeader) assertion = assertion .and. all( abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState) < 1.e-6_RK ) ! LCOV_EXCL_LINE ! by default, the output precision is only 8 digits
 
 #endif
     end function test_runSampler_10
@@ -568,9 +564,7 @@ contains
         assertion = assertion .and. .not. PD2%Err%occurred
         if (.not. assertion) return ! LCOV_EXCL_LINE
 
-        if (PD2%Image%isLeader) then
-            assertion = assertion .and. all( abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState) < 1.e-6_RK ) ! by default, the output precision is only 8 digits
-        endif
+        if (PD2%Image%isLeader) assertion = assertion .and. all( abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState) < 1.e-6_RK ) ! LCOV_EXCL_LINE ! by default, the output precision is only 8 digits
 #endif
     end function test_runSampler_11
 
@@ -645,9 +639,7 @@ contains
         assertion = assertion .and. .not. PD2%Err%occurred
         if (.not. assertion) return ! LCOV_EXCL_LINE
 
-        if (PD2%Image%isLeader) then
-            assertion = assertion .and. all( abs(PD2%RefinedChain%LogFuncState - PD1%RefinedChain%LogFuncState) < 1.e-6_RK ) ! by default, the output precision is only 8 digits
-        endif
[…]
