### 2020-12-10 e99327b
minor enhancements
