+++ src/kernel/tests/Test_ParaDXXX_mod@Test_runSampler_smod.inc.f90
@@ -960,10 +960,115 @@ contains
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     !> \brief
+    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in multichain parallelism.
+    !> Also, ensure the delayed rejection sampling is activated by setting a low value for `adaptiveUpdatePeriod` and
+    !> a large value for `scaleFactor`.
+    !> Also, use uniform proposal, as opposed to Normal in `test_runSampler_17()`.
+    module function test_runSampler_18() result(assertion)
+        use Constants_mod, only: RK
+        implicit none
+        logical             :: assertion
+        type(ParaDXXX_type) :: PD1, PD2
+        real(RK), parameter :: targetAcceptanceRate(*) = [0.2_RK, 0.23_RK]
+        assertion = .true.
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
+
+        ! Run the fresh simulation
+
+        call PD1%runSampler ( ndim = 2_IK &
+                            , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_18" &
+                            , targetAcceptanceRate = targetAcceptanceRate &
+                            , parallelizationModel = "single chain" &
+                            , greedyAdaptationCount = 30_IK & ! This must remain larger than adaptiveUpdatePeriod
+                            , delayedRejectionCount = 4_IK &
+                            , adaptiveUpdatePeriod = 1_IK &
+                            , outputRealPrecision = 16_IK &
+                            , scaleFactor = "5 * gelman" &
+                            , proposalModel = "uniform" &
+                            , randomSeed = 12345_IK &
+                            , sampleSize = 250_IK &
+                            , chainSize = 500_IK &
+                            )
+        assertion = assertion .and. .not. PD1%Err%occurred
+
+        if (.not. assertion) then
+        ! LCOV_EXCL_START
+            if (Test%isDebugMode) then
+                write(Test%outputUnit,"(*(g0,:,' '))")
+                write(Test%outputUnit,"(*(g0,:,' '))")   "process, PD1%Err%occurred(1)", Test%Image%id, PD1%Err%occurred
+                write(Test%outputUnit,"(*(g0,:,' '))")
+            end if
+            return
+        end if
+        ! LCOV_EXCL_STOP
+
+        call Test%Image%sync()
+
+        block
+            use System_mod, only: removeFile
+            call removeFile(PD1%SampleFile%Path%original, PD1%Err) ! delete the sample file
+        end block
+
+        call Test%Image%sync()
+
+        ! restart the simulation with the same configuration
+
+        call PD2%runSampler ( ndim = 2_IK &
+                            , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_18" &
+                            , targetAcceptanceRate = targetAcceptanceRate &
+                            , parallelizationModel = "single chain" &
+                            , greedyAdaptationCount = 30_IK & ! This must remain larger than adaptiveUpdatePeriod
[…]
