### 2020-12-10 0fbe258
minor enhancements
