+++ src/kernel/tests/Test_ParaDXXX_mod.inc.f90
@@ -78,6 +78,7 @@
         module function test_runSampler_17() result(assertion); logical :: assertion; end
         module function test_runSampler_18() result(assertion); logical :: assertion; end
         module function test_runSampler_19() result(assertion); logical :: assertion; end
+        module function test_runSampler_20() result(assertion); logical :: assertion; end
     end interface
 
     interface
@@ -252,6 +253,7 @@ contains
         call Test%run(test_runSampler_17, "test_runSampler_17")
         call Test%run(test_runSampler_18, "test_runSampler_18")
         call Test%run(test_runSampler_19, "test_runSampler_19")
+        call Test%run(test_runSampler_20, "test_runSampler_20")
 
         call Test%run(test_SpecBase_RandomSeed_type_1, "test_SpecBase_RandomSeed_type_1")
         call Test%run(test_SpecBase_RandomSeed_type_2, "test_SpecBase_RandomSeed_type_2")
