+++ src/kernel/tests/Test_ParaDXXX_mod@Test_runSampler_smod.inc.f90
@@ -858,7 +858,7 @@ contains
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     !> \brief
-    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in multichain parallelism.
+    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in singlechain parallelism.
     !> Also, ensure the delayed rejection sampling is activated by setting a low value for `adaptiveUpdatePeriod` and
     !> a large value for `scaleFactor`.
     module function test_runSampler_17() result(assertion)
@@ -960,7 +960,7 @@ contains
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     !> \brief
-    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in multichain parallelism.
+    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in singlechain parallelism.
     !> Also, ensure the delayed rejection sampling is activated by setting a low value for `adaptiveUpdatePeriod` and
     !> a large value for `scaleFactor`.
     !> Also, use uniform proposal, as opposed to Normal in `test_runSampler_17()`.
@@ -1020,7 +1020,7 @@ contains
                             , mpiFinalizeRequested = .false. &
                             , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_18" &
                             , targetAcceptanceRate = targetAcceptanceRate &
-                            , parallelizationModel = "single chain" &
+                            , parallelizationModel = "singlechain" &
                             , greedyAdaptationCount = 30_IK & ! This must remain larger than adaptiveUpdatePeriod
                             , delayedRejectionCount = 4_IK &
                             , adaptiveUpdatePeriod = 1_IK &
@@ -1065,10 +1065,116 @@ contains
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     !> \brief
+    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in multichain parallelism.
+    !> Also, ensure the delayed rejection sampling is activated by setting a low value for `adaptiveUpdatePeriod` and
+    !> a large value for `scaleFactor`.
+    !> Also, avoid delayed rejection, as opposed to what is done in `test_runSampler_17()`.
+    !> Also, request the `maxCumSum` sample refinement and count method.
+    module function test_runSampler_19() result(assertion)
+        use Constants_mod, only: RK
+        implicit none
+        logical             :: assertion
+        type(ParaDXXX_type) :: PD1, PD2
+        real(RK), parameter :: targetAcceptanceRate(*) = [0.2_RK, 0.23_RK]
+        assertion = .true.
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
+
+        ! Run the fresh simulation
+
+        call PD1%runSampler ( ndim = 2_IK &
[…]
