### 2020-12-10 b02c95e
minor enhancements
