+++ src/kernel/tests/Test_ParaDXXX_mod@Test_runSampler_smod.inc.f90
@@ -1085,7 +1085,7 @@ contains
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
                             , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_19" &
-                            , sampleRefinementMethod = "MaxCumSumAutoCorr" &
+                            , sampleRefinementMethod = "MaxCumSumAutoCorr-median" &
                             , targetAcceptanceRate = targetAcceptanceRate &
                             , parallelizationModel = "multi chain" &
                             , greedyAdaptationCount = 2_IK & ! This must remain larger than adaptiveUpdatePeriod
@@ -1125,7 +1125,7 @@ contains
                             , getLogFunc = getLogFuncMVN &
                             , mpiFinalizeRequested = .false. &
                             , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_19" &
-                            , sampleRefinementMethod = "MaxCumSumAutoCorr" &
+                            , sampleRefinementMethod = "MaxCumSumAutoCorr-median" &
                             , targetAcceptanceRate = targetAcceptanceRate &
                             , parallelizationModel = "multi chain" &
                             , greedyAdaptationCount = 2_IK & ! This must remain larger than adaptiveUpdatePeriod
@@ -1171,10 +1171,116 @@ contains
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     !> \brief
+    !> Test the ParaDXXX sampler to restart a complete simulation without an output sample file, in multichain parallelism.
+    !> Also, ensure the delayed rejection sampling is activated by setting a low value for `adaptiveUpdatePeriod` and
+    !> a large value for `scaleFactor`.
+    !> Also, avoid delayed rejection, as opposed to what is done in `test_runSampler_17()`.
+    !> Also, request the `cutoffAutoCorr` sample refinement and count method.
+    module function test_runSampler_20() result(assertion)
+        use Constants_mod, only: RK
+        implicit none
+        logical             :: assertion
+        type(ParaDXXX_type) :: PD1, PD2
+        real(RK), parameter :: targetAcceptanceRate(*) = [0.2_RK, 0.23_RK]
+        assertion = .true.
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
+
+        ! Run the fresh simulation
+
+        call PD1%runSampler ( ndim = 2_IK &
+                            , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_20" &
+                            , sampleRefinementMethod = "cutoffAutoCorr-maximum" &
+                            , targetAcceptanceRate = targetAcceptanceRate &
+                            , parallelizationModel = "multi chain" &
+                            , greedyAdaptationCount = 2_IK & ! This must remain larger than adaptiveUpdatePeriod
+                            , adaptiveUpdatePeriod = 1_IK &
+                            , outputRealPrecision = 16_IK &
+                            , scaleFactor = "5 * gelman" &
+                            , proposalModel = "uniform" &
+                            , randomSeed = 12345_IK &
+                            , sampleSize = -50_IK &
[…]
