### 2020-12-10 20e5771
minor enhancements
