### 2020-12-10 c91371c
minor enhancements
