+++ src/kernel/ParaDXXX_mod@Kernel_smod.inc.f90
@@ -247,9 +247,11 @@ contains
                                 , delimiter = self%SpecBase%OutputDelimiter%val     &
                                 )
             if (self%Err%occurred) then
+                ! LCOV_EXCL_START
                 self%Err%msg = PROCEDURE_NAME//self%Err%msg
                 call self%abort( Err = self%Err, prefix = self%brand, newline = NLC, outputUnit = self%LogFile%unit )
                 return
+                ! LCOV_EXCL_STOP
             end if
 
             if (self%Chain%Count%compact<=CHAIN_RESTART_OFFSET) then
@@ -277,40 +279,44 @@ contains
                     RFN = RandomFileName_type(dir = "", key = self%ChainFile%Path%original//".rst", ext="") ! temporary_restart_copy
                     call copyFile(pathOld=self%ChainFile%Path%original,pathNew=RFN%path,isUnixShell=self%OS%Shell%isUnix,Err=self%err)
                     if (self%Err%occurred) then
+                        ! LCOV_EXCL_START
                         self%Err%msg = PROCEDURE_NAME//self%Err%msg
                         call self%abort( Err = self%Err, prefix = self%brand, newline = NLC, outputUnit = self%LogFile%unit )
                         exit blockLeaderSetup
                         return
+                        ! LCOV_EXCL_STOP
                     end if
 
                     ! reopen the chain file to resume the simulation
 
-                    open( newunit = self%ChainFile%unit             &
-                        , file = self%ChainFile%Path%original       &
-                        , form = self%ChainFile%Form%value          &
-                        , status = self%ChainFile%status            &
-                        , iostat = self%ChainFile%Err%stat          &
+                    open( newunit = self%ChainFile%unit             & ! LCOV_EXCL_LINE
+                        , file = self%ChainFile%Path%original       & ! LCOV_EXCL_LINE
+                        , form = self%ChainFile%Form%value          & ! LCOV_EXCL_LINE
+                        , status = self%ChainFile%status            & ! LCOV_EXCL_LINE
+                        , iostat = self%ChainFile%Err%stat          & ! LCOV_EXCL_LINE
 #if defined IFORT_ENABLED && defined OS_IS_WINDOWS
-                        , SHARED                                    &
+                        , SHARED                                    & ! LCOV_EXCL_LINE
 #endif
                         , position = self%ChainFile%Position%value  )
                     self%Err = self%ChainFile%getOpenErr(self%ChainFile%Err%stat)
                     if (self%Err%occurred) then
+                        ! LCOV_EXCL_START
                         self%Err%msg = PROCEDURE_NAME//": Error occurred while opening "//self%name//" "//self%ChainFile%suffix//" file='"//self%ChainFile%Path%original//"'."
                         call self%abort( Err = self%Err, prefix = self%brand, newline = NLC, outputUnit = self%LogFile%unit )
                         exit blockLeaderSetup
                         return
+                        ! LCOV_EXCL_STOP
                     end if
 
                     ! rewrite the chain file
 
-                    call self%Chain%writeChainFile  ( ndim = nd &
-                                                    , compactStartIndex = 1_IK &
-                                                    , compactEndIndex = self%Chain%Count%compact-CHAIN_RESTART_OFFSET &
-                                                    , chainFileUnit = self%ChainFile%unit &
-                                                    , chainFileForm = self%SpecBase%ChainFileFormat%val &
-                                                    , chainFileFormat = self%ChainFile%format &
-                                                    , adaptiveUpdatePeriod = self%SpecDRAM%AdaptiveUpdatePeriod%val &
+                    call self%Chain%writeChainFile  ( ndim = nd & ! LCOV_EXCL_LINE
[…]
