### 2020-12-10 cec9e28
minor enhancements
