### 2020-12-11 77b0b5b
minor enhancements
