+++ src/kernel/tests/Test_ParaDXXX_mod.inc.f90
@@ -80,6 +80,7 @@
         module function test_runSampler_19() result(assertion); logical :: assertion; end
         module function test_runSampler_20() result(assertion); logical :: assertion; end
         module function test_runSampler_21() result(assertion); logical :: assertion; end
+        module function test_runSampler_22() result(assertion); logical :: assertion; end
     end interface
 
     interface
@@ -256,6 +257,7 @@ contains
         call Test%run(test_runSampler_19, "test_runSampler_19")
         call Test%run(test_runSampler_20, "test_runSampler_20")
         call Test%run(test_runSampler_21, "test_runSampler_21")
+        call Test%run(test_runSampler_22, "test_runSampler_22")
 
         call Test%run(test_SpecBase_RandomSeed_type_1, "test_SpecBase_RandomSeed_type_1")
         call Test%run(test_SpecBase_RandomSeed_type_2, "test_SpecBase_RandomSeed_type_2")
