+++ src/kernel/tests/Test_ParaDXXX_mod@Test_runSampler_smod.inc.f90
@@ -732,7 +732,8 @@ contains
                 return
             end if
 
-            assertion = assertion .and. all(shape(RefinedChain%LogFuncState) == shape(PD%RefinedChain%LogFuncState))
+            ! NOTE: Keep in mind that `PD%RefinedChain` is a weighted chain internally, but unweighted when read from the external file.
+            assertion = assertion .and. size(RefinedChain%LogFuncState(:,1))==size(PD%RefinedChain%LogFuncState(:,1)) .and. sum(PD%RefinedChain%Weight)==size(PD%RefinedChain%LogFuncState(1,:))
 
             if (.not. assertion) then
                 if (Test%isDebugMode) then
@@ -1299,9 +1300,10 @@ contains
 
         call PD%runSampler  ( ndim = NDIM &
                             , getLogFunc = getLogFuncMVN &
-                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_21" &
                             , mpiFinalizeRequested = .false. &
+                            , outputFileName = Test%outDir//"/"//MODULE_NAME//"/test_runSampler_21" &
                             , sampleRefineMentMethod = "batchMeans-med" &
+                            , parallelizationModel = "multichain" &
                             , outputRealPrecision = 15_IK &
                             , outputDelimiter = DELIM &
                             , sampleSize = 10_IK &
@@ -1336,7 +1338,8 @@ contains
             end if
             ! LCOV_EXCL_STOP
 
-            assertion = assertion .and. all(shape(RefinedChain%LogFuncState) == shape(PD%RefinedChain%LogFuncState))
+            ! NOTE: Keep in mind that `PD%RefinedChain` is a weighted chain internally, but unweighted when read from the external file.
+            assertion = assertion .and. size(RefinedChain%LogFuncState(:,1))==size(PD%RefinedChain%LogFuncState(:,1)) .and. sum(PD%RefinedChain%Weight)==size(PD%RefinedChain%LogFuncState(1,:))
 
             if (.not. assertion) then
             ! LCOV_EXCL_START
@@ -1373,5 +1376,68 @@ contains
 
 #endif
     end function test_runSampler_21
+
+!%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+    !> \brief
+    !> Test whether an object of [ChainFileContents_type](@ref paramcmcrefinedchain_mod::chainfilecontents_type)
+    !> class can handle zero-count input MCMC chain for refinement.
+    !> \remark
+    module function test_runSampler_22() result(assertion)
+        use ParaMCMCRefinedChain_mod, only: readRefinedChain, RefinedChain_type
+        implicit none
+        logical                 :: assertion
+        real(RK)    , parameter :: tolerance = 1.e-10_RK
+        character(*), parameter :: DELIM = "delim"
+        integer(IK) , parameter :: NDIM = 2_IK
+        type(ParaDXXX_type)     :: PD
+        type(RefinedChain_type) :: RefinedChain
+        assertion = .true.
+#if defined CODECOV_ENABLED || defined SAMPLER_TEST_ENABLED
+
+        call PD%runSampler  ( ndim = NDIM &
+                            , getLogFunc = getLogFuncMVN &
+                            , mpiFinalizeRequested = .false. &
[…]
