+++ buildParaMonte.sh
@@ -60,10 +60,39 @@ cmakeVersionParaMonteCompatible="3.14.0"
 intelVersionParaMonteCompatible="18.0.0"
 
 ####################################################################################################################################
+#### set up color coding
+####################################################################################################################################
+
+Esc="\e"
+ColorReset="${Esc}[m"
+ColorBold="${Esc}[1m"
+Red="${Esc}[31m"
+Green="${Esc}[32m"
+Yellow="${Esc}[33m"
+Blue="${Esc}[34m"
+Magenta="${Esc}[35m"
+Cyan="${Esc}[36m"
+White="${Esc}[37m"
+BoldRed="${Esc}[1;31m"
+BoldGreen="${Esc}[1;32m"
+BoldYellow="${Esc}[1;33m"
+BoldBlue="${Esc}[1;34m"
+BoldMagenta="${Esc}[1;35m"
+BoldCyan="${Esc}[1;36m"
+BoldWhite="${Esc}[1;37m"
+
+####################################################################################################################################
 #### set up the main root paths
 ####################################################################################################################################
 
 BUILD_NAME="ParaMonte"; export BUILD_NAME
+
+pmcolor="${BoldCyan}"
+pmattn=" ${pmcolor}-- ${BUILD_NAME} -${ColorReset}"
+pmnote="${pmattn} ${BoldMagenta}NOTE:${ColorReset}"
+pmwarn="${pmattn} ${BoldYellow}WARNING:${ColorReset}"
+pmfatal="${pmattn} ${BoldRed}FATAL:${ColorReset}"
+warning="${pmattn} ${BoldYellow}WARNING${ColorReset}"
 
 workingDir="$(pwd)"
 sourceFileDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
@@ -80,9 +109,9 @@ fi
 
 if [[ ! -f "$(pwd)/build${BUILD_NAME}.sh" ]]; then
     echo >&2
-    echo >&2 "-- ${BUILD_NAME} - FATAL: Build failed."
-    echo >&2 "-- ${BUILD_NAME} - FATAL: Please run this script inside the top-level ParaMonte library root directory."
-    echo >&2 "-- ${BUILD_NAME} - FATAL: This is the directory which contains this file in the GitHub repository of ParaMonte."
+    echo >&2 "${pmfatal} Build failed."
+    echo >&2 "${pmfatal} Please run this script inside the top-level ParaMonte library root directory."
+    echo >&2 "${pmfatal} This is the directory which contains this file in the GitHub repository of ParaMonte."
     echo >&2
     exit 1
 fi
@@ -151,41 +180,41 @@ verify() {
         echo >&2 "-- ${BUILD_NAME} - ParaMonte $2 appears to have succeeded."
     else
         echo >&2
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: ParaMonte $2 appears to have failed."
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: If the source of the error cannot be identified,"
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: consider a fresh installation of ParaMonte's required compilers by calling"
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
-        echo >&2 "    -- ${BUILD_NAME} - FATAL:     ./install.sh --fresh"
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: "
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: If the error happens during the installation of ParaMonte prerequisites,"
-        echo >&2 "    -- ${BUILD_NAME} - FATAL: it is possible that the current existing GNU compiler collection installed"
[…]
