+++ buildParaMonte.sh
@@ -92,7 +92,7 @@ pmattn=" ${pmcolor}-- ${BUILD_NAME} -${ColorReset}"
 pmnote="${pmattn} ${BoldMagenta}NOTE:${ColorReset}"
 pmwarn="${pmattn} ${BoldYellow}WARNING:${ColorReset}"
 pmfatal="${pmattn} ${BoldRed}FATAL:${ColorReset}"
-warning="${pmattn} ${BoldYellow}WARNING${ColorReset}"
+warning="${BoldYellow}WARNING${ColorReset}"
 
 workingDir="$(pwd)"
 sourceFileDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
@@ -1007,17 +1007,17 @@ if [ "${CAF_ENABLED}" = "true" ]; then
 
         if [ -z "${intelFortranMpiWrapperPath+x}" ]; then
             echo >&2
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: Failed to identify the Intel MPI library."
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: The Intel MPI library is required to compile"
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: Coarray Fortran applications via Intel compilers."
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: The ParaMonte build will continue at the risk of failing..."
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: Failed to identify the Intel MPI library."
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: The Intel MPI library is required to compile"
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: Coarray Fortran applications via Intel compilers."
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: The ParaMonte build will continue at the risk of failing..."
             echo >&2
         elif [ -z "${intelFortranCompilerPath+x}" ]; then
             echo >&2
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: Failed to identify the Intel Fortran compiler path."
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: The Intel MPI library and Fortran compiler are required"
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: to compile Coarray Fortran applications via Intel compilers."
-            echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: The ParaMonte build will continue at the risk of failing..."
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: Failed to identify the Intel Fortran compiler path."
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: The Intel MPI library and Fortran compiler are required"
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: to compile Coarray Fortran applications via Intel compilers."
+            echo >&2 "-- ${BUILD_NAME}CAF - WARNING: The ParaMonte build will continue at the risk of failing..."
             echo >&2
         else
             cafCompilerPath="${intelFortranCompilerPath}"
@@ -1042,8 +1042,8 @@ if [ "${CAF_ENABLED}" = "true" ]; then
             compareVersions "$cafVersion" "$cafVersionRequired"
             if [ "$?" = "2" ]; then
                 echo >&2
-                echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: The existing OpenCoarrays caf compiler wrapper version is not ParaMonte compatible."
-                echo >&2 "-- ${BUILD_NAME}CAF - ${warning}: A fresh installation of the OpenCoarrays library might be needed."
[…]
