### 2020-12-14 #2741 @shahmoradi
I have addressed @williamfgc comments and I am working on @milancurcic 's comments. I should be able to back with a full revision by tomorrow, hopefully.
