### 2020-12-18 cdslaborg/paramonte#4 @shahmoradi
Milan Thanks for pointing this out. The acknowledgment statement that you see is a relic of the initial custom license of the library. It is now fixed.
