### 2020-12-18 cdslaborg/paramonte#3 @shahmoradi
@milancurcic Thanks for pointing this out. This was apparently an issue with the way OpenCoarrays library creates its setup files, sometimes creating the file and sometimes not, depending on whether it finds a suitable MPI library at installation time. It should be now fixed. If it happens again, please let us know, perhaps along with the specific configuration with which you are building the library.
