### 2020-12-25 #2741 @shahmoradi
@williamfgc

Thank you for your extensive detailed feedback, which significantly improved the quality of this codebase. We will address each of your comments below.

1.  **The library tests**.
    The library originally had a significant number of tests that used to be automatically executed. But at some point, someone (perhaps myself) decided to temporarily disable the tests, which then became permanent since nobody reactivated the tests.

    We have now added a comprehensive set of 866 new unique tests that cover nearly 100% of the entire library as well as the various functionalities of the samplers in the library. Given the three possible major parallelism configurations of the library, there are overall `3 * 866 = 2598` tests for all serial as well as the MPI and Coarray parallel builds. The code coverage report for all three builds for the most recent version of the library is available here:
    + [serial](https://cdslaborg.github.io/paramonte-codecov/kernel/serial/)
    + [MPI parallel](https://cdslaborg.github.io/paramonte-codecov/kernel/mpi/)
    + [Coarray parallel](https://cdslaborg.github.io/paramonte-codecov/kernel/caf/)

    Since the library relies on preprocessing directives to implement different parallelism paradigms (serial, MPI, and Coarray Fortran), the tests had to be also adapted and preprocessed according to the characteristics of the library for different parallelism paradigms. No testing framework that we have inspected so far, met the minimum requirements for testing this package under the different build configurations. We have therefore developed a unit testing framework that is specifically tailored for the needs of this library.
    The tests in the library are separated into two different categories:
    1.  `basic`: Tests that verify the functionality and accuracy of the fundamental backbone modules and procedures of the library.
    1.  `sampler`: Tests that verify the functionality, semantics, and accuracy of the library's samplers.

    By default both categories of tests are deactivated for any particular build for the following reasons:
    1.  The tests can take significant time and memory, in particular, when building the library for multiple different configurations. This can also easily break Continuous Integration workflows like Travis CI that limit the maximum length of the stdout log file.
    2.  Some of the tests such as those that verify the restart functionality of the library's sampler in parallel require external intervention in the regular working of the samplers. This in turn requires the MPI/Coarray processes to be able to communicate the error messages with each other to avoid deadlocks. These error message communications between different processes are not normally needed as the failing process automatically calls `mpi_abort()` in MPI or `error stop` in Coarray to halt the entire program.
    However, such exceptions and errors should be gracefully handled during testing, which requires the addition of extra error handling code snippets to the sampler routines that are otherwise deactivated in normal production runs.
    These extra exception and error handling communications in parallel modes can be quite expensive if they appear in the parallel production code. As such, the library's tests are by default switched off, unless the developer (or the user) readily but explicitly enables the tests via the `-t all` or `--test all` or `--codecov` build flag. For example,
    ```bash
    ./install.sh --lang fortran --build testing --lib dynamic --par none --test all
    ```
    will build the library for usage from the Fortran language with all tests activated.  The tests can be also limited to one category, if desired, to reduce the testing and build time. This is particularly needed during the development of the library when the focus is on one particular functionality in the library. For example,
    ```bash
    ./install.sh --lang c --build testing --lib dynamic --par mpi -t basic
    ```
    will test the library build for parallel usage from the C language and will only activate the basic tests of the library. The default mode, as mentioned above, is,
    ```bash
    ./install.sh -t none
    ```
    However, when generating code coverage, all tests are activated by default. For example,
    ```bash
    ./install.sh --codecov
    ```
    and,
    ```bash
    ./install.sh --codecov -t all
    ```
    are equivalent and,
    ```bash
    ./install.sh --codecov -t none
    ```
    leads to an error message, since a minimal set of tests are required to generate the code coverage report.

2.  **Continuous Integration**.
    We have now enabled [Travis CI](https://travis-ci.org/github/cdslaborg/paramonte/builds/749668477) for this library. All builds are successful. Due to technical difficulties with Coarray Fortran and Windows, CAF parallel builds and builds for Windows are currently excluded from Travis CI. Intel has recently released its new version of Fortran compilers free of charge for all platforms including Windows. This will hopefully enable the testing and CI of this library also on Windows in the near future.

3.  **Contribution policy**.
    We have now laid out a set of guidelines for contributing to the core of library or other contributions:
    +   [Defect reports](https://www.cdslab.org/paramonte/notes/contributing/defects/)
    +   [Enhancement requests](https://www.cdslab.org/paramonte/notes/contributing/enhancements/)
    +   [Development - general guidelines](https://www.cdslab.org/paramonte/notes/contributing/development/general/)
    +   [Development - core library guidelines](https://www.cdslab.org/paramonte/notes/contributing/development/kernel/)

4. **CMake support**.
    CMake is indeed already supported on Linux and macOS. However, direct use of CMake to build the library is suboptimal, in particular, for the end-users as CMake cannot install missing libraries on the user's system if needed. One of our primary goals in the development of this package was to make a compiled-language library as accessible as possible to anyone, even those who do not have experience with compiled languages or CMake or different library dependencies. To achieve this goal, we had to develop multiple layers of build scripts that lay the ground for CMake to successfully build the library. The current build scripts are capable of automatically installing all missing components including the CMake application (if the minimum required version is not detected on the system), GNU compilers, as well as OpenMPI, MPICH MPI, and OpenCoarrays libraries.

    The situation on Windows is slightly more complex. While library build with CMake was originally supported on Windows, we eventually dropped it in favor of developing a home-grown separate Windows Batch build system for the library exclusively on Windows. The reason for this decision was CMake's poor support of the Fortran programming language at the time, especially on Windows. The CMake has dramatically improved over the past three years and given the recent availability of Intel compilers on Windows free of charge, CMake support for Windows is now among top priorities to add to the library.

5.  **Code coverage**.
    Code coverage reports are now added for every new release of the library and are permanently available on GitHub in a separate repository for the three most different builds of the library (the three parallelism paradigms):
    + [serial](https://cdslaborg.github.io/paramonte-codecov/kernel/serial/)
    + [MPI parallel](https://cdslaborg.github.io/paramonte-codecov/kernel/mpi/)
    + [Coarray parallel](https://cdslaborg.github.io/paramonte-codecov/kernel/caf/)

    We have strived for adding an automated code coverage report via [codecov](https://codecov.io/gh/cdslaborg/paramonte) but so far in vain. We will continue our efforts to generate automated code coverage reports and analyses via codecov/TravisCI or similar platforms.

6.  **Documentation**.
    + The library documentation for the end-users is available here: https://www.cdslab.org/paramonte/index.html
    + The library's core API documentation is also available through the same link in the above, or directly at: https://cdslaborg.github.io/paramonte-kernel-doc/html/

7.  **Manuscript's contents and focus**.
    We have now significantly reduced the amount of technical contents of the paper.

8.  **Package redistribution and CI**.
    We absolutely agree that CI is necessary for the success and longevity of this repository and we have now set up an automated build and testing workflow with Travis-CI. We also agree that having a standard distribution method is essential in the long term. There have been, however, a few challenges along our way on identifying and setting up the best method of distribution, most importantly, the many different configurations with which the library can be built, such as the programming languages, the MPI library dependencies, etc. Furthermore, there is the question of which packaging and redistribution method would be the best. On Windows and macOS, the task might be rather straightforward as only a few popular methods exist. The situation, however, is completely different on Linux OS.

    Currently, the users can access the prebuilt versions of the library for almost any production configuration they might need. The prebuilt libraries are always available on [the GitHub release page of the project](https://github.com/cdslaborg/paramonte/releases). Downloading and using these prebuilt libraries takes only three steps from the bash terminal. For example, the prebuilt library for the C programming language can be downloaded and run via,
    ```bash
    libname=libparamonte_c_linux_x64_intel_release_dynamic_heap
    wget https://github.com/cdslaborg/paramonte/releases/latest/download/$libname.tar.gz
    tar xvzf $libname.tar.gz && cd $libname && ./build.sh && ./run.sh && cd ..
    ```

@milancurcic

Thank you for your valuable feedback, which significantly improved the quality of this work. Below, we will address your comments and questions,

1.  **OpenMPI releases**.
    Thank you for pointing this out. As you mentioned, OpenMPI is a popular implementation. We have now added the prebuilt versions of the library with OpenMPI in addition to MPICH. This required a complete revision of the naming convention used in the library across all languages and documentations. The new naming convention automatically suffixes,
    1.  any MPICH build of the library with *_mpich*,
    1.  any OpenMPI build of the library with *_openmpi*,
    1.  any Intel MPI build of the library with *_impi*,
    A complete description of the naming convention used in the library is provided in [the library's documentation](https://www.cdslab.org/paramonte/notes/installation/readme/).

2.  **Build failure**.
    Thank you for pointing this out. This apparently happens on systems where the either the MPI library or the GNU compilers are already installed and meet the minimum version requirements of ParaMonte. As such, the build scripts would not feel the need to create a custom `setup.sh` to set up the environmental variables. However, it would require the setup file's existence, which would subsequently interrupt the build by throwing an exception.

    We believe this issue is now fixed in [the latest release of the library](https://github.com/cdslaborg/paramonte/archive/main.tar.gz).

3.  **license and ACKNOWLEDGMENT.md**.
    Thank you again for pointing this out. The existing `ACKNOWLEDGMENT.md` was a relic of the original license of the library. [This is now fixed](https://github.com/cdslaborg/paramonte#citing-paramonte).

4.  **points 1 and 3 by @williamfgc**.
    Please see our responses to comments by @williamfgc.
