### 2021-01-06 cdslaborg/paramonte#10 @milancurcic
General: How to run in parallel?
I'm playing with the example provided with the binary release (libparamonte_fortran_linux_x64_gnu_release_dynamic_heap_openmpi) and I'm not clear how to run in parallel.

In serial mode things seem okay. However, if I do:

```
./run.sh -n 2
```

I get this message in stdout (among others):

```
        ParaDRAM - NOTE: ParaDRAM is being used in parallel mode but with only one processor. This is computationally inefficient.
        ParaDRAM - NOTE: Consider using the serial version of the code or provide more processes at runtime if it is beneficial.
```

I get the same result if I change the number of processes, or if I run the program directly, `mpiexec -n 2 ./main.exe`.

In reference to https://github.com/openjournals/joss-reviews/issues/2741
