### 2021-01-06 cdslaborg/paramonte#11 @milancurcic
run.sh should error out on a bad CLI argument
Using run.sh provided in libparamonte_fortran_linux_x64_gnu_release_dynamic_heap_openmpi binary release.

Try:

```
$ ./run.sh --bad-arg
```

or similar. The script should give a helpful message to the user, but instead, nothing is output.

On a positive side, the status code is non-zero, as expected:

```
$ ./run.sh --bad-arg
$ echo $?
1
```

This is in reference to https://github.com/openjournals/joss-reviews/issues/2741
