### 2021-01-06 cdslaborg/paramonte#12 @milancurcic
Enhancement: Contributing guidelines
Have contributing guidelines, ideally in a CONTRIBUTING.md file.

https://github.com/openjournals/joss-reviews/issues/2741
