### 2021-01-07 cdslaborg/paramonte#10 @shahmoradi
Milan, thanks for raising this issue. The message you have copied in the above implies that you are not running the executable in parallel, rather, multiple copies of the program are being executed serially. The only other scenario that could lead to such a message being printed on the screen, is when an MPI parallelized sampler is used with only a single processor, like: `./run.sh -n 1`.

Another scenario that can cause an MPI-parallelized program to not properly run in parallel is when an OpenMPI-parallelized executable is invoked via an MPICH `mpiexec` launcher and vice versa. Can you make confirm that you are indeed using the OpenMPI `mpiexec` launcher with a `libparamonte_*` library compiled with `OpenMPI`? I have tested all binaries on the release page on multiple Linux platforms and so far, not seen any problems, as long as both the `mpiexec` and the MPI library version and brand (MPICH / OpenMPI) match the ones used to build the binaries. The MPI library versions are mentioned on the [release page](https://github.com/cdslaborg/paramonte/releases/tag/v1.5.1). To get the MPI library version, try,
```bash
mpiexec --version
```
on the command line.
