### 2021-01-07 cdslaborg/paramonte#11 @shahmoradi
Thanks, Milan for catching this subtle, albeit minor bug. The script was indeed able to capture the wrong input argument, however, an extra `\` prevented the relevant error messages to be displayed despite the script exiting with an error code. Since this is a minor correction to only the example's temporary run scripts, I have modified the existing release with this correction. The new binaries on the release page should resolve this issue now.
