### 2021-01-07 cdslaborg/paramonte#12 @shahmoradi
Thanks, Milan, The contribution guidelines are already provided in [the documentation](https://www.cdslab.org/paramonte/notes/contributing/development/general/). For the sake of completeness, however, I have now added the guidelines to the GitHub repository as well: [CONTRIBUTING.md](https://github.com/cdslaborg/paramonte/blob/main/CONTRIBUTING.md), and [CONTRIBUTIN-KERNELG.md](https://github.com/cdslaborg/paramonte/blob/main/CONTRIBUTING-KERNEL.md)
