### 2021-01-27 cdslaborg/paramonte#13 @williamfgc
Defect: building paramonte main on Ubuntu 20.04 fails with GNU
**The programming language (e.g., all/C/C++/Fortran/Julia/MATLAB/Python/R/...)**:
Fortran

**The compiler/interprerter (e.g., none/GNU/Intel/IBM/NAG/Cray/PGI-NVIDIA/Anaconda/...)**:
GNU

**The operating system (e.g., all/Windows/Linux/macOS/WSL1/WSL2/...)**:
Linux Ubuntu 20.04

**The architecture (e.g., all/x86_x64/Intelx64/ARM/POWER9/...)**:
x86_64

**Describe the bug**
Paramonte `main` fails to build on standard Ubuntu 20 (configures fine with cmake) with error message:

```
$ make -j 8
Scanning dependencies of target paramonte_c_linux_x64_gnu_release_static_stack
make[2]: *** No rule to make target '../src/kernel/ParaMonte_mod@version@kernel.inc.f90', needed by 'obj/CMakeFiles/paramonte_c_linux_x64_gnu_release_static_stack.dir/ParaMonte_mod.f90.o'.  Stop.
make[2]: *** Waiting for unfinished jobs....
[  1%] Building Fortran object obj/CMakeFiles/paramonte_c_linux_x64_gnu_release_static_stack.dir/DateTime_mod.f90.o
[  1%] Building Fortran object obj/CMakeFiles/paramonte_c_linux_x64_gnu_release_static_stack.dir/Constants_mod.f90.o
make[1]: *** [CMakeFiles/Makefile2:171: obj/CMakeFiles/paramonte_c_linux_x64_gnu_release_static_stack.dir/all] Error 2
make: *** [Makefile:130: all] Error 2
```

**To Reproduce**
```
$ cmake -DPMCS=GNU ..
$ make -j
```

**Expected behavior**
Paramonte `main` should build cleanly

**Additional context**
[JOSS review](https://github.com/openjournals/joss-reviews/issues/2741)
