### 2021-01-28 cdslaborg/paramonte#13 @shahmoradi
Thank you for raising this issue. This defect is now resolved. Further guidance is provided [here in this file](https://github.com/cdslaborg/paramonte/blob/main/install.cmake.md) on how to build the library directly via CMake, although we do not recommend doing so as it involves a long sequence of manual steps.
