### 2021-01-28 cdslaborg/paramonte#10 @milancurcic
Thank you for the nudge @shahmoradi and sorry for not acting sooner. I will give this one more try now from a fresh start, and compare MPI versions.
