### 2021-02-10 #2741 @shahmoradi
@williamfgc Thank you. My impression was that @whedon would automatically build the PDF from the [Markdown file](https://github.com/cdslaborg/paramonte/blob/main/pubs/ParaMonte/kernel/JOSS2020/paper.md). Until then, I attach a PDF of the revised draft to this comment. The Markdown file on GitHub is also readable, although not as comfortable as the PDF.
