### 2021-02-17 #2741 @milancurcic
Good to go from me, thank you all. :+1:

A note regarding "Functionality" and why I left it unchecked: I was reading through the docs and played for a while with the example program and configurations, however I wasn't able to get far, even with the docs. I believe this was due to the domain-specific knowledge that I don't have. But I assume that for a domain expert, these parameters are meaningful and they wouldn't have trouble generalizing the example application while following the docs.
