### 2021-02-22 #2741 @VivianePons
Hi thank you both @williamfgc and @milancurcic for the thorough reviews and @shahmoradi for all the improvements.

@williamfgc you left many boxes unchecked even though I understand that you are satisfied with the software. Maybe that's just an oversight? Could you go through the list and see if everything checks (and indeed CI is not required even though it's always good to have)

@milancurcic is there any functionality that you were not able to reproduce? Or anything you wanted to try and were not able to do? Even if you feel you're not a domain expert, your input might be useful for proving the documentation for example (and as an editor, I would really much like seeing this "functionality" box checked to get the paper published!) Also, you left unchecked the "Performance" box (even there is no performance claims in the paper, it should be checked)

Again thank you for your time! I believe we are getting close to final publication!
