### 2021-02-26 #2741 @milancurcic
@VivianePons

> Also, you left unchecked the "Performance" box (even there is no performance claims in the paper, it should be checked)

Sorry, I misunderstood the prompt. Checked!

> @milancurcic is there any functionality that you were not able to reproduce? Or anything you wanted to try and were not able to do?

No, sorry, I wasn't clear. Everything I tried *seems* to work and do things. What I struggled with was fully understanding what it all means, which is the domain-specific part. I will check the "Functionality" box to the best of my knowledge, however, I believe it will take a domain-expert to truly confirm that the software does what it says. I know and trust the Authors so I feel comfortable checking this box.
