### 2021-03-03 #2741 @VivianePons
Thank you very much @williamfgc ! I see that you have left unchecked the Performance box (it need to be checked if there are no performance claims in the paper) and also "automated tests" and "Community guidelines", is there anything to be done still?

Also, @milancurcic was not completely confident checking the Functionality box mostly because he was not a domain expert and did not fully understand the functionality of the software. Can you tell me to what extent you were able to follow on your end?
