### 2021-03-03 #2741 @williamfgc
@VivianePons I checked the boxes and put some follow up comments in bold. I don't want to be the bottleneck for final publication :) . Thanks for guiding the JOSS review process, I enjoyed it.
