### 2021-03-03 #2741 @VivianePons
Hi @shahmoradi

I believe you're missing a word in this sentence "the library was as an open-source project" line 72 of the PDF.

Also, could you check the missing DOI that were flagged by whedon?

Thank you
