### 2021-03-20 #2741 @shahmoradi
>
>
> Good to go from me, thank you all. 👍
>
> A note regarding "Functionality" and why I left it unchecked: I was reading through the docs and played for a while with the example program and configurations, however I wasn't able to get far, even with the docs. I believe this was due to the domain-specific knowledge that I don't have. But I assume that for a domain expert, these parameters are meaningful and they wouldn't have trouble generalizing the example application while following the docs.

@milancurcic Thanks for pointing this out. Such feedbacks are highly valuable, Could you tell me which examples you tried, so that we review them again for accuracy and comprehensibility? Thanks again.
