### 2021-03-20 #2741 @shahmoradi
>
>
> @shahmoradi I just reviewed the revised manuscript and looks appropriate for JOSS. Thanks for addressing the changes and fixing the defect I filed above. Minor thing: Funding and Acknowledgment can be one single section on the paper. Looks good to me, the only one thing I still think could be added in the future is CI, but I'll leave it to @VivianePons as the editor and @milancurcic to decide on final publication. Thanks!

Thank you @williamfgc . Your comments along with @milancurcic 's were highly valuable and crucial to making this a quality software, although there is still ample room for perfection. I will include your minor revision requests in the final draft.
