### 2021-04-14 #2741 @shahmoradi
My Apologies for the delay, I have become fully occupied with some personal
and family matters over the past month. Please give me until this weekend
to fix the outstanding issues and resubmit this paper to JOSS. Thank you.

On Wed, Apr 14, 2021 at 12:50 PM Kyle Niemeyer ***@***.***>
wrote:

> Hello @shahmoradi <https://github.com/shahmoradi>, just checking in
> again. We are essentially ready to accept your submission, but just need
> you to make that change to your article.
>
