### 2021-05-04 #2741 @VivianePons
It seems that you have the same paper appearing twice in the bib file (A Multilevel Empirical Bayesian Approach to Estimating the Unknown Redshifts of 1366 BATSE Catalog Long-Duration Gamma-Ray Bursts)
