### 2021-05-07 #2741 @VivianePons
Well, it's a bit weird because now you have the whole thing as a zip. The archive should be a direct copy of the repo, like you did in previous versions for example here: https://zenodo.org/record/4414285#.YJTwMiY69H4
