### 2021-05-07 #2741 @VivianePons
We need a DOI with the current version of the software, including all changes made during the review so we can't use the old one
