### 2021-05-08 #2741 @shahmoradi
@VivianePons   I think the current zenodo DOI is automatically connected to the latest version of the software. See for example the history of the version of the package on this page: https://zenodo.org/record/4739775#.YJXr9aFOnGr
I can remove the PDF of the paper that I added manually to the zenodo entry yesterday. If this does not help either, then I will create a DOI via figshare or some other platform as I am somewhat confused now on how Zenodo works and how to create new fresh entries of the same repository.
Thank yo
