### 2021-05-11 #2741 @VivianePons
Hi @shahmoradi each DOI is linked to a specific version, I need the current version. The archive is like a picture of the current state of the software, which I need to finalize the publication, so that, in the future it is clear that the paper refers to this specific state of the software. Please either add a proper archive on the current zenodo or make a new one elsewhere.
