### 2021-05-12 #2741 @danielskatz
My suggested changes are in https://github.com/cdslaborg/paramonte/pull/14 - please merge this, or let me know what you disagree with, then we can proceed
