### 2021-05-13 #2741 @danielskatz
👋 @VivianePons - this paper is probably too long and too detailed, but I don't think it's worth cutting it at this point - this is just a thought of something to consider for other future papers
