### 2021-05-20 #2741 @danielskatz
@arfon and @shahmoradi - I just realized (due to a comment by @kyleniemeyer in another review) that I made a mistake when suggesting edits to this paper, which removed some of the references.  Specifically, in https://github.com/cdslaborg/paramonte/pull/14, I put multiple references together with commas, instead of the correct semicolons, which led to all but the first of each set not being included in the pdf.

I'm quite sorry about this.

I'm not sure how to fix this, as I think the paper src has been deleted.

Can we fix this?  If so, how?
