### 2021-05-20 #2741 @shahmoradi
Thanks for the note. I have revived the manuscript Markdown file in the project's repository. Please let me know if I have to do anything else. Thank you.
