### 2021-05-20 #2741 @danielskatz
Thanks @shahmoradi - can you merge https://github.com/cdslaborg/paramonte/pull/15
