### 2021-05-20 #2741 @arfon
> Can we fix this? If so, how?

I think this should be fixed in https://github.com/openjournals/joss-papers/commit/c529ab42dd63f73bccc42b105115deca72b468c8
